### 2024-03-31 #6562 @editorialbot
[PRE REVIEW]: TREvoSim v3: An individual based simulation for generating trees and character data
**Submitting author:** @RussellGarwood (<a href="http://orcid.org/0000-0002-2803-9471">Russell J Garwood</a>)
**Repository:** https://github.com/palaeoware/trevosim/
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v3.0.0-rc.1
