### 2024-04-02 #6562 @Kevin-Mattheus-Moerman
@RussellGarwood I am the AEiC on this track and here to  help with initial steps.

For the moment, could you please consider the following:

- [ ] Try to address the potential DOI issue highlighted above :point_up: . You can add the DOI to you .bib file and call the following in a comment here `@editorialbot check references`.
- [ ] Your notes to the editorial board referred to some prior publications. Where possible, could you link to them here? Is there a pre-print for that recent "journal of Palaeontology" submission?
