### 2024-04-02 #6562 @RussellGarwood
Hi @Kevin-Mattheus-Moerman many thanks for this. Just to say that your request and the other suggestions above are in my to do list, and I will implement over the next day or two. I spent the weekend walking up Scotland to meet a fieldtrip that is traveling up from my university today, and they (hopefully) will be bringing my laptop (currently writing this on the GitHub app, sorry for typos... )
