### 2024-04-14 #6562 @RussellGarwood
Hi @fboehm - as promised, suggestions, including the two from above:

- Euan Furness (Euan-Furness) - reviewed a paper associated with this software release (see below), and has worked on a related, but distinct, piece of (spatially explicit) simulation software called REvoSim
- Bram van Dijk (bramvandijk88) - edited the JOSS paper associated with the last release of that related software REvoSim, which might assist also in reviewing this package
- Thomas Guillerme (TGuillerme) - may be interested in our approach here, and has done a range of research that involves simulating trees and character data
- Emily Dolson (emilydolson) - also edited the JOSS paper associated with the last release of that related software REvoSim, which might assist also in reviewing this package (but I think might be quite time limited until term ends, based on a JOSS review I did for one of her packages).

Hope this helps!
