### 2024-04-19 #6562 @fboehm
> Hi i @fboehm @Kevin-Mattheus-Moerman just a heads up that the paper mentioned above that employs this software has now been accepted. I will soon be able to share an author accepted manuscript version here if that would be useful.

@RussellGarwood  - thank you. Can you share a hyperlink to it, when it becomes available? The reviewers and I may find it useful.
