### 2024-04-22 #6562 @bramvandijk88
Heya,

I'm currently already enrolled in a review process for JOSS, and due to teaching that one has also been moved back a lot. Hopefully, someone else can cover this one for me, as I'll only find the time after May 31st.
