### 2024-04-22 #6562 @fboehm
@Kevin-Mattheus-Moerman - is there a new command for telling editorialbot to print the joss articles that are "most similar" to the current submission?
