### 2024-04-23 #6562 @Euan-Furness
> @Euan-Furness @bramvandijk88 @TGuillerme - would you be willing to review this submission for JOSS? Our reviews are checklist-driven, so you would review the software submission (and accompanying short manuscript) and assess the extent to which it meets the listed criteria. To see the list of criteria, check out this comment, for example: [#6427 (comment)](https://github.com/openjournals/joss-reviews/issues/6427#issuecomment-1973397744). Thank you for considering this invitation!

Hi Frederick

Apologies for the slight delay in responding; yes, I would be happy to review this. I haven't been on the reviewer side of the conversation for JOSS before though, so please do let me know what you need from me!

Best wishes.
