### 2024-04-23 #6562 @fboehm
@Euan-Furness thank you so much! I'll get a second reviewer in the next few days before we start the "review issue" officially. You'll get a GitHub notification when we do start the review, so you may want to verify that your GitHub settings arr set to, say, send an email to you, or show an alert when you log in, for notifications.

Other that that, there's nothing for you to do right now. Once the review begins, our automated software will send you instructions on how to start the review and what criteria need to be considered. Please feel free to mention me, @fboehm at any time with questions
