### 2024-04-24 #6562 @RussellGarwood
Hi @fboehm - thanks for your work towards finding reviewers! I note we're still one short. Two more potential options:

- Lewis Jones (LewisAJones) works on software development in Palaeo (name the palaeoverse package), and he may be interested in this package, and indeed, the JOSS approach to publishing/reviewing
- Alison Cribb (atcribb), like Euan, reviewed the paper associated with this release, and so may be happy to act as a reviewer for this software

Furthermore, I agree with the suggestion of @TGuillerme that if he is happy to do it, Martin Smith would be well placed to review this!

As requested, the author accepted version of the manuscript mentioned above, is now avaliable [here](https://research.manchester.ac.uk/en/publications/exploring-the-macroevolutionary-impact-of-ecosystem-engineers-usi)
