### 2024-04-24 561587b
Some fixes and tweaks to get it compiling on Windows MSVS2019/QT6.x
