### 2024-04-25 #6562 @fboehm
@ms609 @LewisAJones @atcribb - would you be willing to review this submission for JOSS? Our reviews are checklist-driven, so you would review the software submission (and accompanying short manuscript) and assess the extent to which it meets the listed criteria. To see the list of criteria, check out this comment, for example: https://github.com/openjournals/joss-reviews/issues/6427#issuecomment-1973397744. Thank you for considering this invitation!
