### 2024-04-30 #6562 @fboehm
@bethany-j-allen @willgearty - would you be willing to review this submission for JOSS (Journal of Open Source Software, https://joss.theoj.org/)? I think that your expertise in trees and software development positions you to provide a thorough, constructive review. Our reviews are checklist-driven on Github, so you would review the software submission (and accompanying short manuscript) and assess the extent to which it meets the listed criteria. To see the list of criteria, check out this comment, for example: https://github.com/openjournals/joss-reviews/issues/6427#issuecomment-1973397744. Thank you for considering this invitation!
