### 2024-05-03 #6562 @fboehm
> Thanks both! Can this now go into review @fboehm ? I thought I'd check in because the proofs of the paper using this software just arrived (to which we were hoping to add this as a reference).

Yes, thank you, @RussellGarwood  - I apologize for the delay. i fell behind while traveling this week.
