### 2024-05-05 #6722 @RussellGarwood
Thanks for agreeing to review this submission @Euan-Furness and @atcribb - it's much appreciated (and for acting as editor @fboehm!). As I mentioned in the pre-review github issue, we are hoping to cite this release of the software and the associated  JOSS paper in the ecosystem engineering paper which you both kindly reviewed - you can find the author accepted version of that manuscript [here](https://research.manchester.ac.uk/en/publications/exploring-the-macroevolutionary-impact-of-ecosystem-engineers-usi). The proof of that came through on Friday, but I will ask for a delay on that (apologies that timings are a bit tight, I had forgotten there was a pre-review stage at JOSS, and didn't factor in that that might take six weeks from submission to get to review).

To make your life easier as reviewers, I figured I would post pointers here to the materials that your review checklist will ask you to check.

**General checks**

- The repository for the software is [here](https://github.com/palaeoware/trevosim)
- The [licence](https://github.com/palaeoware/trevosim/blob/master/LICENSE.md) is a GNU GPL v3.0, which [is OSI approved](https://opensource.org/license/gpl-3-0)/
- There are author contributions at the end of the article. The article includes no results, and no original data on human subjects or animals.

**Functionality**

- Portable binaries for the software on Mac and Windows can be found [here](https://github.com/palaeoware/trevosim/releases) (under assets for TREvoSim v3.0.0-rc.1), there is guidance for using each in the readme for that release. Should you wish to build the software (which is not required for use), there is documentation for that [here](https://trevosim.readthedocs.io/en/latest/requirements.html#installation) - note that this version requires Qt v6.

**Documentation**

- You can find a statement of need in the article proof.
- Installation instructions are posted above (the wording of the point here seems geared towards e.g. python and R packages, given this is a stand-alone package, all of the 'dependencies' are required to build rather than run the software, and the portables include all files  required to run the software).
- Hopefully the software is easy to use! In terms of example usage, following the single instruction in the [quick start section](https://trevosim.readthedocs.io/en/latest/introduction.html#quick-start) should output, to your desktop, empricially benchmarked character data you can load into a range of packages.
-  You can find documentation of the functionality primarily in sections [4](https://trevosim.readthedocs.io/en/latest/settings.html), [5](https://trevosim.readthedocs.io/en/latest/settings2.html) and [6](https://trevosim.readthedocs.io/en/latest/loggingsim.html) of the documentation.
- Details of [tests](https://trevosim.readthedocs.io/en/latest/tests.html) are included in the documentation. You can run these through the GUI, and they are also included on this release candidate in the build chain. For the oofficial release of v3, we're going to change this so they are run through the Qt Creator GUI to standardise the procedure across operating systems.
- Community guidelines for contributors are linked from the readme, and can be found [here](https://github.com/palaeoware/repoconventions)

**Software paper**

- You can download a proof of this using the links in the post immediately above this one.

I hope this information is useful - and thanks again for taking the time to look at the software!
