### 2024-05-08 #6722 @willgearty
I just realized this is a C++ project. I'm a bit rusty with C++ and only really would feel comfortable reviewing Python or R projects at the moment. Hopefully @ms609 is available!
