### 2024-05-08 #6722 @RussellGarwood
If it helps, @willgearty you can answer all the questions on the checklist by installing and then playing with the software, rather than having to build it or check the code. Even the tests in TREvoSim can be run (and output to) the GUI. (The thought being this would make it more user friendly - so if it is not so, that would also be super useful to find out!)
