+++ docs/requirements.rst
@@ -1,7 +1,7 @@
 .. _requirements:
 
-Compiling, Installation, and Requirements
-==========================================
+Compiling, Testing, Installation, and Requirements
+==================================================
 
 Compiling from Source
 ----------------------
@@ -28,12 +28,22 @@ We recommend you install and use MSYS2 (64-bit) a Windows package manager, based
 #. You should now be able to find the required libraries under "{install location}/mingw64/bin" and the required header (.h) files for QT v5.x.
 #. Open the .pro file in QT Creator, and then use the information above to setup a new 64-bit ming64 kit. Follow standard QT Creator debug/release procedure.
 
-For v3.X - QT Creator + QT v6.x using Qt installer and MinGW (64-bit)
-"""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
+For v3.X - QT Creator + QT v6.x using Qt installer and MSVC2019 (64-bit)
+"""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
 
-For the TREvoSim v3.0.0 release a build was created using an stand install of Qt6 using the QT online installer available from the `Qt website <https://doc.qt.io/qt-6/get-and-install-qt.html>`_. If following this approach, ensuring the installation includes the latest Qt v6 release, and Qt Creator will allow build of the software. To achieve this, open CMakeLists.txt in Qt creator following installation, select to build a release following the `Qt creator documentation <https://doc.qt.io/qtcreator/creator-building-targets.html>`_, and then initiate a build (ctrl + B / Build, then Build Project). This will create an executable that can be launched by double clicking. To create a build that includes the tests, it will be necessary to modify the cmake file as per the instructions provided with the `Google Test framework  <https://google.github.io/googletest/quickstart-cmake.html>`_. 
+1. Install Qt 6.X on your system by running the installer from Qt: https://www.qt.io/download
+2. Install Microsoft Visual Studio 2019 installed. See https://visualstudio.microsoft.com/vs/older-downloads/. Note Visual Studio 2022 should also work but has not yet been tested with TREvoSim.
+3. Download source code.
+4. In QT Creator open a new project by navigating to and openning the TREvoSim CMakeLists.txt file.
+5. Configure build and follow standard debug/release procedure.
+   
+  - See the Qt Windows Deployment tool: https://doc.qt.io/qt-6/windows-deployment.html
 
-*Note:* At the time of release of v3.0.0 the authors temporarily have limited access to windows machines to allow us to make the above changes to cmake, and provide more explicit instructions. We expect to make a patch release with those changes, and featuring improved documentation for Qt v6.x builds, in May 2024. 
+6. To package the application as Windows Installation Binary for distribution we recommend using Inno Setup: https://jrsoftware.org/isinfo.php
+
+  - Inno Setup is a free installer for Windows programs by Jordan Russell and Martijn Laan.
+  - The file TREvoSim_Installer_v3.x.x_dev.iss contained under the ./inno_setup folder is provided as a template. This file should be modified and used to create the Windows Installation Binary. It expects all support .DLLs and the .EXE to be placed in a ./bin folder. The Inno Setup will then create the new install binary in a folder called ./build
+
 
 Ubuntu 22.04 64-bit - QT Creator + QT v6.x using GCC (64-bit)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -80,7 +90,7 @@ MacOS - QT Creator + QT v6.x using Clang and xcode
 
   $ brew install qt
 
-3. To build the software and test suite, download source code, launch Qt Creator, and open the .cmake file.
+3. To build the software and test suite, download source code, launch Qt Creator, and open the CMakeLists.txt file.
[…]
