### 2024-05-10 palaeoware/trevosim#4 @ms609
Software paper: New features
Re: https://github.com/openjournals/joss-reviews/issues/6722

- [x] It would make the manuscript easier to follow if the new features were arranged thematically, rather than alphabetically (especially as the “feature names” are somewhat arbitrary).  For example, it would seem natural to group features that enrich the simulation model (Ecosystem engineering, Expanding playing field) separately from those that relate more to the operation of the software (Logging, Codebase Tests – which in any case is already presented out of sequence).

- [x] I would suggest that the description of new features in the text should be self-contained, so that a reader can understand what the features are and what purpose they serve.  Whilst I appreciate that further details are available in the documentation and in related papers, I didn’t find it easy to find this information, and I think it asks a lot of a reader to search for context that could be provided reasonably concisely in the manuscript itself.

I'll include specific comments on individual sections as individual issues:
- [x] #5
- [x] #6
- [x] #7
- [x] #8
- [x] #9
- [x] #10
