### 2024-05-10 palaeoware/trevosim#3 @RussellGarwood
Amazing, thanks so much for these pointes @ms609. Will work on this very shortly, and tick off the items in your list as they are addressed. Where requried, I will respond in _italics_
