+++ docs/requirementsinstallation.rst
@@ -41,108 +41,3 @@ Ubuntu 64-bit / Linux
 ^^^^^^^^^^^^^^^^^^^^^
 
 For Linux users, the above instructions will allow the software to be built using a limited number of lines of bash. Please contact palaeoware@gmail.com if you encounter any issues.
-
-Compiling from source and Testing
-==================================================
-
-Compiling from Source
-----------------------
-
-Windows 64-bit
-^^^^^^^^^^^^^^
-
-For v1.X,2.X - QT Creator + QT v5.x using MSYS2 (64-bit) and MinGW (64-bit)
-"""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
-
-We recommend you install and use MSYS2 (64-bit) a Windows package manager, based on modern Cygwin (POSIX compatibility layer) and MinGW-w64, that allows easy installation of QT v5.x 64-bit.
-
-#. Download and run the latest version of `MSYS2 <https://www.msys2.org/>`_ for 64-bit Windows. This will be name "MSYS2-x86_64-..." for the 64-bit installer.
-#. Follow the install instructions. We have used the default install location of "C:\mysys64\" and it is here that includes required in the .pro files point. If you install MSYS2 to another location the .pro files will need to be updated to your install location.
-#. Once installed open up MSYS2 shell and run the pacman update command: pacman -Syu Note that as this will almost certainly update pacman itself you may have to close down and restart the MSYS2 shell before re-running the command to finish.
-#. Once MSYS2 and pacman are fully updated run the following command to install QT 5.x and its dependencies: pacman -S mingw-w64-x86_64-qt-creator mingw-w64-x86_64-qt5
-#. Optional - if you intend on debugging the software in QT and wish to use GDB then run the following to install the matching GBD debugger: pacman -S mingw-w64-x86_64-gdb
-#. At this stage you should have the following under the MSYS2 install location:
-    * {install location}/mingw64 (Main ming64 folder)
-    * {install location}/mingw64/bin/qmake.exe (QMake for QT version)
-    * {install location}/mingw64/bin/g++.exe (C++ complier)
-    * {install location}/mingw64/bin/gcc.exe (C complier)
-    * {install location}/mingw64/bin/gdb.exe (Debugger | OPTIONAL)
-#. You should now be able to find the required libraries under "{install location}/mingw64/bin" and the required header (.h) files for QT v5.x.
-#. Open the .pro file in QT Creator, and then use the information above to setup a new 64-bit ming64 kit. Follow standard QT Creator debug/release procedure.
-
-For v3.X - QT Creator + QT v6.x using Qt installer and MSVC2019 (64-bit)
-""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
-
-1. Install Qt 6.X on your system by running the installer from Qt: https://www.qt.io/download
-2. Install Microsoft Visual Studio 2019 installed. See https://visualstudio.microsoft.com/vs/older-downloads/. Note Visual Studio 2022 should also work but has not yet been tested with TREvoSim.
-3. Download source code.
-4. In QT Creator open a new project by navigating to and openning the TREvoSim CMakeLists.txt file.
-5. Configure build and follow standard debug/release procedure.
-   
-  - See the Qt Windows Deployment tool: https://doc.qt.io/qt-6/windows-deployment.html
-
-6. To package the application as Windows Installation Binary for distribution we recommend using Inno Setup: https://jrsoftware.org/isinfo.php
-
-  - Inno Setup is a free installer for Windows programs by Jordan Russell and Martijn Laan.
[…]
