### 2024-05-14 palaeoware/trevosim#12 @ms609
Community guidelines: Licensing
Re: https://github.com/openjournals/joss-reviews/issues/6722

The licensing is unclear.  The "Repository Contribution" document linked from `README.md` stipulates that "All Palaeoware code is Copyright of the respective authors"; the "Copyright and License" section states "All TREvoSim code is copyright 2024 by Russell J. Garwood and Alan R.T. Spencer."; the documentation adds Mark D. Sutton.

This leaves it unclear who would own the copyright of contributed code, e.g. by a pull request.  What would be the copyright status of a line of code contributed by User A and later modified by User B?  Are you asking contributors to transfer their copyright; should pull request instructions include instructions to change the copyright information, and guidelines for what form of commit would qualify for inclusion in the copyright notice? It may be valuable to include a Contributor License Agreement to clarify the status of contributed code, and allow you to make future decisions about licensing.

I have found https://r-pkgs.org/license.html a useful introduction to some of these issues.
