### 2024-05-14 palaeoware/trevosim#11 @RussellGarwood
Amazing, thanks so much for pointing out where our communication here could be improved @ms609. It is clear that the overall summary could be improved by being less terse to allow more detali - I'll check with the editor in the review thread how much wiggle room there is on word counts, or whether a supplemental PDF would be possible. It is clear from your review that otherwise, to understand the algorith, it is required to refer back to both Keating et al. (2020) and Mongiardino Koch et al. (2021), which I'd like to avoid.
