### 2024-05-14 palaeoware/trevosim#13 @ms609
Community guidelines: Support
Re: https://github.com/openjournals/joss-reviews/issues/6722

It would be useful if the documentation contained clear pointers to where a user can report issues or problems with the software or seek support.
