### 2024-05-14 palaeoware/trevosim#14 @ms609
Contribution and authorship: Acknowledging contributions
Re: https://github.com/openjournals/joss-reviews/issues/6722

Individuals' contributions to the project are acknowledged in the Author Contributions section of the manuscript, which is sufficient.

To adopt best practice and make contributions across all versions of the software more explicit (and machine readable), you might consider including a `codemeta.js` file https://codemeta.github.io/codemeta-generator/ .
R packages use a [contribution taxonomy](https://www.loc.gov/marc/relators/relaterm.html) to identify the nature of contributors' roles in the project.
