### 2024-05-15 11d9d9a
Added instruction to install Git and add GIT and Ninja to the Windows PATH
