### 2024-05-20 palaeoware/trevosim#16 @ms609
Documentation: Window layout
Re: https://github.com/openjournals/joss-reviews/issues/6722

The [Window Layout page](https://trevosim.readthedocs.io/en/latest/windowlayout.html) gives a clear explanation of the interface and the structure is intuitive and easy to follow.

- [x] The images are useful, but are rendered at a maximum 696px width (smaller on a narrow browser window), but have an intrinsic width of 1500px.  Key details are thus too small to clearly distinguish.  I would suggest preparing the screen grabs at the size they will be rendered, i.e. resizing the TREvoSim window to the minimum dimensions in which the key features can be seen, so that these are maximally legible to readers of the documentation.
- [x] A textual description of the appearance of the toolbar (e.g. with ""with buttons greyed out" in place of "as follows") would improve accessibility
- [ ] No messages appeared in the Status Bar during or after a run with the default parameters (i.e. open software and press "run").
