### 2024-05-23 a86b7b9
Build folder
Rather than `build-trevosim-Desktop_Qt_6_6_1_MSVC2019_64bit-Release`, Qt 6.7 seems to build the binaries in `build/Desktop_Qt_6_6_1_MSVC2019_64bit-Release`.  Is this a typo, or does the path depends on the version of Qt?
