### 2024-05-26 #6722 @RussellGarwood
Hi @fboehm - I'm working on the actions @ms609 suggests on the basis of his review, and one of which is this:

https://github.com/palaeoware/trevosim/issues/11

He notes that our algorithm description, which is the core of the package, is by necessity terse (we're alrady at the word limit), and given that it has developed over 2 other publications now, a fresh description of the algorithm in depth would be of utility. I agree, but it would take our paper a fair way over the wordcount. Would this be OK, or is there the possibility of adding an SI document? Or should I alternatively create a PDF and place it in my respository, then refer to it from our paper.
