+++ docs/settings.rst
@@ -30,11 +30,20 @@ Simulations
 :Run mode radio: This has two options, run for iteration number or run to taxon number, which switches between the two options below.
 :Run to taxon number: If run mode is taxon number, a simulation will run until this number of species has evolved, and then terminate.
 :Run for iterations number: If run mode is iteration number, a simulation will run until this number of iterations, and then terminate.
-:Species difference: The hamming distance between a selected organism, post mutation, and either the character string when the species originated, or the last species to originate within the lineage (see paper for rationale) is used within the TREvoSim species concept. This setting defines the required hamming distance for a speciation to have occurred.
+:Species difference: The Hamming distance between a selected organism, post mutation, and either the character string when the species originated, or the last species to originate within the lineage (see paper <which one?> for rationale) is used within the TREvoSim species concept. This setting defines the required Hamming distance for a speciation to have occurred.
 :Unresolvable cut off: With low character numbers, especially when TREvoSim is set to strip out uninformative characters, terminals can have the same character string within a matrix, and thus be unresolvable. This setting defines upper limit for identical terminals. If there are more unresolvable taxa than this number in a single run, output files are not written and a warning is provided. If batch mode is underway, the current run is discarded and started again.
-:Fitness target: This is the target value for the count of 1s in the fitness calculation. See Garwood et al (2019) or Keating et al (2020) for details of the fitness algorithm. At low and high values (max being mask number multiplied by character number) fewer genomes will have peak fitness, at half max value, a larger number of peak fitness organisms exist. This can be quantified using fitness histogram menu command.
+:Fitness target: This is the target value for the count of 1s in the fitness calculation. See Garwood et al (2019) or Keating et al (2020) for details of the fitness algorithm.
+
+// MRS comment: can full citations (and links?) to these references be provided?  Now that aspects of the software are described across multiple pages, there might be value in adding a "references" page to the docs, possibly also giving a brief overview of which features are described in which paper.
+
+At low and high values (max being mask number multiplied by character number) fewer genomes will have peak fitness; at half max value, a larger number of peak fitness organisms exist. This can be quantified using fitness histogram menu command.
 :No selection: If this option is enabled, fitnesses are not calculated for digital organisms in the playing field, and instead every iteration and random individual is selected for replication.
-:Sansomian speciation: This option dictates when the character string of a species is recorded: at speciation, or at extinction. Sansomian speciation, the default, is to record the this when a species goes extinct. This ensures if it is a long-surviving species that the recorded characters string best reflects that closest to its sister group. When the option is not selected, genome is recorded at speciation.
+:Sansomian speciation: This option dictates when the character string of a species is recorded: at speciation, or at extinction. Sansomian speciation, the default, is to record <the this> when a species goes extinct. This ensures if it is a long-surviving species that the recorded <characters string> best reflects that closest to its sister group. When the option is not selected, genome is recorded at speciation.
+
+// MRS comment: I'm not familiar with the term "Sansomian speciation"; can a citation be provided?  The description gives me a flavour of how I might choose whether to tick the box, but not enough detail to really follow the logic (I don't see how the genome at extinction is related to a sister group, for example), so I would find (a link to?) a more detailed explanation of the consequences of this decision helpful.
+
 :Discard deleterious mutations: Optionally, TREvoSim can accept only mutations which are neutral, or improve the fitness of an organism.
[…]
