### 2024-05-28 3ab3e28
Comments on settings2.rst
