+++ docs/loggingsim.rst
@@ -32,9 +32,14 @@ This is of utility for studying processes as they occur during a run, as it is o
 
 The next box (body text) is output at the requested frequency and can be used to record anything in the state of the simulation required to study a process of interest. It uses the same keywords (shown below) as the end run log, outputting data in the current state for the iteration at which the log is written. Again, this requires a line break at the end of the string of interest if you would like outputs for each iteration to appear on subsequent lines. 
 
+// MRS comment: "Body text" does not imply to me that this text will be written every _n_ iterations.  Perhaps "Log text" or "iteration text" would be clearer – or the placeholder text could be updated to e.g. "Please enter text to output every N iterations"?  Alternatively, the "Frequency" input could be set to the right of the "Body text" header with the label "Output every [50] iterations".
+
 There are three further options at the base of the dialogue:
 
 :Write running log: By default this log is not output when a simulation runs. In order to write the files this option needs to be enabled. 
+
+// MRS comment: To avoid users overlooking this box and wondering where their log is, I would suggest putting this at the top of the dialog box, and disabling the "text" fields until it is checked.  Separating this from "Write ecosystem engineers" would also clarify that these are two separate processes, and that the "Frequency" input controls the running log, rather than the EE log.
+
 :Write ecosystem engineers: There is a custom log for simulations in which ecosystem engineers are enabled (the nature of this functionality does not allow all required information to be easily output using the running log: this is primarily a convenience function).
 :Frequency: This dictates the frequency, in iterations, with which the running log is written. 
 
@@ -43,9 +48,20 @@ Keywords
 
 Keywords within two vertical bars ( e.g. \|\|Matrix\|\| ) are replaced as a file is written as follows:
 
+// MRS comment: Providing a minimal example of the output created by each keyword would be invaluable in helping a user design an output file.
+
+
 :Character_Number: This outputs the character number.
-:Count: This is replaced with a counter for batch runs; incrementing by one using C++ numbering (i.e. starting from zero).
+:Count: This is replaced with a counter for batch runs; starting from zero and incrementing by one
+
+// MRS comment: The reference to C++ numbering makes this sound more complex than it is!
+
+// MRS comment: This is zero-padded in filenames, is it worth specifying whether `count` is also zero-padded (and provinding a padded keyword if not)?
+
 :Ecosystem_Engineers: This prints a list of species, and their ecosystem engineering status (i.e. whether any members of this species are ecosystem engineers).
+
+// MRS comment: I think this means "This lists, for each species, whether any of its members are ecosystem engineers".  But it could perhaps be read to mean  "This lists all species that are currently ecosystem engineers".  An example of the output format would be particularly helpful here
+
 :Iteration: This outputs the current iteration number. 
 :Matrix: This is replaced with the matrix from the run.
 :MrBayes_Tree: This writes a tree in standard Newick format, including branch lengths (these are based on iteration number throughout), e.g.
@@ -54,9 +70,15 @@ Keywords within two vertical bars ( e.g. \|\|Matrix\|\| ) are replaced as a file
 
[…]
