### 2024-05-28 4a1dcc8
Comment on commands.rst
