### 2024-05-28 #6722 @fboehm
> Sounds good to me. Perhaps it would make sense to integrate this with the documentation as a separate (possibly HTML) page in the online manual?

I like this idea, @ms609. Thank you!

@RussellGarwood - does this - integrating with the online manual - seem reasonable and doable? Thanks again!
