### 2024-05-28 #6722 @RussellGarwood
Thanks both - yes, putting this in the documentation is far more sensible than any of my suggestions :sweat_smile:
