### 2024-05-29 cf298c1
organise build instructions by version more sensible to address issue #3 point 3
