### 2024-05-29 #6722 @RussellGarwood
@atcribb No worries! For this release candidate, we didn't have the chance to check on a range of maps as were all on the move - it was build by @alanspencer and I can't remember what Mac he has. Sadly he left for two back to back fieldtrips yesterday, so can't check, and any mac builds I do will be on a virtual machine on a Linux box, so I suspect won't cut it. I wonder, though, whether @Marksutton has the ability to build on an M3 ?
