### 2024-05-29 e287d2a
Temporary fix for crash in issue #40
Longer term, removal of strip uninformative factor calculation will
obviate the need for this fix.
