### 2024-05-29 db51e52
Redo titles to address issue #3 point 4
