### 2024-05-29 #6722 @RussellGarwood
Aha, from the documentation "Note: this binary was build and tested on an Apple M1 processor as such users on other chipsets might need to compile and build their own TREvoSim binaries." - hopefully Mark may be able to help us.
