### 2024-05-29 419733c
Update readme to address issue #3 point 6
