### 2024-05-29 a34b02a
Update readme to address issue #3 point 7, remove build warning from docs
