+++ docs_build/_sources/requirementsinstallation.rst.txt
@@ -7,7 +7,7 @@ Requirements and Installation from Binaries
 Requirements
 ------------
 
-TREvoSim has no minimum requirements as such, and will run on most standard systems (Windows/Linux/Mac); it however has not been tested on versions of Windows older than Windows 10, before Ubuntu 22.04, and below macOS High Sierra. Performance will benefit from high processor speed and increased number of processor cores, with large amounts (|gt|4GB) of available RAM recommended for large simulations. Graphics card performance is not relevant as GPUs are not currently used in the program's calculation pipeline. A fast hard drive (e.g. SSD) is recommend when intensive logging is enabled; as slow I/O response time can affect the iteration cycle speed.
+TREvoSim has no minimum requirements as such, and will run on most standard systems (Windows/Linux/Mac); however, it has not been tested on versions of Windows older than Windows 10, versions of Ubuntu older than 18.04 (v1,2) or 22.04 (v3), and below macOS High Sierra. Performance will benefit from high processor speed and increased number of processor cores, with large amounts ( |gt| 4GB) of available RAM recommended for large simulations. Graphics card performance is not relevant as GPUs are not currently used in the program's calculation pipeline. A fast hard drive (e.g. SSD) is recommend when intensive logging is enabled; as slow I/O response time can affect the iteration cycle speed.,mmama
 
 Installation from Binaries
 --------------------------
@@ -15,7 +15,7 @@ Installation from Binaries
 Binary download locations
 ^^^^^^^^^^^^^^^^^^^^^^^^^
 
-From the TREvoSim GitHub repository pre-compiled binary releases and packaged installers can be downloaded.
+You can download pre-compiled binary releases and packaged installers from the TREvoSim GitHub repository:
 
   - https://github.com/palaeoware/trevosim/releases/tag/v3.0.0
   - https://github.com/palaeoware/trevosim/releases/tag/v2.0.0
@@ -27,8 +27,8 @@ Windows 64-bit
 
 For Windows users we provide two means of installation:
 
-1. portable binary release within a .ZIP archive - which just needs extracting to a convenient location.
-2. or, a self contained installer - which when run will guide you though the installation process.
+1. A portable binary release within a .ZIP archive. To run this, all you need to do is extract the zip to a convenient location, and double click on the .exe file.
+2. A self contained installer - download and double click on the file, which will guide you though the installation process.
 
 MacOS
 ^^^^^
