### 2024-05-29 palaeoware/trevosim#3 @RussellGarwood
@m609 - all the above points are now addressed bar one (see below) - I'll leave the comment open for a day or two in case you want to check any of the above out, and if I hear nothing further, then close it.

-- Are there alternatives to QT for users who do not

To an extent, yes. You do not need an account to use on an open source OS, so if those T&Cs are an issue, this would be the solution (although I recognise that may not be convenient). Long term, my intention is to replace all QT datastrctures and classes in the core simulation with STL C++, allowing me to compile a headless version for command line use (which relates to issue #31 ) - but that will be in a future release
