### 2024-05-30 3580d06
Add algorithm and concepts page for issues #37 and #11
