+++ docs/algorithmconcepts.rst
@@ -1,0 +1,6 @@
+.. _algorithmconcepts:
+
+Algorithm and Concepts
+=======================
+
+To do. 
