### 2024-05-31 35027cb
Print root option for issue #37
