### 2024-05-31 c026c6a
Update documentation to reflect rooting addition for issue #37
