### 2024-05-31 fddb1e5
Start work on EE test
I did testing for the paper manually, and just realised I had not added
a proper test for EE to the test suite. Start this
