### 2024-06-01 5ee4def
EE test - apply EE once
