### 2024-06-02 5b39df2
Finish EE tests - all in order
