### 2024-06-03 487c624
Add extra context to clarify character options in v3 - issue #5
