+++ JOSS/paper.md
@@ -90,7 +90,7 @@ TREvoSim v3.0.0 includes a suite of new features that allow the investigation of
 
 ## Character options
 
-New options allow finer control of TREvoSim functions that employ genome characters. In addition to a user-defined number of total characters, a separate limit on character count used for species selection and/or the fitness calculation can now be applied (in the default settings, all are the same). When this differs from the total character number, only a subset of characters (those from zero to the limit) are included in the defined operations, and others can evolve independent of these processes (i.e., in the absence of selective forces, akin to more neutral drift-like processes).  
+New options allow finer control of TREvoSim functions that employ genome characters. Characters in TREvoSim are used in several portions of the algorithm -- they form the basis of calculating fitness of organisms, and are also employed in the identification of species. In previous versions of REvoSim, all characters were used for both functions, through a user-defined total character number. From v3, a separate limit on the character count used for species selection and/or the fitness calculation can be applied (in the default settings, all are the same). When this differs from the total character number, only a subset of characters (those from zero to the limit) are included in the defined operations, and others can evolve independent of these processes (i.e., in the absence of selective forces, akin to more neutral drift-like processes).  
 
 ## Default simulation parameters
 
