### 2024-06-03 1248d4a
Try to clarify the potential impact of adding a mask in EE for issue #4
