+++ JOSS/paper.md
@@ -98,7 +98,7 @@ New default values for simulation parameters are introduced with this release. T
 
 ## Ecosystem engineering
 
-A new ecosystem engineering system allows organism-environment feedback to be investigated. When the ecosystem engineering option is enabled (it is disabled in default settings), a species is assigned ecosystem engineering status halfway through a run, and passes this status to descendents. When this occurs, the genome of that organism is either used to overwrite an environmental mask, or added to the environment as an additional mask. This can occur once (‘one-shot’ ecosystem engineering) or repeatedly (‘persistent’) after this point. When a mask is added in the first application, it is overwritten in subsequent applications when ecosystem engineers are persistent. A new facility to log the ecosystem-engineering status of individuals is provided. 
+A new ecosystem engineering system allows organism-environment feedback to be investigated. When the ecosystem engineering option is enabled (it is disabled in default settings), a species is assigned ecosystem engineering status halfway through a run, and passes this status to descendents. When this occurs, the genome of that organism is either used to overwrite an environmental mask, or added to the environment as an additional mask. Overwriting a mask reduces the hamming distance between engineers and masks, and -- assuming a low fitness target -- thus directly improves their fitness relative to non-engineers. In contrast, adding a mask changes the fitness landscape changes the nature of the fitness landscape for all organisms, but with a weaker direct benefit to ecosystem engineers. Ecosystem engineering can occur just once (‘one-shot’ ecosystem engineering) or can be repeated after the first application (‘persistent’). When a mask is added in the first application, it is overwritten in subsequent applications when ecosystem engineers are persistent. A new facility to log the ecosystem-engineering status of individuals is provided. 
 
 ## Expanding playing field 
 
