### 2024-06-03 54c051f
Improve description of expanding playing field for issue #8
