### 2024-06-03 bd3465d
Clarify what is new in logging to address issue #9
