+++ JOSS/paper.md
@@ -110,7 +110,7 @@ In previous versions, TREvoSim only allowed outputs at the end of a run. From v3
 
 ## Match peaks
 
-When there are multiple environments for any given playing field, by default masks are seeded with random numbers, and may thus have different peak fitness values. This new option instead seeds each playing field with environments that have the same peak fitness. Additionally, it uses a heuristic algorithm to generate an initial organism that has the same fitness in each environment (in >99% of cases). This option allows finer control of the fitness landscape. 
+When there are multiple environments for any given playing field, by default masks are seeded with random numbers, and may thus have different peak fitness values. This new option instead seeds each playing field with environments that have the same peak fitness. This is achieved by doing a site-wise randomisation of the sequence of zeros and ones across masks, between environments. For example, with three masks in a given environment, the first bit may be 1,0,0 for masks 1,2 and 3 respectively -- when this option is enabled, that may be moved to bit three between the first and second environment, and this is repeated for all sites. This operation ensures that the best achievable fitness will remain the same, but will be achieved by a different genome across environments. As the simulation progresses and mutations to the masks occur, matching peaks are no longer guaranteed. Additionally, the simulation uses a heuristic algorithm to generate an initial seed organism that has the same fitness in each environment (in >99% of cases) when this option is selected. In general, this option allows finer control of the fitness landscape. 
 
 ## No selection
 
