### 2024-06-06 4729738
Tests done for mixing with two playingfields.
Just need to do three.
