### 2024-06-08 #6722 @RussellGarwood
Hi @fboehm thanks for checking in!

Updates-wise, I'm making my way through the suggestions @ms609 made - I've made a number of changes to the manuscript (see commits on JOSS branch of repository, or check out the associated issues), and also addressed some of the changes to the code base. I have a fair few more to go though, and am currently on vacation, so progress is slower than it would be if I had a solid chunk of time to spend on this - I hope to have everything addressed by next Friday or so :)

On the mac build, the last I heard from @marksutton suggested he had access to a suitable mac to do a build for an M3 chip - hopefully he can update us when he has the chance.
