### 2024-06-08 3212b27
Add more information to the mixing test
