### 2024-06-09 c97a8bd
Add a new section to introduction to address issue #13
