### 2024-06-09 palaeoware/trevosim#13 @RussellGarwood
The above commit adds a section to the documentation introduction page which points towards the github issues page.
