### 2024-06-09 palaeoware/trevosim#14 @RussellGarwood
Thanks for this suggestion. I had not come acxross a codemeta.js file before - looks like a great idea. I will generate one as soon as the JOSS paper is published, thus allowing me to link the codemeta.js file to the JOSS contribution. To this end I'll be keeping this issue open until that is included
