### 2024-06-09 18d5c22
Add test descriptions
Dropdown for test selection now includes a brief description as
requested in issue #19
