### 2024-06-10 e3fce2d
Add stop button to address issue #17
