### 2024-06-10 #6722 @marksutton
Ah, sorry, have now looked at this _properly_. No I don't have any capacity to build for an M3- I only have access to an M1 mac, and  I'm far from an expert on Macs in general. I don't think there is much I can do to help. I'm more than a bit surprised that the M3 machine won't run M1 code, but if this really is a compatibility issue between chip generations, I can only suggest that you might need to build it yourself on your machine :-(

Apologies for not being able to do more (and for not properly reading the request first time round)
