### 2024-06-11 #6722 @RussellGarwood
@fboehm Yes, I shall add the point about M3 chips to the release notes - sadly I don't have the money to buy the latest mac just to build software (and virtual machines for Macs are tricky)

-- what is required to build and use the software on a Mac, if one can't use the binary? Do you list instructions and dependencies?

Instructions here:

https://trevosim.readthedocs.io/en/latest/buildingfromsource.html
