### 2024-06-11 c0eeb71
Implement pause and cancel for fitness histogram - see issue #17
