### 2024-06-11 67665d5
Ensure that write file setting is honoured to fix issue #33
