### 2024-06-11 dc01d02
Add end; command to data block in nexus file for issue #28, plus other nexus file cleaning
