### 2024-06-12 213009f
Make strip uninformative factor default to five
