### 2024-06-12 83dc462
for issue #27 RUF is now default of five (last commit), and there is no longer a tickbox to recalculate on settings close
