### 2024-06-12 ebc8a9d
All required changes to code base for issue #27 - new messages on fail, and menu option to recalculate factor
