### 2024-06-12 f703bea
Update documentation to reflect changes for issue #27
