### 2024-06-12 36329df
standardise terminal names as per issue #36
