### 2024-06-13 aa32f4a
Allow user to modify the name of the tree file for issue #38
