### 2024-06-13 ab0861a
Send output docs link to the logging URL for issue #28
