### 2024-06-13 bb80dbf
Update documentation to include changes made for issues #28 and #38 - also modify docs and underlying code to rationalise how running log works. There were some aspects of this that didn't make sense when I looked at them again
