+++ output.cpp
@@ -33,7 +33,7 @@ Output::Output(QWidget *parent, simulationVariables *simSettings) :
     ui->intro_01->setText("This output dialogue allows you to set the text for up to two custom output files to reflect the state of a completed simulation. For example, you can output the matrix, the number of characters or terminals, and the tree (which can also be output as a separate file - see below). You can include commands for downstream analysis of the data. For more details of how to write outputs, please see the TREvoSim documentation (available by clicking the button below).");
 
     ui->intro_02->setWordWrap(true);
-    ui->intro_02->setText("In addition to the two custom files above, this option allows you to save the tree as a .nex file.");
+    ui->intro_02->setText("In addition to the two custom files above, this option allows you to save the tree as a .nex file. Provide a base name here - the word tree_ will be appended to the string you provide.");
 
     QObject::connect(ui->docsPushButton, &QPushButton::clicked, this, &Output::docs);
 
@@ -66,8 +66,7 @@ Output::Output(QWidget *parent, simulationVariables *simSettings) :
     ui->c_write_running_log->setChecked(settings->writeRunningLog);
     ui->c_write_ee->setChecked(settings->writeEE);
 
-    ui->running_log_header->insertPlainText(settings->runningLogHeader);
-    ui->running_log_body->insertPlainText(settings->runningLogBody);
+    ui->running_log_body->insertPlainText(settings->runningLogString);
 
     ui->c_work_log->setChecked(settings->workingLog);
 }
@@ -121,8 +120,7 @@ void Output::on_buttonBox_accepted()
     settings->writeFileTwo = ui->c_write_file_02->isChecked();
     settings->writeRunningLog = ui->c_write_running_log->isChecked();
 
-    settings->runningLogHeader =  ui->running_log_header->toPlainText();
-    settings->runningLogBody = ui->running_log_body->toPlainText();
+    settings->runningLogString = ui->running_log_body->toPlainText();
 
     settings->runningLogFrequency = ui->s_frequency->value();
 }
