+++ simulation_variables.h
@@ -75,8 +75,7 @@ public:
     QString logFileString01;
     QString logFileString02;
     QString logFileString03;
-    QString runningLogHeader;
-    QString runningLogBody;
+    QString runningLogString;
 
     //And saving files
     QString savePathDirectory;
