+++ docs_build/_sources/loggingsim.rst.txt
@@ -6,37 +6,36 @@ Logging the Simulation
 Logging system
 --------------
 
-**End run log**
+End run log
+^^^^^^^^^^^
 
 TREvoSim has a versatile logging system which allows the user to define outputs in a range of formats required for phylogenetic inference packages, or e.g. R, provided these allow plain text inputs. Clicking on the output button of the toolbar will launch the output dialogue, which has two tabs. The first is a log that is output at the end of the run:
 
 .. figure:: _static/output_01.png
     :align: center
 
-A typical use case for this is when you would like to use the outputs of a finished simulation to do downstream analyses for e.g. phylogenetic methods. This tab provides options for two custom log files, which are placed, after a run, within the *TREvoSim_output* folder created on the save path. Basename defines the start of the filename, which then either includes a number, which iterates with subsequent runs when output is not set to append, or *_batch* if the outputs are set to append and a batch is being run. The file extension is also defined here. The two text boxes then allow custom file content to be written: text is written as provided to the file -- for example with run instructions or program commands -- and any of the keywords shown below included within two vertical bar ( | ) symbols are replaced as a file is written. The default outputs are shown in the figure above -- they create a vanilla nexus file which allows you to, for example, load the character matrix into R. 
+A typical use case for this is when you would like to use the outputs of a finished simulation to do downstream analyses for e.g. phylogenetic methods. This tab provides options for two custom log files, which are placed, after a run, within the *TREvoSim_output* folder created on the save path. Basename defines the start of the filename. TREvoSim then adds a number to this name which iterates with subsequent runs. The file extension is also defined here. 
 
-There are two further options at the base of the dialogue:
+The two text boxes allow custom file content to be written: text from these is written as provided to the file -- for example with run instructions or program commands -- and any of a series of keywords defined below included within two vertical bar ( | ) symbols are replaced as a file is written. The default outputs are shown in the figure above -- they create a vanilla nexus file which allows you to, for example, load the character matrix into R, and a TNT file which conducts a parsimony search on TREvoSim data before outputting the most parsimonious trees. 
 
-:Output NEX tree file: This outputs a tree for each run in a standard nexus format with the tree and translate block, as well as a comment with the settings of the run written to it. These do not append, but each file has the run number at the end of the file name.
-:Output working log: When this is checked, TREvoSim outputs a text file outlining many of the steps each iteration, such as the state of the playing field, the environment, and the processes the software is going through. This helps understand and fact check any given run, but for significant playing field sizes, taxon numbers, or character numbers, it creates a significant (10s -- 100s of MB) text file.
+In addition to these two custom files, TREvoSim can output a tree for each run in a standard nexus format. The options at the bottom of the output tab allow you to define the basename for this file (the word tree is appended to the name provided). The output includes the tree and translate block, as well as a comment with the settings of the run written to it. 
 
-
-**Running log**
+Running log
+^^^^^^^^^^^
 
 The other tab is a running log which is output during a run:
 
 .. figure:: _static/output_02.png
     :align: center
 
-This is of utility for studying processes as they occur during a run, as it is output on a regular basis as the simulation progresses. The text entered in the header text region is output once, when the file is created -- so if, for example, a CSV output is required and a series of strings are comma separated, these will form the column headers. Note that in order to write the first data output on the next line, a newline is required at the end of this string. 
-
[…]
