### 2024-06-13 da8afc9
Make default tnt output for issue #28
