### 2024-06-13 4765717
Add link to docs from settings dialogue, as mentioned in issue #28
