### 2024-06-13 0cedc74
Changes to batch dialogue to address issue #32
