### 2024-06-14 0a75d88
Complete required coding for issue #32
