### 2024-06-14 1528b76
Add documentation of stochastic layer depth for issue #21
