### 2024-06-14 9b2d059
For issue #18 - move random seed from menu to option in settings toolbox. Ensure this settings works in all ways expected of a setting in simulationVariavbles object including save and load
