### 2024-06-14 946b9d0
Update docs to reflect changes for issue #18
