### 2024-06-14 #6722 @RussellGarwood
@ms609 Just a heads up that I cleaned up the issues on the TREvoSim repository, and have closed 15 that I have finished addressing - descriptions of my changes are in each. Do shout if you want me to reopen any:

https://github.com/palaeoware/trevosim/issues?q=is%3Aissue+is%3Aclosed
