### 2024-06-14 fcecf78
Improve documentation to reflect issue #20
