### 2024-06-14 37f9b5f
update runs behaviour to reflect issue #35
