### 2024-06-14 375d7f3
Update documentation to include some background information on software behaviour
