### 2024-06-14 d93a002
fix silly errors in previous changes for issue #28
