### 2024-06-14 7958a9a
Make set uninformative factor default to the value of the simsettigns variable for issue #27
