### 2024-06-14 755c4f9
Make set uninformative factor message more useful for issue  #27
