### 2024-06-14 689051c
Replace qasconst with std::as_const for issue #47
