+++ simulation.cpp
@@ -113,12 +113,12 @@ simulation::simulation(int runsCon, const simulationVariables *simSettingsCon, b
     for (int p = 0; p < simSettings->playingfieldNumber; p++) playingFields.append(new playingFieldStructure);
     //This should either be to playingfieldSize if PFs do not expand
     if (!simSettings->expandingPlayingfield)
-        for (auto p : qAsConst(playingFields))
+        for (auto p : std::as_const(playingFields))
             for (int i = 0; i < simSettings->playingfieldSize; i++)
                 p->playingField.append(new Organism(runGenomeSize, simSettings->stochasticLayer));
     //Or just one, if they do expand
     else
-        for (auto p : qAsConst(playingFields))
+        for (auto p : std::as_const(playingFields))
             p->playingField.append(new Organism(runGenomeSize, simSettings->stochasticLayer));
     //Note organisms are populated here, but actually set during the initialisation function
 
@@ -131,7 +131,7 @@ simulation::simulation(int runsCon, const simulationVariables *simSettingsCon, b
     //So this is playing field number / environment number / mask number / bits
     //This will fill the masks using random numbers as requried - if we need to match fitness peaks, we will do so below.
     if (!simSettings->matchFitnessPeaks)
-        for (auto p : qAsConst(playingFields))
+        for (auto p : std::as_const(playingFields))
         {
             for (int k = 0; k < simSettings->environmentNumber; k++)
             {
@@ -154,7 +154,7 @@ simulation::simulation(int runsCon, const simulationVariables *simSettingsCon, b
     else
     {
         //First let's populate them with false / 0 bits
-        for (auto p : qAsConst(playingFields))
+        for (auto p : std::as_const(playingFields))
         {
             for (int k = 0; k < simSettings->environmentNumber; k++)
             {
@@ -169,7 +169,7 @@ simulation::simulation(int runsCon, const simulationVariables *simSettingsCon, b
 
         //Now let's do the ones
         //We need to shuffle their position between environments to allow multiple fitness peaks - keep a record
-        for (auto p : qAsConst(playingFields))
+        for (auto p : std::as_const(playingFields))
         {
             QVector <QVector <bool> > used;
             for (int i = 0; i < simSettings->environmentNumber; i++)
@@ -244,8 +244,8 @@ bool simulation::run()
         warning("Oops", "This has not initialised this correctly. Please try different settings or email RJG.");
         return false;
     }
-    else for (auto p : qAsConst(playingFields))
-            for (auto o : qAsConst(p->playingField)) *o = bestOrganism;
+    else for (auto p : std::as_const(playingFields))
+            for (auto o : std::as_const(p->playingField)) *o = bestOrganism;
     \
 
     if (simSettings->test == 3) return true;
@@ -296,7 +296,7 @@ bool simulation::run()
 
         int playingFieldNumber = -1;
         //Do the iteration for all playing fields
-        for (auto pf : qAsConst(playingFields))
+        for (auto pf : std::as_const(playingFields))
         {
             /******** Sort playing field by fitness *****/
             //std::sort(playingField.begin(),playingField.end(),[](const Organism* OL, const Organism* OR){return OL->fitness<OR->fitness;});
@@ -392,7 +392,7 @@ bool simulation::run()
         // For each species, Loop through playing field and count and when one instance of a species exists, overwrite genome in species list, and update display
         QHash<QString, QVector <int> > extinct = checkForExtinct(speciesList);
 
[…]
