### 2024-06-14 3408306
Replace qasconst with std::as_const for issue 47
