+++ testinternal.cpp
@@ -132,7 +132,7 @@ bool testinternal::testZero(QString &outString)
     out << "Organism genome is: " << x.printGenomeString(&org) << "\n";
 
     //Now set masks in simulation to 1
-    for (auto p : qAsConst(x.playingFields))
+    for (auto p : std::as_const(x.playingFields))
         for (int k = 0; k < simSettings.environmentNumber; k++)
             for (int j = 0; j < simSettings.maskNumber; j++)
                 for (auto &i : p->masks[k][j]) i = true;
@@ -163,7 +163,7 @@ bool testinternal::testZero(QString &outString)
     if (error) testFlag = false;
 
     //Now set masks in simulation to 1
-    for (auto p : qAsConst(y.playingFields))
+    for (auto p : std::as_const(y.playingFields))
         for (int k = 0; k < simSettings.environmentNumber; k++)
             for (int j = 0; j < simSettings.maskNumber; j++)
                 for (auto &i : p->masks[k][j]) i = true;
@@ -186,12 +186,12 @@ bool testinternal::testZero(QString &outString)
     if (error) testFlag = false;
 
     //Now set masks in environment 0 simulation to 1
-    for (auto p : qAsConst(z.playingFields))
+    for (auto p : std::as_const(z.playingFields))
         for (int j = 0; j < simSettings.maskNumber; j++)
             for (auto &i : p->masks[0][j]) i = true;
 
     //Now set masks in environment 1 simulation to 0
-    for (auto p : qAsConst(z.playingFields))
+    for (auto p : std::as_const(z.playingFields))
         for (int j = 0; j < simSettings.maskNumber; j++)
             for (auto &i : p->masks[1][j]) i = false;
 
@@ -378,8 +378,8 @@ bool testinternal::testThree(QString &outString)
     x.run();
 
     Organism org(*x.playingFields[0]->playingField[0]);
-    for (auto p : qAsConst(x.playingFields))
-        for (auto o : qAsConst(p->playingField))
+    for (auto p : std::as_const(x.playingFields))
+        for (auto o : std::as_const(p->playingField))
             if (!(*o == org)) testFlag = false;
 
     QString flagString = testFlag ? "true" : "false";
@@ -392,8 +392,8 @@ bool testinternal::testThree(QString &outString)
     y.run();
 
     org = *y.playingFields[0]->playingField[0];
-    for (auto p : qAsConst(y.playingFields))
-        for (auto o : qAsConst(p->playingField))
+    for (auto p : std::as_const(y.playingFields))
+        for (auto o : std::as_const(p->playingField))
             if (!(*o == org)) testFlag = false;
     flagString = testFlag ? "true" : "false";
     out << "Mode independent - checked if all organisms are the same after initialisation and returned " << flagString << ".\n";
@@ -404,8 +404,8 @@ bool testinternal::testThree(QString &outString)
     if (error) return false;
     z.run();
     org = *z.playingFields[0]->playingField[0];
-    for (auto p : qAsConst(z.playingFields))
-        for (auto o : qAsConst(p->playingField))
+    for (auto p : std::as_const(z.playingFields))
+        for (auto o : std::as_const(p->playingField))
             if (!(*o == org)) testFlag = false;
     flagString = testFlag ? "true" : "false";
     out << "Mode identical at start - checked if all organisms are the same after initialisation and returned " << flagString << ".\n";
@@ -436,20 +436,20 @@ bool testinternal::testFour(QString &outString)
     Organism org(20, true);
     org.initialise(20, simSettings.stochasticMap);
     out << "\nStochastic genome is: ";
[…]
