### 2024-06-15 5778ce1
Fix test dialogue for issue #48
