### 2024-06-16 a1f4411
Standardise test return for issue #49
