### 2024-06-17 #6722 @RussellGarwood
@fboehm Thanks for checking on progress! This is proving an iterative process (probably made more confusing by the fact I'm addressing issues in a random order based on how much time I have available at any point that I can work on the codebase or manuscript).

@ms609 Thanks ever so much for all the pointers so far. This is all greatly improving both the consistency of the code and the description of the software - I really do appreciate the time you're taking on this.
