### 2024-06-17 #6722 @atcribb
Hi @RussellGarwood - not totally sure this is specifically an M3 issue. This morning I tried to open the .dmg on an M2 and got a "You can't use this version of the application TREvoSim with this version of macOS" error so it may be more an operating system issue there. Unfortunately I don't think I have access to any machines that 3.0 is going to run on, so it may be best for you and @fboehm to find someone else with a compatible computer. I am really sorry I have not been able to help more!
