### 2024-06-17 #6722 @RussellGarwood
@atcribb - No worries at all. Yeah, I think this reflects the difficulty of building for macs in recent years as new chips keep arriving. This was built on an M1 I recall, and so I guess the only way we can build it for M2 and M3 is to get hold of one to which we have admin rights. Hopefully sometime down the line we will manage this and you can use the software if you so wish!

@fboehm - Given the depth of @ms609's review (the trevosim issues and issue history attest to this) I wonder if one review would be sufficient in this circumstance? It'll take me another week at least to complete addressing those, so there is not a mad rush; my comment more based on the fact that reviewers' time is a valuable resource, and if we can avoid adding to the overall review time burden in this instance, perhaps that would be good for the system as a whole?
