### 2024-06-17 palaeoware/trevosim#52 @RussellGarwood
Mac M2 and M3 builds
Add a note to release to reflect likely issues with M2 and M3 chips
