### 2024-06-17 palaeoware/trevosim#52 @alanspencer
As I'm now back I'll take a look at this issue - I'm sure there is a way to cross-complie btween chipsets and OS versions... we just don't seem to be doing this at the moment. Will try to take a look one evenign this week.
