### 2024-06-17 #6722 @RussellGarwood
FYI, I've made an issue to reflect this. We will either add lack of compatability to the documentation and release notes, or @alanspencer will manage to cross compile the software as per his comment here (thanks Alan!)

https://github.com/palaeoware/trevosim/issues/52
