### 2024-06-17 palaeoware/trevosim#52 @RussellGarwood
Cool, thanks man :) Code base is changing quite a bit ATM, I'm doing it all on the master branch, FYI - though can of course switch to another branch and start merging if that would make stuff easier on your end.
