### 2024-06-17 db6d6f4
Update default nexus out string so it loads in claddis. Change benchmarking wording to comparison to empirical, start modifying R script to run out of the box
