+++ comparison_to_empirical_data/TREvoSim_comaprison_to_empirical_data.r
@@ -12,24 +12,22 @@ rm(list = ls())
 
 ########################################################################################################################
 
-# This script analyses TREvoSim runs, benchmarking them against empirical datasets. 
-# It was written by Nicolás Mongiardino Koch and Russell Garwood, extra steps function courtesy of Joe Keating.
+# This script analyses TREvoSim runs, and then compares them against empirical datasets.
+# It was written by Nicolás Mongiardino Koch and Russell Garwood, with an extra steps function courtesy of Joe Keating.
 # The empirical datasets are sourced from this paper:
 # Mongiardino Koch, N., Garwood, R.J. and Parry, L.A., 2021. Fossils improve phylogenetic analyses of morphological characters. Proceedings of the Royal Society B, 288(1950), p.20210044.
-# The above paper cites the source of each dataset used for the analysis
+# The above paper cites the source of each dataset used for the analysis.
 
 ########################################################################################################################
 
 #Directory with empirical data:
-empiricalWD <- "./Empirical_data/"
+empiricalWD <- "./Empirical_data"
 #Directory with TREvoSim outputs in it:
-sourceWD <- "/home/russell/Desktop/Programming/TREvoSim/analysis/"
+sourceWD <- "./TREvoSim_output_defaults_2024"
 #Directory to output graphs
-outputWD <- "/home/russell/Desktop/"
+outputWD <- "./Graphs_out/"
 
 #You can use the following switches to determine whether it:
-###Calculates rates using an MK model
-calcRates <- FALSE
 ###Counts the number of extant taxa
 countExtant <- TRUE
 ###Counts the number of parsimony steps on the true tree
@@ -91,25 +89,25 @@ steps <- function(tree, mdata) {
 empiricalTreeness <- function() {
   #Get empirical data
   trees <- list.files(empiricalWD, pattern = ".tre", full.names = TRUE)
-  
+
   #Treeness for empirical
   empiricalTreenessVector = vector(length = length(trees))
-  
+
   for (i in 1:length(trees)) {
     tree <- read.tree(trees[i])
     empiricalTreenessVector[i] = treeness(tree)
   }
-  
+
   return(empiricalTreenessVector)
 }
 
 empiricalColless <- function() {
   #Get empirical data
   trees = list.files(empiricalWD, pattern = ".tre", full.names = TRUE)
-  
+
   #Treeshape for empirical
   empiricalTreeshapeVector = vector(length = length(trees))
-  
+
   for (i in 1:length(trees)) {
     tree = read.tree(trees[i])
     #Use the corrected colless so number of tips is not an isssue. See:
@@ -118,50 +116,50 @@ empiricalColless <- function() {
     n = length(tree$tip.label)
     empiricalTreeshapeVector[i] = colless(tree)/((n - 1) * (n - 2)/2)
   }
-  
+
   return(empiricalTreeshapeVector)
 }
 
 empiricalSteps <- function() {
   #Get emprical data
   empiricalStepsDF <- data.frame()
-  
+
   if (length(list.files(empiricalWD, pattern = "empirical_steps.csv")) == 1) {
[…]
