### 2024-06-17 3a5665c
modify default string added in last commit to excape qauotes and thus allow software to compile
