### 2024-06-17 36c7d58
Add zero padding to the tree string so the analysis script works
