### 2024-06-18 37a4aaa
Update R script for comparison to empirical, minor changes to default values, update the files included output under defaults, update graphs - next need to crteate shell script to call R script in folder
