### 2024-06-19 e66832c
Switch treeshape in comparison script to calculating total cophenetic index
