### 2024-06-19 5ab1a42
More updates to script, and some to documentation
