### 2024-06-19 78c05a9
Add details of empirical comparison to the documentation. With this, and other commits in the last 48 hours, most - though not quite all - elements of issue #6 have been addressed. I will add a comment to the issue when back online
