+++ comparison_to_empirical_data/TREvoSim_comaprison_to_empirical_data.r
@@ -321,7 +321,7 @@ if (calcTreeshape) {
     theme_minimal() +
     theme(panel.border = element_rect(color = "black", fill = NA)) +
     ylim(0, 1) +
-    labs(title = "TREvoSim vs Empirical treeshape", x = "Data type", y = "Treeshape (normalised total cophenetic index)") +
+    labs(title = "TREvoSim vs Empirical tree asymmetry", x = "Data type", y = "Tree asymmetry (normalised total cophenetic index)") +
     theme(plot.title = element_text(hjust = 0.5)) +
     theme(legend.position = "none")
 
