+++ comparison_to_empirical_data/TREvoSim_comaprison_to_empirical_data.r
@@ -1,5 +1,10 @@
 ## If packages aren't installed, install them, then load them
-packages <- c("phangorn", "phytools", "phylobase", "Claddis", "tidyverse", "vapoRwave", "treestats", "TreeTools")
+packages <- c("cli", "phangorn", "phytools", "phylobase", "Claddis", "tidyverse", "treestats", "TreeTools")
+# the automatic package installation is nice. vapoRwave is not in CRAN,
+# so presumably will need installing from moldach/vapoRwave 
+# (which would require a user to set up devtools, build tools, etc).
+# If this is just used for colour palettes, might it be simpler to just use
+# inbuilt R palettes from e.g. hcl.pals()?
 if(length(packages[!packages %in% installed.packages()[,"Package"]]) > 0){
   install.packages(packages[!packages %in% installed.packages()[,"Package"]])
 }
@@ -14,7 +19,9 @@ library(treestats)
 library(TreeTools)
 
 #Clear environment
-rm(list = ls())
+#rm(list = ls()) # This is generally discouraged, in case the user has a 
+# precious variable saved in their environment and accidentally deletes it.
+
 
 #If not running from the included bash script, set working directory here
 #To run on the default outputs, this should be the absolute path to the folder comparison_to_empirical_data within the TREvoSim repository
@@ -55,19 +62,11 @@ calcTreeshape <- TRUE
 #Function to calculate treeness
 #i.e. sum of internal branch lengths/total branch lengths
 treeness <- function(tree) {
-  return(1 - (sum(tree$edge.length[which(tree$edge[, 2] %in% c(1:length(tree$tip.label)))])/sum(tree$edge.length)))
+  # `which` and `c` are redundant here; possibly this is more readable?
+  return(1 - (sum(tree$edge.length[tree$edge[, 2] <= length(tree$tip.label)])/sum(tree$edge.length)))
 }
 
-########################################################################################################################
-#Function to calculate FitMK - this approach works because it doesn't matter what order we spit out the values
-#mclapply won't work if you just send the columns to mclapply, but this way we can user an iterator - here characterNo
-doFitMK <- function(characterNo, fSimTree, fSimMatrix) {
-  #Print to show progress is happening
-  ###Note this will only work when called from terminal - won't have any effect in R studio
-  cat("*")
-  #Fit MK to data
-  return(fitMk(simTree, matrix[, characterNo], model = "ER")$rates)
-}
+# doFitMK is unused
 
 ########################################################################################################################
 #Function to compute the number of excess steps for each character of a binary morphological character matrix.
@@ -79,21 +78,36 @@ doFitMK <- function(characterNo, fSimTree, fSimMatrix) {
 #mdata<-read.nexus.data(matrixFiles[1])
 #tree<-read.nexus("/home/russell/Desktop/TREvoSim_output_unrooted/TREvoSim_tree_000.nex")
 #mdata<-read.nexus.data("/home/russell/Desktop/TREvoSim_output_unrooted/TREvoSim_000.nex")
-
-steps <- function(tree, mdata) {
-  char_results <- data.frame(character = 0, steps = 0)
-  levelInData <- levels(as.factor(unlist(mdata)))
-  mdata_df <- as.data.frame(mdata)
-  for (z in 1:length(unlist(mdata[1]))) {
[…]
