### 2024-06-20 palaeoware/trevosim#54 @ms609
Software paper: Summary
re. openjournals/joss-reviews/issues/6722

- [x] The summary paragraph at present feels overly general.  The opening paragraphs on the role of simulations in interrogating evolution and methods are effective – though I wonder whether it might be justified to draw a tighter focus, and assume that the reader already has some concept that simulations may in general be useful?

> simulations can create a known phylogenetic tree alongside associated character data, and therefore play a powerful role in assessing the adequacy of approaches for inferring trees from character data

- [x] This sentence sounds like it is just one more thing that simulations in general can do, rather than one of the distinctive features of TrevoSim.

- [x] It's not immediately obvious what "individual-based" means in the context of line 25.  Something akin to "Trevosim simulates reproducing individuals in order to create a known phylogenetic tree alongside character data, which can be a useful tool when..." might combine these two points to give a reader a clearer immediate picture of the motivation of TREvoSim.

- [x] The final sentence of the Summary paragraph is a little hard to decipher; I think it means "whereas previously TREvosim could only be used to study phylo methods, now it can also help study processes"; it's not clear whether the new features also help to better test phylogenetic methods.

- [x] It would be useful if the summary could summarize what features are introduced in this release - tricky, I appreciate, as there are a fair number and few words available.  But I can imagine a future researcher using v7 of the software wanting to see which release introduced feature X – it would be helpful if the abstract gave some indication as to the specifics of the new additions.
