### 2024-06-20 palaeoware/trevosim#55 @ms609
Software Paper: Statement of need
Re. openjournals/joss-reviews/issues/6722#issuecomment-2104775251

- [x] The Statement of Need feels like it describes the TREvoSim software as a whole, rather than the Need for features that are new in v3.  (This also holds for other aspects of the paper.)
Perhaps @fboehm could comment as to which of these end members JOSS anticipates?

> phylogenetic simulations are conducted

- [x] It's not quite clear what a "phylogenetic simulation" is here; do you have in mind the simulation of character data, a tree (which it itself a form of data), or both?

> such as birth-death models or randomly generated data

> data generated are different in a number of ways to those created using a stochastic model

- [x] TREvoSim uses stochastic / random processes to generate data; its data are not deterministic.  These sentences do not do enough to distinguish [data simulated within] Trevosim from other approaches.

> a level of model misspecification resembling that expected from empirical datasets

- [x] I can see what you are driving at here, but it might be worth spelling out to a reader why this is a desirable feature – if I read this without my brain engaged, I might think that you were advertising that the data are bad because they are a poor match to common models (rather than suggesting that current models are bad because real data don't fit their assumptions).
- [x]  More substantively, it would be nice to somehow demonstrate that the nature of misspecification is similar to and equivalent to that of empirical datasets – though I'm not sure how one would go about that.  Your text implies that TREvoSim data are a good proxy for real data because they are both an equally bad fit to existing models, but if the data are a bad fit to the models for different reasons, I don't think it follows that they are equivalent.  And there are other ways to generate misspecified data – e.g. to use the GTR model to simulate data and the Mk to analyse it.
- I think what's really at stake here is whether TRevoSim data is more "realistic" than data simulated by other approaches, and the data validation offers a good test of this. (It would be interesting to compare the properties of TRevosim simulations with data simulated by other means).

> (true) phylogenetic trees and character data are an emergent property of the simulation
- [x] This to me is the unique selling point of TREvoSim, and the SoN could do more to emphasize how this allows a fundamentally different nature of question to be addressed (and perhaps to lead with this).
