### 2024-06-20 palaeoware/trevosim#57 @ms609
Software paper: References
re. openjournals/joss-reviews/issues/6722

Overall, no references are wasted.

- [x] In "such as birth-death models or randomly generated data", please add a reference to birth-death models, and some sort of reference to help the reader understand what you have in mind by "randomly generated data" (see also #55)

- [x] Optionally, and if there is space, it might add value to the paper if you were to include references that address evolutionary scenarios to those that the new features endeavour to model.  For example, you might pick a couple of interesting case studies to illustrate "scenarios where evolutionary dynamics are driven by variations in the rate of environmental change" (line 121).
A clearer picture of the evolutionary scenarios that TREvoSim v3 can now address might also help to bolster the Statement of Need.
