### 2024-06-22 a347fd4
Changes required for issue #46
