+++ simulation.cpp
@@ -581,23 +581,44 @@ bool simulation::run()
     QList <int> uninformativeCoding;
     QList <int> uninformativeNonCoding;
 
-    //Test for informative
+    //Test for informative characters - this is a seperate operation to actually stripping them
     testForUninformative(speciesList, uninformativeCoding, uninformativeNonCoding);
 
     int uninformativeNumber = uninformativeCoding.length() + uninformativeNonCoding.length();
+    informativeCharacters = runGenomeSize - uninformativeNumber;
+    //If calculateStripUninformativeFactor is running, all we care about is uninformativeCharacters, so return here
+    if (calculateStripUninformativeFactorRunning) return true;
 
-    if (theMainWindow != nullptr)
+    if (theMainWindow != nullptr && !simSettings->stripUninformative && uninformativeNumber > 0)
     {
         QString statusString =
             QString ("There are %1 uninformative characters in your matrix, and the software is not set to strip these out. This is not necessarily a problem, but I thought you should know.").arg(
                 uninformativeNumber);
-
-        if (!simSettings->stripUninformative && uninformativeNumber > 0) theMainWindow->setStatus(statusString);
+        theMainWindow->setStatus(statusString);
     }
 
     //Strip the characters if requried
     if (simSettings->stripUninformative)
     {
+        //Check whether we have enough coding and non coding characters
+        bool requiredCharacterNumber = testForCharacterNumber(uninformativeCoding, uninformativeNonCoding);
+        if (!requiredCharacterNumber && !simSettings->test)
+        {
+            if (theMainWindow != nullptr)
+            {
+                QString label = "It seems there are not enough informative characters to pull this off.\n\n"
+                                "By default, TREvoSim over generates characters by a factor of 5x before trying to strip down to those that are parsimony uninformative. "
+                                "Under your current settings (in which the strip uninformative factor is " + QString::number(simSettings->stripUninformativeFactor) +
+                                ") TREvoSim has not managed to recover enough informative characters. It has only recovered " + QString::number(speciesList[0]->genome.size()) +
+                                " characters. Choosing the menu option \'Recalculate uninformative factor for current settings\' will allow you to recalculate this factor for the current settings, and \'Set uninformative factor' will allow you to set it manually to a large number.\n\n"
+                                "Alternatively, this may be a one off - you could try running a batch of 1, and the program will try repeatedly with these settings - though after ten or more repeats you may want to cancel and change the settings.";
+                warning("Oops", label);
+            }
+            if (simSettings->workingLog) workLogTextStream << "Return at !requiredCharacterNumber\n";
+            clearVectors(playingFields, speciesList);
+            return false;
+        }
+
         bool stripped = stripUninformativeCharacters(speciesList, uninformativeCoding, uninformativeNonCoding);
         if (!stripped)
         {
@@ -605,7 +626,6 @@ bool simulation::run()
             return false;
         }
     }
-    if (calculateStripUninformativeFactorRunning) return true;
 
     /******** Check for intrisnically unresolvable clades *****/
     int unresolvableCount = 0;
@@ -1445,8 +1465,23 @@ void simulation::testForUninformative(QVector <Organism *> &speciesList, QList <
         }
 }
 
+bool simulation::testForCharacterNumber(QList <int> &uninformativeCoding, QList <int> &uninformativeNonCoding)
+{
+    //Deal with informative v.s. uninformative characters
+    int requiredNonCodingCharacterNumber = simSettings->genomeSize - simSettings->fitnessSize;
[…]
