+++ simulation.h
@@ -96,6 +96,7 @@ private:
     void applyPlayingfieldMixing(QVector<Organism *> &speciesList);
     void applyEcosystemEngineering(QVector<Organism *> &speciesList, bool writeEcosystemEngineers);
     void testForUninformative(QVector <Organism *> &speciesList, QList <int> &uninformativeCoding, QList <int> &uninformativeNonCoding);
+    bool testForCharacterNumber(QList <int> &uninformativeCoding, QList <int> &uninformativeNonCoding);
     bool stripUninformativeCharacters(QVector<Organism *> &speciesList, const QList <int> &uninformativeCoding, const QList <int> &uninformativeNonCoding);
     bool checkForUnresolvableTaxa(QVector<Organism *> &speciesList, QString &unresolvableTaxonGroups, int &unresolvableCount);
     bool writeFile(const QString logFileNameBase, const QString logFileExtension, const QString logFileString, const QHash<QString, QString> &outValues, const QVector<Organism *> &speciesList);
