### 2024-06-22 11861ab
Move process events to main thread, not write gui, and also remove timer from gui update as it is making cancel non responsive
