### 2024-06-22 cd3149f
update the software so it updates the gui 10ms or so
