+++ simulation.h
@@ -39,7 +39,6 @@ private:
     int perturbationStart, perturbationEnd, perturbationOccurring;
     int ecosystemEngineeringOccurring;
     int environmentalPerturbationCopyRate;
-    int GUIUPdateTime;
     //One vector to keep a copy of masks, the other to keep a copy of which elements of the original masks have been copied over the perturbation masks
     QVector <QVector <QVector <QVector <bool> > > > environmentalPerturbationMasksCopy;
     QVector <QVector <QVector <QVector <bool> > > > environmentalPerturbationOverwriting;
