### 2024-06-22 1e01560
Found start process event commands that have been slowing down GUI updates by several orders of magnitude since 2017. Oh my, how embarassing.
