### 2024-06-22 a1d09d8
Fix GUI problems reported in issue #44, and also sort out the recalculate functionality to work better if cancelled
