### 2024-06-22 69187d1
Fix window priority issue and repeated warnings for issue #45
