### 2024-06-22 433726a
Fix wording on error message for issue #45
