### 2024-06-23 1a99f07
change wording of fitness histogram docs for issue #29
