### 2024-06-23 612141a
Change default number of bits for fitness histogram - issue #29
