+++ mainwindow.cpp
@@ -710,7 +710,7 @@ void MainWindow::countPeaks()
     //Create a new simulation object - sending it important settings.
     bool error = false;
 
-    int genomeSize = QInputDialog::getInt(this, "Fitness histogram...", "How many bits?", 32, 1, 64, 1, &error);
+    int genomeSize = QInputDialog::getInt(this, "Fitness histogram...", "How many bits?", 16, 1, 64, 1, &error);
     if (!error) return;
 
     int repeats =  QInputDialog::getInt(this, "Fitness histogram...", "How many repeats?", 1, 1, 100000, 1, &error);
