### 2024-06-23 a0e5ac3
Add success message for fitness histogram, except when it is cancelled for issue #29
