### 2024-06-23 7d07bc7
Add warning if histogram will overwrite previous outputs for issue #29
