+++ docs/filecommands.rst
@@ -31,7 +31,7 @@ Commands
 Fitness histogram
 ^^^^^^^^^^^^^^^^^
 
-The fitness algorithm for TREvoSim is described in Keating et al. (2020). The fitness landscape varies depends on the simulation settings. This menu command (shortcut ctrl + shift + h) provides a basic assessment of that landscape by showing the distribution of fitnesses for every possible genome for the current settings. It provides an option to select how many bits in the test organism's genome: run times increase exponentially and larger genome sizes (e.g. |gt| 20) can take a significant amount of time.
+The fitness algorithm for TREvoSim is described in Keating et al. (2020). The fitness landscape varies depends on the simulation settings, and the randomly chosen sequences that are used to create the environmental masks. This menu command (shortcut ctrl + shift + h) provides a basic assessment of that landscape by showing the distribution of fitnesses for every possible genome for the current settings. It provides an option to select how many bits in the test organism's genome: run times increase exponentially, and larger genome sizes (e.g. |gt| 20) can take a significant amount of time. After testing the fitness for every possible rearrangement of bits, it outputs the results to a file with the base name TREvoSim_fitness_histogram in the save path defined in the toolbar. For all genome lengths, this text file includes the settings and a histogram of the number of fitnesses for each potential fitness. For genome sizes below 25, it also includes a full list of every genome for every fitness level. Note that the lower the number assigned by the fitness algorithm, the fitter the organism is. 
 
 Run tests
 ^^^^^^^^^
