### 2024-06-23 a8bede5
Make search to only local trevosim output folder; ensure fiiles are overwritten if exist following this warning for issue #29
