### 2024-06-23 8961ed0
Above 20 bits, do 100000 random genomes for fitness hhistogram docs - issue #29
