### 2024-06-24 c4eaa01
change wording of output tab for issue #38
