### 2024-06-24 #6722 @fboehm
> Hi @RussellGarwood - not totally sure this is specifically an M3 issue. This morning I tried to open the .dmg on an M2 and got a "You can't use this version of the application TREvoSim with this version of macOS" error so it may be more an operating system issue there. Unfortunately I don't think I have access to any machines that 3.0 is going to run on, so it may be best for you and @fboehm to find someone else with a compatible computer. I am really sorry I have not been able to help more!

Thank you, @atcribb - have you been able to build from source the software package? Have you encountered difficulties in doing so? It sounds like you're saying that you have been unable to use the binary file that the authors share. Is that right? Have you also tried to create a binary file on your computer from the source code? I apologize if I've overlooked previous comments from you. Thank you for any clarification.
