### 2024-06-25 18b4c13
This commit incldues a frankly stupid way of doing non repeating random numbers which just doesn't work, and was never going to, for large genome sizes in count peaks function. Pushing it as a gentle reminder of my own inability to think things through at times...
