### 2024-06-27 12d799b
Update doc to reflect changes to fitness histogram for issue #29
