### 2024-06-27 9d16019
Redo wording out count peaks output
