### 2024-06-27 #6722 @fboehm
@atcribb - Just wanted to check to see if you've succeeded in building the software from the source code. I am under the impression that you've found that the binary file that the authors prepared doesn't work with your computer, however it remains unclear to me if you've been able to build the software on your computer from the source code.

I appreciate any clarifications that you might offer. Thanks again!
