### 2024-06-28 a57721b
Minor changes to count peaks, but nothing that will solve a segfault
