### 2024-06-28 e5bd0d0
Fix crash on windows for fitness histogram - issue #29
