### 2024-06-28 ab20362
Modify so fitness histrogram replicates persists within session for issue #29
