### 2024-06-29 fe8baf3
Allow paths in base filenames for issue #34
