### 2024-06-29 4dfe81d
Improve descriptions in window layout page for issue #16
