### 2024-06-29 palaeoware/trevosim#16 @RussellGarwood
With above commit I've improved the wording of this page, thanks for the pointer.

I've also marked as completed the "No messages" point, because I believe this resulted from the way I had previously coded the software to deal with slow GUI updates. This was fixed in commits about a week ago, and certainly on Linux the status bar updates every iteration to give the iteration number. It would be good to know if this isn't the case on windows (i.e. that this issue wasn't related to my old approach to writing the GUI).
