### 2024-06-29 cb505a7
Standardise tests a little
