+++ simulation.cpp
@@ -1337,8 +1337,8 @@ void simulation::applyPlayingfieldMixing(QVector<Organism *> &speciesList)
                 int selectOverwrite = QRandomGenerator::global()->bounded(playingFields[selectPlayingfield]->playingField.count());
                 int sourceForOverwrite = QRandomGenerator::global()->bounded(playingFields[i]->playingField.count());
                 //Check for extinction
-                speciesExtinction(speciesList[playingFields[selectPlayingfield]->playingField[selectOverwrite]->speciesID], playingFields[selectPlayingfield]->playingField[selectOverwrite], iterations,
-                                  simSettings->sansomianSpeciation, simSettings->stochasticLayer);
+                if (!simSettings->test) speciesExtinction(speciesList[playingFields[selectPlayingfield]->playingField[selectOverwrite]->speciesID], playingFields[selectPlayingfield]->playingField[selectOverwrite],
+                                                              iterations, simSettings->sansomianSpeciation, simSettings->stochasticLayer);
 
                 //OVerwerite selected with current organism
                 *playingFields[selectPlayingfield]->playingField[selectOverwrite] = *playingFields[i]->playingField[sourceForOverwrite];
@@ -1689,9 +1689,9 @@ void simulation::speciesExtinction(Organism *speciesListOrganism, const Organism
 {
     if (speciesListOrganism->speciesID != playingFieldOrganism->speciesID)
     {
-        if (!test) qInfo() << "There is a mismatch between species list species ID and playingfield species ID. Can't do extinction. Species list ID " << speciesListOrganism->speciesID <<
-                               " playingfield species ID " << playingFieldOrganism->speciesID << " born of " << playingFieldOrganism->parentSpeciesID << " at iteration " << playingFieldOrganism->cladogenesis << " current it " <<
-                               iterations;
+        if (!simSettings->test) qInfo() << "There is a mismatch between species list species ID and playingfield species ID. Can't do extinction. Species list ID " << speciesListOrganism->speciesID <<
+                                            " playingfield species ID " << playingFieldOrganism->speciesID << " born of " << playingFieldOrganism->parentSpeciesID << " at iteration " << playingFieldOrganism->cladogenesis << " current it " <<
+                                            iterations;
         return;
     }
 
