### 2024-06-29 dd75f1b
Update docs tests = issue #43
