### 2024-06-29 957b0b4
Update docs logging - issue #43
