### 2024-06-29 7ef2787
Update docs settings ages 1 and 2 - issue #43 and issue #16
