### 2024-06-29 1c67bbe
Change order of testflag and fglagstring in test 5 as per PR #50
