+++ testinternal.cpp
@@ -495,10 +495,10 @@ bool testinternal::testFive(QString &outString)
     }
 
     double mean = static_cast<double>(cnt) / 10000;
+    if (mean < 1.25 || mean > 1.31) testFlag = false;
     QString flagString = testFlag ? "true" : "false";
     out << "Ran 10000 mutations of a 128 bit organism. At a rate of " << simSettings.organismMutationRate << " mutation per hundred characters per iteration this resulted in a mean of ";
     out << mean << " mutations. TREvoSim expects this to be between 1.25 and 1.31 and returned " << flagString << "\n";
-    if (mean < 1.25 || mean > 1.31) testFlag = false;
 
     simSettings.organismMutationRate = 2.;
 
@@ -517,10 +517,10 @@ bool testinternal::testFive(QString &outString)
     }
 
     mean = static_cast<double>(cnt) / 10000;
+    if (mean < 2.5 || mean > 2.62) testFlag = false;
     flagString = testFlag ? "true" : "false";
     out << "Ran 10000 mutations of a 128 bit organism. At a rate of " << simSettings.organismMutationRate << " mutations per hundred characters per iteration this resulted in a mean of ";
     out << mean << " mutations. TREvoSim expects this to be between 2.5 and 2.62 and returned " << flagString << "\n";
-    if (mean < 2.5 || mean > 2.62) testFlag = false;
 
 
     simSettings.environmentNumber = 2;
