### 2024-06-29 3b6fc55
change wording of random seed as per comment on PR # 30
