### 2024-06-29 3bae87d
Change behaviour and organisation of output window as per PR #30
