### 2024-06-29 94f8a25
Clarify zero padding on count, and update wording on C++ numbering
