### 2024-06-29 3bbbf9f
Add examples and make other changes requested to logging docs for PR #30
