+++ docs/loggingsim.rst
@@ -28,12 +28,15 @@ The other tab is a running log which is output during a run:
 .. figure:: _static/output_02.png
     :align: center
 
-This is of utility for studying processes as they occur during a run, as it is output on a regular basis as the simulation progresses. Any text entered into the Custom running log text box is output to a new file (appended with an iterator) at the requested frequency. This can be used to record anything in the state of the simulation required to study a process of interest. It uses the same keywords (shown below) as the end run log, outputting data in the current state for the iteration at which the log is written.
-
-There are three further options at the base of the dialogue:
+This is of utility for studying processes as they occur during a run, as it is output on a regular basis as the simulation progresses. Any text entered into the Custom running log text box is output to a new file (appended with an iterator) at the requested frequency. This can be used to record anything in the state of the simulation required to study a process of interest. At the top of the dialogue are two options:
 
 :Write running log: By default the running log is not created or written to when a simulation runs. In order to write the files this option needs to be enabled.
 :Frequency: This dictates the frequency, in iterations, with which a new running log file is written. 
+
+Below this is a free text box. This works the same way, and uses the same keywords (shown below) as the end run log, outputting data in the current state for the iteration at which the log is written -- it is only possible to enter text into this box when the Write running log checkbox is ticked.
+
+There are then two further options at the base of the dialogue:
+
 :Write full working log: When this is checked, TREvoSim outputs a text file that it appends to as it runs. This file outlines every step of each iteration, such as the state of the playing field, the environment, and the processes the software is going through. This helps understand and fact check any given run, but for significant playing field sizes, taxon numbers, or character numbers, it creates a large text file (10s -- 100s of MB). 
 :Write ecosystem engineers: There is a custom log for simulations in which ecosystem engineers are enabled (the nature of this functionality does not allow all required information to be easily output using the running log: this is primarily a convenience function).
 
@@ -43,31 +46,95 @@ Keywords
 Keywords within two vertical bars ( e.g. \|\|Matrix\|\| ) are replaced as a file is written as follows:
 
 :Character_Number: This outputs the character number.
-:Count: This is replaced with a counter for batch runs; starting from zero and incrementing by one. This is padded with leading zeroes to three figures.
-:Ecosystem_Engineers: This prints a list of species, and their ecosystem engineering status (i.e. whether any members of this species are ecosystem engineers).
+:Count: This is replaced with a counter for batch runs; starting from zero and incrementing by one. This is padded with leading zeroes to three figures
+:Ecosystem_Engineers: This prints a list of species, and their ecosystem engineering status (i.e. whether any members of this species are ecosystem engineers). A minimal example is shown below.
+
+::
+
+  Applying ecosystem engineers for 1 time on iteration 10
+  Masks before ecosystem engineers :
+  Playingfield 0
+  Environment 0
+  Mask number 0 :	0001101111
+  Mask number 1 :	0111110101
+  Mask number 2 :	1111001000
+
+
+  Playing field(s):
+
+  Playing field number,Playingfield position,Species ID,Ecosystem engineer,Genome
+  0,0,0,0,0111111101
[…]
