### 2024-06-29 618d9be
Changes to settings2 for PR #30
