### 2024-06-29 eca2438
Changes to settings for PR #30, add links to papers throughout
