+++ docs/settings.rst
@@ -24,11 +24,11 @@ Simulations
 :Run mode radio: This has two options, run for iteration number or run to taxon number, which switches between the two options below.
 :Run to taxon number: If run mode is taxon number, a simulation will run until this number of species has evolved, and then terminate.
 :Run for iterations number: If run mode is iteration number, a simulation will run until this number of iterations, and then terminate.
-:Species difference: The hamming distance between a selected organism, post mutation, and either the character string when the species originated, or the last species to originate within the lineage (see paper for rationale) is used within the TREvoSim species concept. This setting defines the required hamming distance for a speciation to have occurred.
+:Species difference: The Hamming distance between a selected organism, post mutation, and either the character string when the species originated, or the last species to originate within the lineage (see Keating et al. [`2020 <https://doi.org/10.1093/sysbio/syaa012>`_] for rationale) is used within the TREvoSim species concept. This setting defines the required Hamming distance for a speciation to have occurred.
 :Unresolvable cut off: With low character numbers, especially when TREvoSim is set to strip out uninformative characters, terminals can have the same character string within a matrix, and thus be unresolvable. This setting defines upper limit for identical terminals. If there are more unresolvable taxa than this number in a single run, output files are not written and a warning is provided. If batch mode is underway, the current run is discarded and started again.
-:Fitness target: This is the target value for the count of 1s in the fitness calculation. See Garwood et al (2019) or Keating et al (2020) for details of the fitness algorithm. At low and high values (max being mask number multiplied by character number) fewer genomes will have peak fitness, at half max value, a larger number of peak fitness organisms exist. This can be quantified using fitness histogram menu command.
+:Fitness target: This is the target value for the count of 1s in the fitness calculation. See Garwood et al. (`2019 <https://doi.org/10.1111/pala.12420>`_) or Keating et al (2020) for details of the fitness algorithm. At low and high values (max being mask number multiplied by character number) fewer genomes will have peak fitness, at half max value, a larger number of peak fitness organisms exist. This can be quantified using the :ref:`Fitness histogram` menu command.
 :No selection: If this option is enabled, fitnesses are not calculated for digital organisms in the playing field, and instead every iteration and random individual is selected for replication.
-:Sansomian speciation: This option dictates when the character string of a species is recorded: at speciation, or at extinction. Sansomian speciation, the default, is to record the this when a species goes extinct. This ensures if it is a long-surviving species that the recorded characters string best reflects that closest to its sister group. When the option is not selected, genome is recorded at speciation.
+:Sansomian speciation: This option dictates when the character string of a species is recorded: at speciation, or at extinction. Sansomian speciation (named because it was Dr Rob Sansom's idea, rather than being a wider concept), the default, is to record the genome/characters of a species when that species goes extinct. This is useful, in particular with respect to species zero -- that used to seed the simulation. Because this tends to survive for longer than other species (see discussion in :ref:`Tree Rooting`), the genome at the start of a species lifespan could differ significantly from that at its extinction. Recording the genome of the last individual of the species ensures in such long-surviving species that the recorded characters are closest that of its sister species (and also the last common ancestor of these two species). When the option is not selected, genome is recorded at speciation.
 :Discard deleterious mutations: Optionally, TREvoSim can accept only mutations which are neutral, or improve the fitness of an organism.
[…]
