### 2024-06-30 ba37ab5
Make changes to R script - remove fit mk, add cli, simplify treeness as requested
