### 2024-06-30 bd9d60f
Other requested additioins to script
