+++ comparison_to_empirical_data/TREvoSim_comaprison_to_empirical_data.r
@@ -26,6 +26,8 @@ library(TreeTools)
 
 # This script analyses TREvoSim runs, and then compares them against empirical datasets.
 # It was written by Nicolás Mongiardino Koch and Russell Garwood, with an extra steps function courtesy of Joe Keating.
+# It was reviewed for a JOSS paper by Martin Smith who made a number of very valuable improvements.
+
 # The empirical datasets are sourced from this paper:
 # Mongiardino Koch, N., Garwood, R.J. and Parry, L.A., 2021. Fossils improve phylogenetic analyses of morphological characters. Proceedings of the Royal Society B, 288(1950), p.20210044.
 # The above paper cites the source of each dataset used for the analysis.
@@ -93,16 +95,26 @@ excessSteps <- function(tree, mdata) {
 empiricalTreeness <- function() {
   #Get empirical data
   trees <- list.files(empiricalWD, pattern = ".tre", full.names = TRUE)
+  nTrees <- length(trees)
+  if (!nTrees) {
+    stop("No trees found in ", empiricalWD)
+  }
 
   #Treeness for empirical
   empiricalTreenessVector = vector(length = length(trees))
 
-  for (i in 1:length(trees)) {
+  for (i in 1:seq_along(trees)) {
     tree <- read.tree(trees[i])
     empiricalTreenessVector[i] = treeness(tree)
   }
 
   return(empiricalTreenessVector)
+}
+
+TCI <- function(tree) {
+  context <- TCIContext(tree)
+  minimum <- context$minimum
+  (TotalCopheneticIndex(tree) - minimum) / (context$uniform.exp - minimum)
 }
 
 empiricalTreeshape <- function() {
@@ -115,21 +127,24 @@ empiricalTreeshape <- function() {
   for (i in 1:length(trees)) {
     # When first coded (19 June 24) there is an error in the maximum value of the index for the trees below, and thus these are skipped
     # This was immediately fixed in TreeTools, so this skipping code can be removed when that filters through to a release
-    if (i==4 || i==9) next
+    if (packageVersion("TreeTools") <= "1.11.1") {
+      if (i==4 || i==9) next
+    }
     tree = read.tree(trees[i])
-    #Use the corrected colless so number of tips is not an isssue. See:
-    #library(treebalance) collessI( method = 'corrected')
-    #And associated book
-    n = length(tree$tip.label)
-    empiricalTreeshapeVector[i] = TotalCopheneticIndex(tree) / TCIContext(tree)$maximum
-    #empiricalTreeshapeVector[i] = colless(tree)/((n - 1) * (n - 2)/2)
+    empiricalTreeshapeVector[i] <- TCI(tree)
   }
-
   return(empiricalTreeshapeVector)
 }
 
+#J1 balance (shape) index for empirical trees - courtesy of Martin Smith
+empiricalJ1 <- function() {
+  stopifnot(packageVersion("TreeTools") >= "1.11.1.9001")
+  vapply(list.files(empiricalWD, pattern = ".tre", full.names = TRUE),
+         function(tr) J1Index(read.tree(tr)), double(1), USE.NAMES = FALSE)
+}
+
 empiricalSteps <- function() {
-  #Get emprical data
+  #Get empirical data
   empiricalStepsDF <- data.frame()
 
   if (length(list.files(empiricalWD, pattern = "empirical_steps.csv")) == 1) {
@@ -170,11 +185,11 @@ empiricalSteps <- function() {
 
     for (j in 1:length(to_calculate)) {
[…]
