### 2024-06-30 5623a76
Fix some errors in script
