### 2024-06-30 39b12c8
Move simulated viuolin to right for extra steps plot, add line to differentiate from empirical
