### 2024-06-30 02d33f3
Update docs in incorporate changes to graphs from PR #50 and issue #6
