### 2024-06-30 05cf8f0
Add note about new defaults
