### 2024-06-30 palaeoware/trevosim#4 @RussellGarwood
I've addressed all of the individual "New features" issues linked from this, and in doing so have tried throughout to - whilst thinking of the word limit - also make this all self contained. I hope this has improved the clarity throughout. I have also rearranged the new features as suggested. As ever, feedback welcome, and thanks for the pointers!
