### 2024-06-30 843a0a8
Clarify references for other approaches to generating data for issue #57
