+++ JOSS/paper.md
@@ -141,8 +141,8 @@ Simulation-termination can now be configured to occur after either a user-define
 TREvoSim v3.0.0 introduces a test suite covering all aspects of the simulation mechanics. These are included in the build pipeline, which will fail if any tests return a failure. 
 
 # Statement of need
-
-Typically, phylogenetic simulations are conducted using deterministic or stochastic approaches [e.g. @Guillerme_Puttick_Marcy_Weisbecker_2020; @Puttick_O_Reilly_Pisani_Donoghue_2019], such as birth-death models or randomly generated data. TREvoSim complements these by using a selection-driven, agent-based approach: the data generated are different in a number of ways to those created using a stochastic model [@Keating_Sansom_Sutton_Knight_Garwood_2020]. The data generated by the software is likely to violate the assumptions of many common models used in the process of phylogenetic inference, incorporating a level of model misspecification resembling that expected from empirical datasets. Default settings have also been validated to reflect a number of features of empirical data matrices and trees. Given that (true) phylogenetic trees and character data are an emergent property of the simulation, the software is particularly well suited to simulation studies that can be analysed through phylogenetic trees and character data matrices. These include, for example: the impact of missing data on phylogenetic inference; the impact of rates of environmental change on character evolution; and the nature of evolution under different fitness landscapes. 
+Guillerme_2024
+Typically, phylogenetic simulations are conducted using deterministic or stochastic approaches such as birth-death models [e.g. Guillerme_2024] or data based on random numbers [e.g. @Puttick_O_Reilly_Pisani_Donoghue_2019]. TREvoSim complements these by using a selection-driven, agent-based approach: the data generated are different in a number of ways to those created using a stochastic model [@Keating_Sansom_Sutton_Knight_Garwood_2020]. The data generated by the software is likely to violate the assumptions of many common models used in the process of phylogenetic inference, incorporating a level of model misspecification resembling that expected from empirical datasets. Default settings have also been validated to reflect a number of features of empirical data matrices and trees. Given that (true) phylogenetic trees and character data are an emergent property of the simulation, the software is particularly well suited to simulation studies that can be analysed through phylogenetic trees and character data matrices. These include, for example: the impact of missing data on phylogenetic inference; the impact of rates of environmental change on character evolution; and the nature of evolution under different fitness landscapes. 
 
 # Current associated projects 
 
