+++ JOSS/paper.bib
@@ -1,6 +1,6 @@
  @article{Barido-Sottani_Saupe_Smiley_Soul_Wright_Warnock_2020, title={Seven rules for simulations in paleobiology}, volume={46}, ISSN={0094-8373, 1938-5331}, DOI={10.1017/pab.2020.30}, abstractNote={Simulations are playing an increasingly important role in paleobiology. When designing a simulation study, many decisions have to be made and common challenges will be encountered along the way. Here, we outline seven rules for executing a good simulation study. We cover topics including the choice of study question, the empirical data used as a basis for the study, statistical and methodological concerns, how to validate the study, and how to ensure it can be reproduced and extended by others. We hope that these rules and the accompanying examples will guide paleobiologists when using simulation tools to address fundamental questions about the evolution of life.}, number={4}, journal={Paleobiology}, publisher={Cambridge University Press}, author={Barido-Sottani, Joëlle and Saupe, Erin E. and Smiley, Tara M. and Soul, Laura C. and Wright, April M. and Warnock, Rachel C. M.}, year={2020}, month=nov, pages={435–444}, language={en} }
  @article{Dolson_Ofria_2021, title={Digital Evolution for Ecology Research: A Review}, volume={9}, ISSN={2296-701X}, DOI={10.3389/fevo.2021.750779}, url={https://www.frontiersin.org/articles/10.3389/fevo.2021.750779}, abstractNote={In digital evolution, populations of computational organisms evolve via the same principles that govern natural selection in nature. These platforms have been used to great effect as a controlled system in which to conduct evolutionary experiments and develop novel evolutionary theory. In addition to their complex evolutionary dynamics, many digital evolution systems also produce rich ecological communities. As a result, digital evolution is also a powerful tool for research on eco-evolutionary dynamics. Here, we review the research to date in which digital evolution platforms have been used to address eco-evolutionary (and in some cases purely ecological) questions. This work has spanned a wide range of topics, including competition, facilitation, parasitism, predation, and macroecological scaling laws. We argue for the value of further ecological research in digital evolution systems and present some particularly promising directions for further research.}, journal={Frontiers in Ecology and Evolution}, author={Dolson, Emily and Ofria, Charles}, year={2021} }
[…]
