### 2024-07-02 #6722 @fboehm
Hello, @atcribb - can you clarify the status of your attempts to build the software from source? Have you encountered difficulties when trying to do this? If so, can you describe them here in the thread?

thank you again!
