### 2024-07-04 #6722 @atcribb
Hi @fboehm, my apologies. I have not been able to build the software from source on my computer.
