### 2024-07-04 #6722 @atcribb
Hi @RussellGarwood – I tried to follow the manual but ran into some snags with Qt. This could definitely just be a skill issue though – I don't have any other prior experience.
