### 2024-07-04 #6722 @RussellGarwood
Thanks @atcribb - if useful I could happily zoom with you to try and figure out the issue (and if I can't we can try and coerce @AlanSpencer into helping). Do shout if this would be useful!
