### 2024-07-07 64aa094
Started algorithm and concepts outline of algorithm
