+++ docs/algorithmconcepts.rst
@@ -6,7 +6,81 @@ Algorithm and Concepts
 TREvoSim Algorithm Description
 ------------------------------
 
-To do. 
+TREvoSim has been developed across several publications, and the algorithm was first described in Keating et al. (2020, v1.0.0), with modifications outlined in Mongiardino Koch et al (2021, v2.0.0). The simulation approach also shares features with the spatially explicit eco-evolutionary simulation package REvoSim (Garwood et. al. 2019, Furness et. al. 2023). The latest version of the algorithm is described here to avoid having to build a picture by accessing multiple publications. Note, this is modified after those previous descriptions, and so employs the same wording in places.
+
+TREvoSim employs digital organisms in the form of binary strings of a user-defined length (e.g. 0010110011101010011011). The zeros and ones comprising the string are here referred to as the genome of the digital organism, and ultimately provide the binary characters that can be used at the end of a simulation for, for example, phylogenetic inference, and they are also used to assign fitness to organisms relative to an environment, as outlined below. When a simulation runs, organisms compete, replicate, and mutate. The simulation incorporates a species concept, and speciation is emergent in the simulation, allowing the software to output a phylogenetic tree showing species relationships during or at the end of a run. 
+
+This overview employs figures from Garwood et al. (2024) -- in all, user defined variables are shown in green font, and the values shown are the default. For clarity, the description eschews listing every potential modification to the algorithm that can be achieved through modifying the software settings, as this would obfuscate important functions in potentially unnecessary detail (although it does mention a number of key choices). Where a setting modifies the algorithm outlined below, this is described fully in the information included on the settings pages in this documentation. The implications of changing a setting are also described in these pages, wherever possible.
+
+Data structures
+^^^^^^^^^^^^^^^
+
+Organisms reside in a structure called the playing field -- a list of organisms that, once a simulation is running, are sorted by fitness. The length of this list is user defined (this size is the limit on the number of organisms 'alive' at any given point during the simulation, and can include members of different species, as well as multiple organisms from the same species). All organisms within a single playing field compete with each other, and the default size of the playing field is 20 organisms. Associated with the playing field is one (by default), or several, environments. Environments in TREvoSim are a user-defined quantity of random numbers (= masks) that default to the same length as the organism genome. The default TREvoSim environment consists of 3 masks, of 128 characters. See below for a summary. 
+
+.. figure:: _static/data_structures.png
+    :width: 400
+    :align: center
+
+If so desired, simulations can employ more than one playing field, which can have the same, or different, environment(s). 
+
+If you wish to modify TREvoSim code, bear in mind that there are numerous other data structures present in the underlying code. You can find an overview of these at the start of mainwindow.cpp
+
+Fitness calculation 
+^^^^^^^^^^^^^^^^^^^
+
[…]
