### 2024-07-08 ed9d4df
Finish writing about iteration phase
