### 2024-07-08 76709c7
More typings
