### 2024-07-08 757e194
First draft of algorithm overview done in documentation for issue #11
