+++ docs_build/_sources/algorithmconcepts.rst.txt
@@ -15,7 +15,7 @@ This overview employs figures from Garwood et al. (2024) -- in all, user defined
 Data structures
 ^^^^^^^^^^^^^^^
 
-Organisms reside in a structure called the playing field -- a list of organisms that, once a simulation is running, are sorted by fitness. The length of this list is user defined (this size is the limit on the number of organisms 'alive' at any given point during the simulation, and can include members of different species, as well as multiple organisms from the same species). All organisms within a single playing field compete with each other, and the default size of the playing field is 20 organisms. Associated with the playing field is one (by default), or several, environments. Environments in TREvoSim are a user-defined quantity of random numbers (= masks) that default to the same length as the organism genome. The default TREvoSim environment consists of 5 masks, of 128 characters. See below for a summary. 
+Organisms reside in a structure called the playing field -- a list of organisms that, once a simulation is running, are sorted by fitness. The length of this list is user defined (this size is the limit on the number of organisms 'alive' at any given point during the simulation, and can include members of different species, as well as multiple organisms from the same species). All organisms within a single playing field compete with each other, and the default size of the playing field is 20 organisms. Associated with the playing field is one (by default), or several, environments. Environments in TREvoSim are a user-defined quantity of random numbers (= masks) that default to the same length as the organism genome. The default TREvoSim environment consists of 3 masks, of 128 characters. See below for a summary. 
 
 .. figure:: _static/data_structures.png
     :width: 400
@@ -28,7 +28,7 @@ If you wish to modify TREvoSim code, bear in mind that there are numerous other 
 Fitness calculation 
 ^^^^^^^^^^^^^^^^^^^
 
-Every iteration, the fitness of every organism in the playing field is calculated (using an approach first described in Garwood et al. 2019, and further detailed in Keating et al. 2020). To calculate fitness, TREvoSim uses an exclusive OR function (i.e. one that takes two inputs, outputs 1 where they differ, and 0 when they are the same) to calculate the `Hamming distance  <https://en.wikipedia.org/wiki/Hamming_distance>`_ between the genome of the organism, and each mask in an environment in turn. For example, if there are five masks, the genome is compared to each in turn. The output of this operation has a one wherever inputs differ. The ones are summed across all output strings, so with five masks, by comparing every bit of the organism genome to the equivalent bit of each mask and summing the results, TREvoSim calculates a value between 0 and 5n. Fitness is then an integer calculated as the distance from a used-defined target value: by default that is zero, creating a sharp fitness peak in the environment (a small number of possible genome configurations can achieve maximal fitness), but setting this to e.g. 2.5n will allow a far wider range of genomic configurations to achieve maximal fitness. This approach is summarised in the figure below.
[…]
