+++ JOSS/paper.md
@@ -77,12 +77,11 @@ bibliography: paper.bib
 
 # Summary
 Simulations provide valuable insights into the patterns and processes of evolution, and the performance of analytical methods used to investigate empirical data
-[@Wright_Hillis_2014; @Barido-Sottani_Saupe_Smiley_Soul_Wright_Warnock_2020; @Dolson_Ofria_2021]. This is particularly true for building phylogenies (evolutionary trees) where the true relationships between groups are unknowable in empirical settings and using empirical data.  Simulations can create a known phylogenetic tree alongside associated character data, and therefore play a powerful role in assessing the adequacy of approaches for inferring trees from character data. Here we present TREvoSim v3.0.0: an individual-based model written in C++, with a focus on creating phylogenetic character data and trees. Previous versions have been used to study phylogenetic inference methods [@Keating_Sansom_Sutton_Knight_Garwood_2020; @Mongiardino_Koch_Garwood_Parry_2021; @Mongiardino_Koch_Garwood_Parry_2023]; the v3.0.0 release adds a range of new features that allow the study of evolutionary processes, in addition to phylogenetic methods.  
+[@Wright_Hillis_2014; @Barido-Sottani_Saupe_Smiley_Soul_Wright_Warnock_2020; @Dolson_Ofria_2021]. This is particularly true for building phylogenies (evolutionary trees) where the true relationships between groups are unknowable in empirical settings and using empirical data. Simulations can create a known phylogenetic tree alongside associated character data, and therefore play a powerful role in assessing the adequacy of approaches for inferring trees from character data. Here we present TREvoSim v3.0.0: an individual-based model written in C++, with a focus on creating phylogenetic character data and trees. Previous versions have been used to study phylogenetic inference methods [@Keating_Sansom_Sutton_Knight_Garwood_2020; @Mongiardino_Koch_Garwood_Parry_2021; @Mongiardino_Koch_Garwood_Parry_2023]; the v3.0.0 release adds a range of new features that allow the study of evolutionary processes, in addition to phylogenetic methods.  
 
 # Background 
 
-
-TREvoSim v1.0.0 was developed to investigate the accuracy and precision of phylogenetic inference methods [@Keating_Sansom_Sutton_Knight_Garwood_2020]. After further development, TREvoSim v2.0.0 was used to investigate the impact fossils have on phylogenetic inference and evolutionary timescales [@Mongiardino_Koch_Garwood_Parry_2021; @Mongiardino_Koch_Garwood_Parry_2023]. In brief, TREvoSim is a non spatially-explicit model in which organisms — which consist of a genome of binary characters — compete within a structure called the playing-field to echo natural selection (Figure 1). Their chance of replication is dictated by a fitness algorithm that assesses organismal fit against a series of random numbers (masks, constituting an environment). On replication, organisms have a chance of mutation, and descendents overwrite a current member of the playing field. The simulation has a lineage-based species concept, and at the end of a simulation can output trees and characters (species genomes), as well as logging the simulation state as the model runs.
[…]
