### 2024-07-09 palaeoware/trevosim#16 @RussellGarwood
Just to be sure, does this include the message "Iteration: XX" when a simulation runs?

![image](https://github.com/palaeoware/trevosim/assets/1557073/89c32ff6-2913-4e7a-be64-2bfee0861a26)

If so, I think the status bar itself doesn't seem to work on windows, and I should have a dig around!
