### 2024-07-09 #6722 @RussellGarwood
@ms609 - not quite finished on the changes you suggest, on the paper, FYI, so there are a few that haven't been actioned in the above PDF. Should be done with those soon, I hope!
