### 2024-07-09 palaeoware/trevosim#16 @ms609
No "Iteration" message is visible; the bar remains blank.
