### 2024-07-09 palaeoware/trevosim#4 @ms609
This looks great – this section now feels coherent and easy to follow, and strikes an excellent balance between brevity and exposition.
