### 2024-07-09 palaeoware/trevosim#16 @RussellGarwood
Hmm, on a windows build here by @alanspencer we definitely have status messages appearing: is it possible this has not updated all elements since you pulled the latest version? One way to check would be to clear or rename your build directory and then rebuild.
