### 2024-07-09 palaeoware/trevosim#16 @ms609
I've just pulled the latest version, deleted the `build` directory, and re-built; still no status messages.
