+++ CMakeLists.txt
@@ -12,7 +12,8 @@ set(CMAKE_AUTOMOC ON)
 set(CMAKE_AUTORCC ON)
 
 # Set the macOS binary to both intel and arm
-set(CMAKE_OSX_ARCHITECTURES "arm64;x86_64" CACHE STRING "" FORCE)
+set(CMAKE_OSX_ARCHITECTURES "x86_64;arm64" CACHE STRING "" FORCE)
+set(CMAKE_OSX_DEPLOYMENT_TARGET "13.0" CACHE STRING "Minimum OS X deployment version")
 
 #TREvoSim
 project(${TREVOSIM_APP} VERSION ${TREVOSIM_VERSION} LANGUAGES CXX)
