### 2024-07-10 palaeoware/trevosim#52 @alanspencer
Here is a link to a new .dmg host on GoogleDrive that should be a universal MacOS build (arm64 and x86_64). The XCode requiment is set to SDK 13, which should cover all newer MacOS versions (i.e. iOS 15, iPadOS 15, tvOS 15, watchOS 8, and macOS Big Sur 11.3 or above).

https://drive.google.com/file/d/1ud8cXWXtPbKD3BsIvEwHinWhjk5Vnz0y/view?usp=drive_link
