### 2024-07-11 palaeoware/trevosim#11 @RussellGarwood
With the above push, a first draft of the concepts and algorithm overview is available. It probably still requires a proofread, but I don't think there is anything I had been planning to add missing
