### 2024-07-11 d4613f1
Update docs to include settings persistent
