### 2024-07-12 c76b5de
Modify summery for issue #54
