+++ JOSS/paper.md
@@ -77,7 +77,7 @@ bibliography: paper.bib
 
 # Summary
 Simulations provide valuable insights into the patterns and processes of evolution, and the performance of analytical methods used to investigate empirical data
-[@Wright_Hillis_2014; @Barido-Sottani_Saupe_Smiley_Soul_Wright_Warnock_2020; @Dolson_Ofria_2021]. This is particularly true for building phylogenies (evolutionary trees) where the true relationships between groups are unknowable in empirical settings and using empirical data. Simulations can create a known phylogenetic tree alongside associated character data, and therefore play a powerful role in assessing the adequacy of approaches for inferring trees from character data. Here we present TREvoSim v3.0.0: an individual-based model written in C++, with a focus on creating phylogenetic character data and trees. Previous versions have been used to study phylogenetic inference methods [@Keating_Sansom_Sutton_Knight_Garwood_2020; @Mongiardino_Koch_Garwood_Parry_2021; @Mongiardino_Koch_Garwood_Parry_2023]; the v3.0.0 release adds a range of new features that allow the study of evolutionary processes, in addition to phylogenetic methods.  
+[@Wright_Hillis_2014; @Barido-Sottani_Saupe_Smiley_Soul_Wright_Warnock_2020; @Dolson_Ofria_2021]. Here we present TREvoSim v3.0.0: an agent- or individual-based model written in C++, in which digital organisms evolve, creating phylogenetic character data and trees. Trees inferred from empirical data always carry uncertainty, but TREvoSim can create a known tree alongside associated character data, allowing assessment of the inference methods used [@Keating_Sansom_Sutton_Knight_Garwood_2020; @Mongiardino_Koch_Garwood_Parry_2021; @Mongiardino_Koch_Garwood_Parry_2023]. The v3.0.0 release adds a range of new features to allow greater flexibility in simulating phylogenetic trees and character data (more logging options, finer control over character character and simulation parameters), and to facilitate the study of broader evolutionary topics (e.g. ecosystem engineering, adaptive landscapes, selection).  
 
 # Background 
 
@@ -123,7 +123,7 @@ This implements a limited period of increased rates of environmental change that
 
 ### Character limits
 
-New options allow finer control of TREvoSim functions that employ genome characters. Characters in TREvoSim are used in several portions of the algorithm -- they form the basis of calculating fitness of organisms, and are also employed in the identification of species. In previous versions of TREvoSim, all characters were used for both functions, through a user-defined total character number. From v3, a separate limit on the character count used for species selection and/or the fitness calculation can be applied (in the default settings, all are the same). When  either - or both - differ from the total character number, only a subset of characters (those from zero to the limit) are included in the defined operations, and others can evolve independent of these processes (i.e., in the absence of selective forces, akin to more neutral drift-like processes).
+New options allow finer control of TREvoSim functions that employ genome characters. Characters in TREvoSim are used in several portions of the algorithm -- they form the basis of calculating fitness of organisms, and are also employed in the identification of species. In previous versions of TREvoSim, all characters were used for both functions, through a user-defined total character number. From v3, a separate limit on the character count used for species selection and/or the fitness calculation can be applied (in the default settings, all are the same). When  either -- or both -- differ from the total character number, only a subset of characters (those from zero to the limit) are included in the defined operations, and others can evolve independent of these processes (i.e., in the absence of selective forces, akin to more neutral drift-like processes).
 
 ### Default simulation parameters
 
