### 2024-07-12 0f710cd
Remove one more word for issue #54
