+++ JOSS/paper.md
@@ -77,7 +77,7 @@ bibliography: paper.bib
 
 # Summary
 Simulations provide valuable insights into the patterns and processes of evolution, and the performance of analytical methods used to investigate empirical data
-[@Wright_Hillis_2014; @Barido-Sottani_Saupe_Smiley_Soul_Wright_Warnock_2020; @Dolson_Ofria_2021]. Here we present TREvoSim v3.0.0: an agent- or individual-based model written in C++, in which digital organisms evolve, creating phylogenetic character data and trees. Trees inferred from empirical data always carry uncertainty, but TREvoSim can create a known tree alongside associated character data, allowing assessment of the inference methods used [@Keating_Sansom_Sutton_Knight_Garwood_2020; @Mongiardino_Koch_Garwood_Parry_2021; @Mongiardino_Koch_Garwood_Parry_2023]. The v3.0.0 release adds a range of new features to allow greater flexibility in simulating phylogenetic trees and character data (more logging options, finer control over character character and simulation parameters), and to facilitate the study of broader evolutionary topics (e.g. ecosystem engineering, adaptive landscapes, selection).  
+[@Wright_Hillis_2014; @Barido-Sottani_Saupe_Smiley_Soul_Wright_Warnock_2020; @Dolson_Ofria_2021]. Here we present TREvoSim v3.0.0: an agent- or individual-based model written in C++, in which digital organisms evolve, creating phylogenetic character data and trees. Trees inferred from empirical data always carry uncertainty, but TREvoSim can create a known tree alongside associated character data, allowing assessment of the inference methods used [@Keating_Sansom_Sutton_Knight_Garwood_2020; @Mongiardino_Koch_Garwood_Parry_2021; @Mongiardino_Koch_Garwood_Parry_2023]. The v3.0.0 release adds a range of features to allow greater flexibility in simulating phylogenetic trees and character data (more logging options, finer control over character character and simulation parameters), and to facilitate the study of broader evolutionary topics (e.g. ecosystem engineering, adaptive landscapes, selection).  
 
 # Background 
 
@@ -143,7 +143,7 @@ TREvoSim v3.0.0 introduces a test suite covering all aspects of the simulation m
 
 # Statement of need
 
-Typically, phylogenetic simulations are conducted using deterministic or stochastic approaches such as birth-death models [e.g. @Guillerme_2024] or data based on random numbers [e.g. @Puttick_O_Reilly_Pisani_Donoghue_2019]. TREvoSim complements these by using a selection-driven, agent-based approach: the data generated are different in a number of ways to those created using a stochastic model [@Keating_Sansom_Sutton_Knight_Garwood_2020]. The data generated by the software is likely to violate the assumptions of many common models used in the process of phylogenetic inference, incorporating a level of model misspecification resembling that expected from empirical datasets. Default settings have also been validated to reflect a number of features of empirical data matrices and trees. Given that (true) phylogenetic trees and character data are an emergent property of the simulation, the software is particularly well suited to simulation studies that can be analysed through phylogenetic trees and character data matrices. These include, for example: the impact of missing data on phylogenetic inference; the impact of rates of environmental change on character evolution; and the nature of evolution under different fitness landscapes. 
[…]
