### 2024-07-12 palaeoware/trevosim#54 @RussellGarwood
With the above commits I have rewriten the abstract/summary with these changes in mind. I did find it quite challenging but I hope I have addressed these points without expanding the summary too much by doing the following:

- I have reduced the amount of general information regarding simulations in the abstract, and more specifically tied one of the examples highlighted to TREvoSim, reordering the sentences to achieve this
- I incorporated a brief description of what an individual based model is in the abstract, however, I also elected to keep individual-based in, and added agent-based as well, given that these are specific terms used in ecology / digital evolution communities that might appreciate this software
- I've provided, I hope, clarity on the nature of the new additions and how they might be helpful, whilst keeping additional words to a minimum.
