+++ JOSS/paper.md
@@ -143,7 +143,7 @@ TREvoSim v3.0.0 introduces a test suite covering all aspects of the simulation m
 
 # Statement of need
 
-Typically, phylogenetic trees or character data are simulated using stochastic approaches such as birth-death models [e.g. @Guillerme_2024] or data based on random numbers [e.g. @Puttick_O_Reilly_Pisani_Donoghue_2019]. TREvoSim complements these by using a selection-driven, agent-based approach: the data generated are different in a number of ways to those created using a stochastic model [@Keating_Sansom_Sutton_Knight_Garwood_2020]. The data generated by the software is likely to violate the assumptions of many common models used in the process of phylogenetic inference, incorporating a level of model misspecification resembling that expected from empirical datasets. Default settings have also been validated to reflect a number of features of empirical data matrices and trees. Given that (true) phylogenetic trees and character data are an emergent property of the simulation, the software is particularly well suited to simulation studies that can be analysed through phylogenetic trees and character data matrices. These include, for example: the impact of missing data on phylogenetic inference; the impact of rates of environmental change on character evolution; and the nature of evolution under different fitness landscapes. 
+ TREvoSim employs a selection-driven, agent-based approach: it incorporates key elements of biological evolution (selection, reproduction, heritability and mutation). The (true) phylogenetic trees and character data are an emergent property of a TREvoSim simulation, and as such the software is particularly well suited to simulation studies that can be analysed through phylogenetic trees and character data matrices. These include, for example: the impact of missing data on phylogenetic inference; the impact of rates of environmental change on character evolution; and the nature of evolution under different fitness landscapes. This complements other simulation approaches where, typically, phylogenetic trees or character data are simulated using stochastic tools such as birth-death models [e.g. @Guillerme_2024] or data based on random numbers [e.g. @Puttick_O_Reilly_Pisani_Donoghue_2019] that do not incorporate, for example, selection acting on individuals. The data TREvoSim generates are different in a number of ways to those created using stochastic models [@Keating_Sansom_Sutton_Knight_Garwood_2020], and are also likely to violate the assumptions of models commonly used in phylogenetic inference. Incorporating a level of model misspecification resembling that expected from empirical datasets is desirable in simulation studies that assess the efficacy of inference methods. Given the complexity of morphological evolution, the subsequent impact of character coding, and our current understanding of the patterns present in empirical character data, it is challenging to demonstrate, beyond the inclusion of empirically grounded concepts in its generation, the naturalism of TREvoSim data. The default settings have been chosen to reflect a number of features of empirical data matrices and trees to try and minimise the mismatch between TREvoSim and real world data, however, there are a broad range of potential alternative means of quantifying outputs. Which of these is most appropriate is likely to depend on the area of study and specific question at hand, and as such, TREvoSim provides granular control over the simulation parameters, allowing users to generate data that best serve their needs. TREvoSim is intended as a versatile platform that might be used to study a broad range of topics. 
 
 # Current associated projects 
 
