### 2024-07-13 palaeoware/trevosim#55 @RussellGarwood
Thanks for all the comments @ms609 - they are really helpful, and indeed, have been the basis of a lot of thought since you have posted them. I have distilled those thoughts into a rewritten statement of need, which I hope addresses them, and clarifies numerous aspects of where I think (hope) TREvoSim may be useful. Specific points:

-- The Statement of Need feels like it describes the TREvoSim software as a whole

Indeed, this is as intended. Given this is the first paper solely dedictated to TREvoSim, it felt appropriate to take this approach, so hopefully JOSS and @fboehm are OK with this!

--  It's not quite clear what a "phylogenetic simulation" is here; do you have in mind the simulation of character data, a tree (which it itself a form of data), or both?

I think the new statement is clearer on the kinds of data I'm trying to talk about throughout - usually both trees and characters.

--  TREvoSim uses stochastic / random processes to generate data; its data are not deterministic. These sentences do not do enough to distinguish [data simulated within] Trevosim from other approaches.

This was poorly phrased, and my thoughts were not clearly expressed on the page. Apologies. What I was getting at is that an important difference between this and other approaches is that this has stochasticity, but that it is filtered through selection, whereas e.g. random data or a birth deah model don't have this additional filter.

--  More substantively, it would be nice to somehow demonstrate that the nature of misspecification is similar to and equivalent to that of empirical datasets  / I think what's really at stake here is whether TRevoSim data is more "realistic" than data simulated by other approaches, and the data validation offers a good test of this. (It would be interesting to compare the properties of TRevosim simulations with data simulated by other means).

This is a really nice point, and I don't think my statement did a good job getting across my take on this. The accuracy/naturalism/realism of TREvoSim data is not a message I would particularly like to push too hard. I hope that by incorporating - to quote a new addition to the paper in light of this series of points, "key elements of biological evolution (reproduction, heritability and mutation)" - its data may be more comparable to empirical data than if it did not. However, I want to be careful about claims of accuracy because there is so much we haven't qunatified yet in empirical data! Rather than implying this is necessarily a *good* proxy for real data I tend to think of it as equivalent to striving towards a potentially more suitable one hatn alternatives. A parallel (if we knew the true nature of our empriical data) would be aiming for for less inacuraccy as opposed to a high level of accuracy. If we use models that are similar to generate data and then infer trees, we might expect (I think, correctly) those to perform better than they would in a situation where the generating model is significantly different, no matter what those specific differences are. Hence TREvoSim provides one option (alongside others you mention) for doing creating such data that we do expect to be mismatched. That can still be better than using than a model without the mismatch, even if it is not a brilliant proxy for real world data. Thus I don't think the difficulty in validating all aspects of simulated data to real world data invalidates its use in this kind of context.

On the basis of your comment, I don't think I had communicated this particularly well (and indeed, it could be that I had not thought this through in as much depth as I needed to be able to clarify these points until your comment forced me to do so - thanks!). This is possibly because I *do* think TREvoSim is nice (and I hope I'm not wasting my time on writing it)  because incorporates some key biological concepts I mentioned before, and I wasn't clear enough that it is hard to then assess whether this makes it a better proxy for empricial data. I have rewritten all of the appropriate parts in the statement of need to try and better communicate these themes.

I do note that down the line, I would love to do a more thorough investigation of the nature of data simulated under different models, and comparing all of that to empirical data (but I don't think this is the venue for that). If you would be interested in this too, we should talk about it further sometime.

--  This to me is the unique selling point of TREvoSim, and the SoN could do more to emphasize how this allows a fundamentally different nature of question to be addressed (and perhaps to lead with this).

I have restructured to lead with this, thanks.
