### 2024-07-13 fbe5610
Fix a typo in the manuscript, update funding information
