### 2024-07-13 1a76559
Add most pertinent references for issue #57
