+++ JOSS/paper.bib
@@ -1,4 +1,6 @@
  @article{Barido-Sottani_Saupe_Smiley_Soul_Wright_Warnock_2020, title={Seven rules for simulations in paleobiology}, volume={46}, ISSN={0094-8373, 1938-5331}, DOI={10.1017/pab.2020.30}, abstractNote={Simulations are playing an increasingly important role in paleobiology. When designing a simulation study, many decisions have to be made and common challenges will be encountered along the way. Here, we outline seven rules for executing a good simulation study. We cover topics including the choice of study question, the empirical data used as a basis for the study, statistical and methodological concerns, how to validate the study, and how to ensure it can be reproduced and extended by others. We hope that these rules and the accompanying examples will guide paleobiologists when using simulation tools to address fundamental questions about the evolution of life.}, number={4}, journal={Paleobiology}, publisher={Cambridge University Press}, author={Barido-Sottani, Joëlle and Saupe, Erin E. and Smiley, Tara M. and Soul, Laura C. and Wright, April M. and Warnock, Rachel C. M.}, year={2020}, month=nov, pages={435–444}, language={en} }
+ @article{Butterfield_2011, title={Animals and the invention of the Phanerozoic Earth system.}, volume={26}, DOI={10.1016/j.tree.2010.11.012}, abstractNote={Animals do not just occupy the modern biosphere, they permeate its structure and define how it works. Their unique combination of organ-grade multicellularity, motility and heterotrophic habit makes them powerful geobiological agents, imposing myriad feedbacks on nutrient cycling, productivity and environment. Most significantly, animals have “engineered” the biosphere over evolutionary time, forcing the diversification of, for example, phytoplankton, land plants, trophic structure, large body size, bioturbation, biomineralization and indeed the evolutionary process itself. This review surveys how animals contribute to the modern world and provides a basis for reconstructing ancient ecosystems. Earlier, less animal-influenced biospheres worked quite differently from the one currently occupied, with the Ediacaran-Cambrian radiation of organ-grade animals marking a fundamental shift in macroecological and macroevolutionary expression.}, number={2}, journal={Trends in ecology & evolution}, author={Butterfield, Nicholas J}, year={2011}, month=feb, pages={81–7} }
+ @article{Condamine_Rolland_Morlon_2013, title={Macroevolutionary perspectives to environmental change}, volume={16}, ISSN={1461-0248}, DOI={10.1111/ele.12062}, abstractNote={Ecology Letters is a broad-scope ecology journal considering all taxa, in any biome and geographic area, and spanning community, microbial & evolutionary ecology.}, journal={Ecology Letters}, publisher={John Wiley & Sons, Ltd}, author={Condamine, Fabien L. and Rolland, Jonathan and Morlon, Hélène}, year={2013}, month=jan, pages={72–85}, language={en} }
[…]
