### 2024-07-13 palaeoware/trevosim#57 @RussellGarwood
Thanks for these suggestions - I clarified the first point with the above commit (and this has since been further modifed in my changes in light of #55), and I was pleased for the opportunity to add some key references for future directions. Because  the JOSS format encourages a concise reference list, I limited these to the points where I thought they might be most useful (some areas I could potentially have rerefenced are so vast that I didn't really know where to start) - but the additions clearly do help link this to the ideas already out there. Cheers!
