### 2024-07-15 palaeoware/trevosim#55 @ms609
Hi Russell, glad that the comments have been useful.  The revised manuscript is much improved.  A few follow-up comments:

> with three masks in a given environment, the first bit may
78 be 1,0,0 for masks 1,2 and 3 respectively – when this option is enabled, that may be moved to
79 bit three between the first and second environment, and this is repeated for all sites.

I think follow this, after a few readings – I think this is saying that when <the match peaks> option is enabled, <the pattern 1, 0, 0> may be <assigned?> to <the third bit> in the second environment.  In general, I often find that the subject of 'this' or 'that' can be unclear, so it may be worth specifying more explicitly in the text.

I initially misread the sense due to the coincidence between "bit three" and the fact there are three masks listed.  As the bit chosen in the example is arbitrary, perhaps choose an integer that hasn't been used for something else (bit 7, say).  Using the same terminology ("first bit / bit 1" vs "seventh bit / bit 7") to refer to bit position might also avoid confusion.

Perhaps rephrasing to "The first bit in masks 1, 2 and 3 may be 1, 0, and 0" would also be clearer, as I got confused when squaring the singular `bit` with the three entries `1, 0, 0`.

What is not clear to me from this description is that shuffling the site patterns in the environment has any tangible difference from shuffling the site patterns in the observed genomes at the end of the simulation.  Functionally, I don't see how the 'match peaks' is intended to differ from simply performing _n_ replicates of an identical starting environment.

> [The stochastic layer] removes direct control of fitness from the organism genome

This feels slightly too strong; the organism genome is still the primary control on fitness – i.e. fitness does not change unless the genome does – but the strength of the relationship is less prominent.  I feel like another half-sentence spelling out many-to-one mapping would help the reader here; what I understand from this is that the stochastic layer is something akin to a codon table where several distinct genome entries may correspond to an equivalent fitness level.

>  I would love to do a more thorough investigation of the nature of data simulated under different models, and comparing all of that to empirical data

This would certainly be interesting to explore; the challenge I suspect will come in defining the properties of empirical data in a quantifiable way. Mulvey et al. have a [preprint](https://dx.doi.org/10.1101/2024.01.25.577179) that makes an important contribution here, and I think there'd be mileage in continuing the exploration of potential test statistics.
