### 2024-07-15 palaeoware/trevosim#54 @ms609
Looks great – you accomplish a lot with few words!

I had to read one sentence twice to understand that the "inference methods used" are not used by TREvoSim; perhaps removing 'the' from 'allowing assessment of <strike>the</strike> inference methods used' would clarify this?
