### 2024-07-23 6dcfe99
Add clarification to the summary as per comment on issue #54
