### 2024-07-23 palaeoware/trevosim#54 @RussellGarwood
Thanks - tried to clarify with above push!
