### 2024-07-23 ae34039
Update text in light of comments on issue #55
