### 2024-07-23 e092613
update copyright in docs
