### 2024-07-23 fd5deda
Update readme to provide further details about contributing to the code for issue #12
