### 2024-07-23 palaeoware/trevosim#12 @RussellGarwood
Thanks for these pointers, much aprpeciated! I have now standardised the copyright statments, followed the CLA approach you used in TreeTools (thanks for this!), and also added acknowledgements to contributions to the readme.md. I also took the opportunity to add a code of conduct to the repo contributors repository.
