### 2024-07-23 palaeoware/trevosim#16 @RussellGarwood
Apologies, we have not been able to reproduce the lack of status message on any windows build on this end - as such we're at a bit of a loss as to what to do about it!
