### 2024-07-23 palaeoware/trevosim#55 @RussellGarwood
Thanks for the continued feedback on this front @ms609 - it is, as ever, appreciated.

-- I think follow this, after a few readings – I think this is saying that when option is enabled, <the pattern 1, 0, 0> may be <assigned?> to in the second environment. In general, I often find that the subject of 'this' or 'that' can be unclear, so it may be worth specifying more explicitly in the text.

I've tried to clarify this further by consistently referring to sites, rather than "bits" when the former is what I actually mean. This was sloppy terminology on my part, and I hope the distinction I have made makes it easier to follow.

-- What is not clear to me from this description is that shuffling the site patterns in the environment has any tangible difference from shuffling the site patterns in the observed genomes at the end of the simulation. Functionally, I don't see how the 'match peaks' is intended to differ from simply performing n replicates of an identical starting environment.

Shuffling the patterns in the environment will ensure that the gnome(s) capable of peak fitness differ between environments at the start of the run, but that they will have the same absolute value in terms of fitness. As such the hope is that in the early stages of a simulation, the evolutionary trajectories of the early population will not be weighted towards adapting to just one of multiple environments. What happens at the end of the simulation is the result of the evolution within the system as a simulation has run, so essentially we can try and start assessing evolutionary outcomes based on the adaptive landscape at the start of a simulation with a reasonable hope that the early evolution, at least, is representative of that adaptive landscape (it proved difficult to keep up the matched peaks approach without lots of user intervention). So if more environments == flatter landscape, however you want to define that, we can maximise the chances of the outcomes representing that, rather than just the population adapting to the environment which has the highest possible fitness peak.

-- This feels slightly too strong

Yes, I guess this depends on your reading of the word "direct". I have reworded to try and be more explicit as you suggest!

-- This would certainly be interesting to explore; the challenge I suspect will come in defining the properties of empirical data in a quantifiable way. Mulvey et al. have a preprint that makes an important contribution here, and I think there'd be mileage in continuing the exploration of potential test statistics.

Indeed, there is so much in emprical data that we don't have a full - or even - partial handle on. Thanks for the link to this preprint, it looks like important reading!
