### 2024-07-24 9c96947
Implement simple command line function for issue #31
