### 2024-07-26 c9ff6a1
Some tweaks to help with building on Mac
