+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.16)
+cmake_minimum_required(VERSION 3.16...3.16)
 
 set(TREVOSIM_VERSION 3.0.0)
 set(TREVOSIM_APP TREvoSim)
@@ -41,16 +41,20 @@ include_directories("${gtest_SOURCE_DIR}/include"
                     "${gmock_SOURCE_DIR}/include"
 )
 
-# The MACOSX_BUNDLE_ICON_FILE variable is added to the Info.plist
-# generated by CMake. This variable contains the .icns file name,
-# without the path.
-set(MACOSX_BUNDLE_ICON_FILE icon.icns)
-# And the following tells CMake where to find and install the file itself.
-set(app_icon_macos "${CMAKE_CURRENT_SOURCE_DIR}/resources/icon.icns")
-set_source_files_properties(${app_icon_macos} PROPERTIES MACOSX_PACKAGE_LOCATION "Resources")
+if (APPLE)
+    # The MACOSX_BUNDLE_ICON_FILE variable is added to the Info.plist
+    # generated by CMake. This variable contains the .icns file name,
+    # without the path.
+    set(MACOSX_BUNDLE_ICON_FILE icon.icns)
+    # And the following tells CMake where to find and install the file itself.
+    set(app_icon_macos "${CMAKE_CURRENT_SOURCE_DIR}/resources/icon.icns")
+    set_source_files_properties(${app_icon_macos} PROPERTIES MACOSX_PACKAGE_LOCATION "Resources")
+endif()
 
-# Set the icon for Windows
-set(app_icon_resource_windows "${CMAKE_CURRENT_SOURCE_DIR}/resources/trevosim.rc")
+if(WIN32)
+    # Set the icon for Windows
+    set(app_icon_resource_windows "${CMAKE_CURRENT_SOURCE_DIR}/resources/trevosim.rc")
+endif()
 
 # Add QT required components
 find_package(Qt6 REQUIRED COMPONENTS Core Gui Widgets Concurrent Core5Compat)
@@ -81,7 +85,6 @@ set(app_sources
 set( UI
     about.ui  batchdialog.ui  mainwindow.ui  output.ui  settings.ui
 )
-
 
 # Defines the main TREvoSim executable
 qt_add_executable(${TREVOSIM_APP}
