### 2024-07-26 palaeoware/trevosim#55 @ms609
Great, the match peaks now makes sense, and your explanation has helped me to get my head around the motivation.

I can follow §Stochastic Layer now, but it's still a little obscure; "to create a 1 bit in the genome calculation" is still a slightly tricky phrasing to understand, though I can see what you are getting at. And there's a couple of typos in "whilst are 1100 may mao to 0".
