### 2024-07-26 palaeoware/trevosim#63 @ms609
Feature validation
Per https://github.com/openjournals/joss-reviews/issues/6722

I've been working through the reviewer checklist – almost all the boxes are now ticked!

I am stuck on the **Automated tests** point, "the functionality of the software can be verified".  I couldn't see an obvious way to validate the Match Peaks and No Selection features.  Are these covered by existing tests, or is there a manual way to evaluate this?
