### 2024-07-26 8075d3e
Updates the MacOS build instructions
- Updates the CMake.txt file to separate out the Windows and MacOS instructions more clearly.
- Update the MacOS build instructions.
- Moves the MacOS and Windows build utility files to a new folder.
- Adds instructions and custom bash script to generate MacOS .dmg installer.
