+++ docs/buildingfromsource.rst
@@ -72,7 +72,7 @@ QT Creator + QT v6.x using Qt installer and MSVC2019 (64-bit).
 #. [Optional] To package the application as Windows Installation Binary for distribution we recommend using Inno Setup v6.x: https://jrsoftware.org/isinfo.php
 
    * Inno Setup is a free installer for Windows programs by Jordan Russell and Martijn Laan. 
-   * The file TREvoSim_Installer_v3.x.x_dev.iss under the ./inno_setup folder is provided as a template. Copy this file to the deploy folder you created earlier.
+   * The file TREvoSim_Installer_v3.x.x_dev.iss under the ./build_utils/windows/inno_setup folder is provided as a template. Copy this file to the deploy folder you created earlier.
    * Open this file in InnoSetup - if you associated .ISS files with InnoStup during its installation you will be able to double click this file. It will look similar to this:
   
    .. figure:: _static/innosetup.png
@@ -193,23 +193,70 @@ Compiling v1.X, v2.X, v3.X using Qt Creator
 MacOS 
 ^^^^^
 
-QT Creator + QT v6.x using Clang and xcode.
+QT Creator + QT v6.x using Clang and Xcode.
 
 For v1.X, 2.X, v3.X
 """""""""""""""""""
 
-1. Xcode can be downloaded from the Apple Store or Apple's developer website (including older versions of Xcode): https://developer.apple.com/download/. We recommend always using the latest Xcode available from Apple that has been tested with the Qt version you are using.
-2. Install Qt 6.X and Qt Creator on your system by running the installer from Qt: https://www.qt.io/download. An alternative may be to install via homebrew: 
+#. Xcode can be downloaded from the Apple Store or Apple's developer website (including older versions of Xcode): https://developer.apple.com/download/. We recommend always using the latest Xcode available from Apple that has been tested with the Qt version you are using.
+#. Install Qt 6.X and QT Creator on your system by running the open source development installer from Qt: https://www.qt.io/download-open-source. An alternative may be to install via homebrew: 
 
 .. code-block:: console
 
   $ brew install qt
 
-3. To build the software and test suite, download source code, launch Qt Creator, and open the CMakeLists.txt file.
-4. Configure build and follow standard debug/release procedure.
+#. Download or clone using Git the TREvoSim source code from the GitHub repository: https://github.com/palaeoware/trevosim
+#. Run QT Creator and configure the build pipeline. QT Creator should automatically detect your Xcode installation and configure the build chain. However, if you have multiple versions installed - or if you get errors in the steps subsequent to this one, you might need to manually change QT creator to use the desired Xcode Toolkit. To do this, select projects in the left hand toolbar of Qt Creator (ctrl + 5), and then select the kit labelled with the version of Qt you installed, plus "for macos" under the 'Build and Run' section.
+#. In QT Creator using the main menu 'File |rarr| Open File or Project…' navigating to and select the TREvoSim CMakeLists.txt file in the root of the source code folder. This will open the TREvoSim project within the workspace. 
+#. From QT Creator you can build debug and release version of TREvoSim. The easiest way is via the bottom left 'Build' button. Alternatively you can use the 'Build |rarr| Build All Projects…' from the main menu.
+#. Once you have built the release version of TREvoSim through QT Creator, and you if wish to make the binary useable outside of the QT Creator environment (either as a standalone distribution or to package in a .dmg), you will need to run the Qt MacOS Deployment tool. This tool automates the process of creating a deployable applicaion package that contains all QT-related dependencies to make TREvoSim run (e.g., .dysys libraries).
+
[…]
