+++ docs/buildingfromsource.rst
@@ -209,14 +209,14 @@ For v1.X, 2.X, v3.X
 #. Run QT Creator and configure the build pipeline. QT Creator should automatically detect your Xcode installation and configure the build chain. However, if you have multiple versions installed - or if you get errors in the steps subsequent to this one, you might need to manually change QT creator to use the desired Xcode Toolkit. To do this, select projects in the left hand toolbar of Qt Creator (ctrl + 5), and then select the kit labelled with the version of Qt you installed, plus "for macos" under the 'Build and Run' section.
 #. In QT Creator using the main menu 'File |rarr| Open File or Project…' navigating to and select the TREvoSim CMakeLists.txt file in the root of the source code folder. This will open the TREvoSim project within the workspace. 
 #. From QT Creator you can build debug and release version of TREvoSim. The easiest way is via the bottom left 'Build' button. Alternatively you can use the 'Build |rarr| Build All Projects…' from the main menu.
-#. Once you have built the release version of TREvoSim through QT Creator, and you if wish to make the binary useable outside of the QT Creator environment (either as a standalone distribution or to package in a .dmg), you will need to run the Qt MacOS Deployment tool. This tool automates the process of creating a deployable applicaion package that contains all QT-related dependencies to make TREvoSim run (e.g., .dysys libraries).
+#. Once you have built the release version of TREvoSim through QT Creator, and you if wish to make the binary useable outside of the QT Creator environment (either as a standalone distribution or to package in a .dmg), you will need to run the Qt MacOS Deployment tool. This tool automates the process of creating a deployable applicaion package that contains all QT-related dependencies to make TREvoSim run (e.g., QT libraries).
 
    * First step is to copy to a new folder (we called ours 'deploy' and will refer to this by name) the newly generated binaries (TREvoSim.app and TREvoSimTest.app) from your build folder. The build folder should be called 'build\\Desktop_Qt_6_6_2_for-macOS-Release' if using the QT defaults, where Qt_6_6_2 will change to match your current QT 6 version.
    * From the terminal command line '''cd''' into the '{path to your QT installation}/{QT 6 version}/macos/bin/' folder containing the macdeployqt6 binary and run the following commands:
 	
    .. code-block:: console
 	
-	 ./macdeployqt6 {full path to TREvoSim binary}/TREvoSim.app -verbose=2
+	  $ ./macdeployqt6 {full path to TREvoSim binary}/TREvoSim.app -verbose=2
 	
 
    * You can view the contents of the application package by right clicking the TREvoSim.app and selecting 'Show Package Contents'. The folder structure should now look like this:
@@ -232,7 +232,7 @@ For v1.X, 2.X, v3.X
 
    .. code-block:: console
 
-   ./macdeployqt6 {full path to TREvoSim binary}/TREvoSim.app -dmg -verbose=2
+    $ ./macdeployqt6 {full path to TREvoSim binary}/TREvoSim.app -dmg -verbose=2
 
    2. Using a seperate .dmg packaging tool and the generate_dmg bash script: 
    
@@ -240,15 +240,15 @@ For v1.X, 2.X, v3.X
    
     .. code-block:: console
 
-    brew install create-dmg
+      $ brew install create-dmg
 
     * The .dmg installer package can then be generated by running the following commands:
 
     .. code-block:: console
 
-    cd {full path to TREvoSim repo}/build_utils/macos/
-    chmod 755 generate_dmg
-    ./generate_dmg
+      $ cd {full path to TREvoSim repo}/build_utils/macos/
+      $ chmod 755 generate_dmg
+      $ ./generate_dmg
 
     * The .dmg will be created in the 'deploy' folder and will be called 'TREvoSim_Installer.dmg'.
     * Note you may need to edit the variables in the generate_dmg bash script to match your folder structure.
