### 2024-07-26 palaeoware/trevosim#63 @RussellGarwood
Oh, thanks for catching this - I suspect both are missing having traled through the git history on the development repo, though I thought I had done them at some point. In theory you should be able to check both of these using the complete working log that outputs everything TREvoSim does, but this is unweildy and it would be much better to have tests - I shall implement some in coming days!
