### 2024-07-26 166381a
Fix some typos in the new building from source instructions, standardise numbering to be automated and fix indentation to render properly throughout£
