+++ docs_build/_sources/buildingfromsource.rst.txt
@@ -72,7 +72,7 @@ QT Creator + QT v6.x using Qt installer and MSVC2019 (64-bit).
 #. [Optional] To package the application as Windows Installation Binary for distribution we recommend using Inno Setup v6.x: https://jrsoftware.org/isinfo.php
 
    * Inno Setup is a free installer for Windows programs by Jordan Russell and Martijn Laan. 
-   * The file TREvoSim_Installer_v3.x.x_dev.iss under the ./inno_setup folder is provided as a template. Copy this file to the deploy folder you created earlier.
+   * The file TREvoSim_Installer_v3.x.x_dev.iss under the ./build_utils/windows/inno_setup folder is provided as a template. Copy this file to the deploy folder you created earlier.
    * Open this file in InnoSetup - if you associated .ISS files with InnoStup during its installation you will be able to double click this file. It will look similar to this:
   
    .. figure:: _static/innosetup.png
@@ -124,30 +124,30 @@ Ubuntu
 Compiling v3.X using command line
 """""""""""""""""""""""""""""""""
 
- This has been tested on Ubuntu 22.04 64-bit with QT v6.x using GCC (64-bit).
+This has been tested on Ubuntu 22.04 64-bit with QT v6.x using GCC (64-bit).
 
-1. Install GCC and Qt using system packages:
+#. Install GCC and Qt using system packages:
 
-.. code-block:: console
+    .. code-block:: console
 
-  sudo apt-get install build-essential libgl1-mesa-dev
-  sudo apt install qt6-base-dev libqt6core5compat6-dev
+      sudo apt-get install build-essential libgl1-mesa-dev
+      sudo apt install qt6-base-dev libqt6core5compat6-dev
 
 
-2. Download source code and navigate to folder, or alternatively clone using Git:
+#. Download source code and navigate to folder, or alternatively clone using Git:
 
-.. code-block:: console
+    .. code-block:: console
 
-  git clone https://github.com/palaeoware/trevosim.git
-  cd trevosim
+      git clone https://github.com/palaeoware/trevosim.git
+      cd trevosim
 
-3. Within TREvoSim folder, run the following command to build the software:
+#. Within TREvoSim folder, run the following command to build the software:
 
-.. code-block:: console
+    .. code-block:: console
 
- cmake --build .
+      cmake --build .
 
-4. Launch the software by double clicking on the TREvoSim binary that has been created in this folder.
+#. Launch the software by double clicking on the TREvoSim binary that has been created in this folder.
 
 
 Compiling v1.X & v2.X using command line
@@ -155,61 +155,108 @@ Compiling v1.X & v2.X using command line
 
 This has been tested on Ubuntu 18.04 and 2204, with QT v5.x using GCC (64-bit).
 
-1. Install GCC and Qt using system packages:
+#. Install GCC and Qt using system packages:
 
-.. code-block:: console
+    .. code-block:: console
 
-  sudo apt-get install build-essential libgl1-mesa-dev
-  sudo apt install qt5-default
+      sudo apt-get install build-essential libgl1-mesa-dev
+      sudo apt install qt5-default
 
-2. Download source code and navigate to folder, or alternatively clone using Git:
+#. Download source code and navigate to folder, or alternatively clone using Git:
 
-.. code-block:: console
+    .. code-block:: console
 
-  git clone https://github.com/palaeoware/trevosim.git
-  cd trevosim
+      git clone https://github.com/palaeoware/trevosim.git
+      cd trevosim
 
-3. Within REvoSim folder create makefile:
+#. Within REvoSim folder create makefile:
 
-.. code-block:: console
+    .. code-block:: console
 
-  qmake ./trevosim.pro
+      qmake ./trevosim.pro
 
[…]
