+++ docs/buildingfromsource.rst
@@ -124,30 +124,30 @@ Ubuntu
 Compiling v3.X using command line
 """""""""""""""""""""""""""""""""
 
- This has been tested on Ubuntu 22.04 64-bit with QT v6.x using GCC (64-bit).
+This has been tested on Ubuntu 22.04 64-bit with QT v6.x using GCC (64-bit).
 
-1. Install GCC and Qt using system packages:
+#. Install GCC and Qt using system packages:
 
-.. code-block:: console
+    .. code-block:: console
 
-  sudo apt-get install build-essential libgl1-mesa-dev
-  sudo apt install qt6-base-dev libqt6core5compat6-dev
+      sudo apt-get install build-essential libgl1-mesa-dev
+      sudo apt install qt6-base-dev libqt6core5compat6-dev
 
 
-2. Download source code and navigate to folder, or alternatively clone using Git:
+#. Download source code and navigate to folder, or alternatively clone using Git:
 
-.. code-block:: console
+    .. code-block:: console
 
-  git clone https://github.com/palaeoware/trevosim.git
-  cd trevosim
+      git clone https://github.com/palaeoware/trevosim.git
+      cd trevosim
 
-3. Within TREvoSim folder, run the following command to build the software:
+#. Within TREvoSim folder, run the following command to build the software:
 
-.. code-block:: console
+    .. code-block:: console
 
- cmake --build .
+      cmake --build .
 
-4. Launch the software by double clicking on the TREvoSim binary that has been created in this folder.
+#. Launch the software by double clicking on the TREvoSim binary that has been created in this folder.
 
 
 Compiling v1.X & v2.X using command line
@@ -155,40 +155,40 @@ Compiling v1.X & v2.X using command line
 
 This has been tested on Ubuntu 18.04 and 2204, with QT v5.x using GCC (64-bit).
 
-1. Install GCC and Qt using system packages:
+#. Install GCC and Qt using system packages:
 
-.. code-block:: console
+    .. code-block:: console
 
-  sudo apt-get install build-essential libgl1-mesa-dev
-  sudo apt install qt5-default
+      sudo apt-get install build-essential libgl1-mesa-dev
+      sudo apt install qt5-default
 
-2. Download source code and navigate to folder, or alternatively clone using Git:
+#. Download source code and navigate to folder, or alternatively clone using Git:
 
-.. code-block:: console
+    .. code-block:: console
 
-  git clone https://github.com/palaeoware/trevosim.git
-  cd trevosim
+      git clone https://github.com/palaeoware/trevosim.git
+      cd trevosim
 
-3. Within REvoSim folder create makefile:
+#. Within REvoSim folder create makefile:
 
-.. code-block:: console
+    .. code-block:: console
 
-  qmake ./trevosim.pro
+      qmake ./trevosim.pro
 
-4. Build by running the make command:
+#. Build by running the make command:
 
-.. code-block:: console
+    .. code-block:: console
 
-  make
+      make
 
-5. Navigate to bin folder (e.g. trevosim/bin) and launch software by double clicking on file.
+#. Navigate to bin folder (e.g. trevosim/bin) and launch software by double clicking on file.
 
 
 Compiling v1.X, v2.X, v3.X using Qt Creator
 """""""""""""""""""""""""""""""""""""""""""
 
-1. Install Qt 5.X (v1.X, v2.X) or Qt 6.X (v3.X) on your system by running the installer from Qt: https://www.qt.io/download
-2. Download source code, launch Qt Creator, and open the CMakeLists.txt file. Configure build and follow standard debug/release procedure.
[…]
