+++ docs_build/_sources/buildingfromsource.rst.txt
@@ -48,19 +48,18 @@ QT Creator + QT v6.x using Qt installer and MSVC2019 (64-bit).
    * First step is to copy to a new folder (we called ours 'deploy' and will refer to this by name) the newly generated binaries (TREvoSim.exe and TREvoSimTest.exe) from your build folder. The build folder should be called 'build\\Desktop_Qt_6_6_1_MSVC2019_64bit-Release' if using the QT defaults, where Qt_6_6_1 will change to match your current QT 6 version.
    * From the windows command line '''cd''' into the '{path to your QT installation}\\{QT 6 version}\\msvc2019_64\\bin\\' folder containing the windeployqt6.exe binary and run the following commands:
 	
-   .. code-block:: console
-	
-	 qtenv2.bat
-	 "C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\VC\Auxiliary\Build\vcvars64.bat"
+      .. code-block:: console
+      
+        qtenv2.bat
+        "C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\VC\Auxiliary\Build\vcvars64.bat"
 	 
    
-   where 'C:\\Program Files (x86)\\Microsoft Visual Studio' is your path to the installed VS2019.
+    (where 'C:\\Program Files (x86)\\Microsoft Visual Studio' is your path to the installed VS2019)
 	
-   .. code-block:: console
+        .. code-block:: console
+        
+          windeployqt6 {full path to TREvoSim binary}\TREvoSim.exe  --release
 	
-	 windeployqt6 {full path to TREvoSim binary}\TREvoSim.exe  --release
-	
-
    * Your deploy folder should now look like this:
 	
    .. figure:: _static/post_windeploy.png
@@ -252,7 +251,7 @@ For v1.X, 2.X, v3.X
 
     * The .dmg will be created in the 'deploy' folder and will be called 'TREvoSim_Installer.dmg'.
     * Note you may need to edit the variables in the generate_dmg bash script to match your folder structure.
-    * On opening the .dmg installer package you will be created by the following drag n' drop installation window:
+    * On opening the .dmg installer package you will see the following drag n' drop installation window:
 	
    .. figure:: _static/installer_macdmg.png 
 
