+++ docs/buildingfromsource.rst
@@ -270,6 +270,6 @@ TREvoSim |gt| = v3 come with a suite of development tests that verify the core s
       
         $ ./TREvoSimTest.app/Contents/MacOS/TREvoSimTest
 
-#. From the QT Creator application. Once the test program has been build (see point 1 above) you can run the test suite with the QT Creator by going to the 'Tools |rarr| Tests |rarr| Run All Tests' from the main menu. This will build the application if not already done so and display the results in the 'Test Result' tab which normally appears at the bottom of the workspace.
+#. From the QT Creator application. Once the test program has been built (see point 1 above) you can run the test suite with the QT Creator by going to the 'Tools |rarr| Tests |rarr| Run All Tests' from the main menu. This will build the application if not already done so and display the results in the 'Test Result' tab which normally appears at the bottom of the workspace.
 
 Note: The tests that run are the same that can be called from the TREvoSim program via the 'Tests' button on the main tool bar.
