### 2024-07-26 9066963
Update test six to address issue #63 in part by adding no selection mode. This is now presented as the lasty column in the table printed to screen, and included in the tests. Also reworded outputs to reflect changes in terminology in documentation, and to avoid dividing by zero and changed an instance of incorrect list format in the docs
