+++ testinternal.cpp
@@ -30,6 +30,7 @@ testinternal::testinternal(MainWindow *theMainWindowCon)
     testList.insert(16, "Print tree");
     testList.insert(17, "Ecosystem engineers");
     testList.insert(18, "Playing field mixing");
+    testList.insert(19, "Match Peaks");
 }
 
 QString testinternal::testDescription(int testNumber)
@@ -590,12 +591,12 @@ bool testinternal::testSix(QString &outString)
     if (theMainWindow)
     {
         theMainWindow->addProgressBar(0, 100000);
-        theMainWindow->setStatus("Testing coin toss.\n");
+        theMainWindow->setStatus("Testing geometric distribution.\n");
     }
 
     bool testFlag = true;
     QTextStream out(&outString);
-    out << "Testing coin toss.\n\n";
+    out << "Testing geometric distribution (== coin toss in previous versions).\n\n";
 
     simulationVariables simSettings;
     simSettings.selectionCoinToss = 2.;
@@ -619,11 +620,11 @@ bool testinternal::testSix(QString &outString)
     QVector<int> hits2(pfSize, 0);
     QVector<int> hits3(pfSize, 0);
     QVector<int> hits4(pfSize, 0);
+    QVector<int> hits5(pfSize, 0);
 
     for (int i = 0; i < 100000; i++)
     {
-        if (theMainWindow)
-            theMainWindow->setProgressBar(i);
+        if (theMainWindow) theMainWindow->setProgressBar(i);
         hits0[x.coinToss(x.playingFields[0])]++;
         hits1[x.coinToss(x.playingFields[1])]++;
         hits2[x.coinToss(x.playingFields[2])]++;
@@ -632,14 +633,27 @@ bool testinternal::testSix(QString &outString)
 
     //Try also with a different probability
     simSettings.selectionCoinToss = 3.0;
-    for (int i = 0; i < 100000; i++)hits4[x.coinToss(x.playingFields[0])]++;
+    for (int i = 0; i < 100000; i++)
+    {
+        if (theMainWindow) theMainWindow->setProgressBar(i);
+        hits4[x.coinToss(x.playingFields[0])]++;
+    }
+
+    //And with no selection enabled, which should make these all equal
+    simSettings.noSelection = true;
+    for (int i = 0; i < 100000; i++)
+    {
+        if (theMainWindow) theMainWindow->setProgressBar(i);
+        hits5[x.coinToss(x.playingFields[0])]++;
+    }
 
     out << "<table><tr><th>Position</th><th>"
         "</th><th>Fittest at top - selection 2</th><th>"
         "</th><th>Fittest at top - selection 3</th><th>"
         "</th><th>Equal Fitness</th><th>"
         "</th><th>Alternating fitness</th><th>"
-        "</th><th>Three fitnesses</th></tr>";
+        "</th><th>Three fitnesses</th><th>"
+        "</th><th>No selection enabled</th></tr>";
 
     for (int i = 0; i < pfSize; i++)
     {
@@ -647,16 +661,23 @@ bool testinternal::testSix(QString &outString)
         out << "</td><td>&nbsp;&nbsp;&nbsp;&nbsp;</td><td>" << hits4[i];
         out << "</td><td>&nbsp;&nbsp;&nbsp;&nbsp;</td><td>" << hits1[i];
         out << "</td><td>&nbsp;&nbsp;&nbsp;&nbsp;</td><td>" << hits2[i];
-        out << "</td><td>&nbsp;&nbsp;&nbsp;&nbsp;</td><td>" << hits3[i] << "</tr>";
+        out << "</td><td>&nbsp;&nbsp;&nbsp;&nbsp;</td><td>" << hits3[i];
[…]
