+++ docs_build/_sources/algorithmconcepts.rst.txt
@@ -56,21 +56,21 @@ TREvoSim algorithm - Iteration
 
 After initialisation, a simulation can be run, either until the requested number of species has evolved, or until a user defined iteration. The simulation iteration algorithm employs the following steps, which are applied every iteration to every playing field (if there are more than one) independently.
 
-1) Calculate fitness for organisms in playing field, and sort playing field by fitness, with the fittest organisms at the top of the list. If a number of organisms have the same fitness (e.g., at initialization), these are randomly ordered.
+#. Calculate fitness for organisms in playing field, and sort playing field by fitness, with the fittest organisms at the top of the list. If a number of organisms have the same fitness (e.g., at initialization), these are randomly ordered.
 
-2) An organism is picked to be duplicated via from a geometric distribution with p = 0.5 by default (i.e. there is a 50% chance of selecting the first organism in the list, then if that is not chosen, a 50% chance of selecting the second, and so on). If the simulation reaches the end of the playing field without selecting an organism, it starts from the beginning again. Alternatively, there is the option to select a random individual from the playing field, without linking this to fitness ('No selection' mode). 
+#. An organism is picked to be duplicated via from a geometric distribution with p = 0.5 by default (i.e. there is a 50% chance of selecting the first organism in the list, then if that is not chosen, a 50% chance of selecting the second, and so on). If the simulation reaches the end of the playing field without selecting an organism, it starts from the beginning again. Alternatively, there is the option to select a random individual from the playing field, without linking this to fitness ('No selection' mode). 
 
-3) The organism selected for duplication has a user-defined chance of mutation (defined as mutations per hundred characters per iteration; default 5.0 mutations per iteration per 100 genome bits). The user can select whether deleterious mutations are accepted.
+#. The organism selected for duplication has a user-defined chance of mutation (defined as mutations per hundred characters per iteration; default 5.0 mutations per iteration per 100 genome bits). The user can select whether deleterious mutations are accepted.
 
-4) If the duplicated organism, after mutation, is sufficiently different to its character string at origination it is defined as a new species -- the amount of difference required is the user defined variable species difference (default 4). Note that if this is not the first speciation in the lineage, distance is measured from the last species to diverge from its lineage -- see discussion below. If it is a new species this is recorded, and the tree display on the graphical user interface is updated.
+#. If the duplicated organism, after mutation, is sufficiently different to its character string at origination it is defined as a new species -- the amount of difference required is the user defined variable species difference (default 4). Note that if this is not the first speciation in the lineage, distance is measured from the last species to diverge from its lineage -- see discussion below. If it is a new species this is recorded, and the tree display on the graphical user interface is updated.
 
-5) The duplicated, mutated organism is then returned to the playing field, overwriting the least fit organism in the playing field by default (or randomly selecting one of the least fit if multiple least fit organisms exist). Alternatively ('Random overwrite') a random member of the playing field can be selected for overwriting.
+#. The duplicated, mutated organism is then returned to the playing field, overwriting the least fit organism in the playing field by default (or randomly selecting one of the least fit if multiple least fit organisms exist). Alternatively ('Random overwrite') a random member of the playing field can be selected for overwriting.
 
-6) Organisms in the playing field typically represent multiple species once a simulation is running. Thus the playing field is checked each iteration, and any species that have become extinct are identified. Dy default the characters of the last surviving organism are appended to the character matrix on extinction (this is the 'Sansomian speciation' option, and ensures that if a single lineage has given birth to multiple species, the recorded character set is closest to the most recently branching terminal; when this is turned off, the characters recorded for a species are its genome at origination).
[…]
