+++ docs_build/_sources/buildingfromsource.rst.txt
@@ -227,7 +227,7 @@ For v1.X, 2.X, v3.X
    
 #. [optional] You can now generate a .dmg package to aid distribution of the TREvoSim.app. There are two ways to do this:
    
-   #. Adding the -dmg flag to the macdeploy6 command as part of the previous step:
+   #. Adding the -dmg flag to the macdeployqt6 command as part of the previous step:
 
       .. code-block:: console
 
@@ -260,7 +260,7 @@ For v1.X, 2.X, v3.X
 Development Testing
 -------------------
 
-TREvoSim |gt| = v3 come with a suite of development tests that verify the core simulator/generator code. These can be called via one of two methods during the devlopment phase:
+TREvoSim |gt| = v3 come with a suite of development tests that verify the core simulator/generator code. These can be called via one of two methods during the development phase:
 
 #. As a standalone program - for this we leverage the GoogleTest Mocking and Testing Framework (https://github.com/google/googletest) which is pulled in automatically by CMake and built alongside the TREvoSim main program - creating a separate standalone test program called TREvoSimTest. This program can be called from the command line on Mac/Linux or from the QT Creator environment on Windows.
   
@@ -270,6 +270,6 @@ TREvoSim |gt| = v3 come with a suite of development tests that verify the core s
       
         $ ./TREvoSimTest.app/Contents/MacOS/TREvoSimTest
 
-#. From the QT Creator application. Once the test program has been build (see point 1 above) you can run the test suite with the QT Creator by going to the 'Tools |rarr| Tests |rarr| Run All Tests' from the main menu. This will build the application if not already done so and display the results in the 'Test Result' tab which normally appears at the bottom of the workspace.
+#. From the QT Creator application. Once the test program has been built (see point 1 above) you can run the test suite with the QT Creator by going to the 'Tools |rarr| Tests |rarr| Run All Tests' from the main menu. This will build the application if not already done so and display the results in the 'Test Result' tab which normally appears at the bottom of the workspace.
 
 Note: The tests that run are the same that can be called from the TREvoSim program via the 'Tests' button on the main tool bar.
