### 2024-07-26 palaeoware/trevosim#63 @RussellGarwood
With above push, no selection mode is rolled into test 6, as it fundamentally works by modify the individual selected for replication. I also took the opportunity to, hopefully, clarify the reporting of the test, but given its nature  I suspect it will remain hard to follow without looking at the code!
