### 2024-07-27 c30abbc
Some modifications to count peaks to allow it to take an environment for testing purposes
