### 2024-07-31 17925bd
Add test 19 to test march peaks for issue #63. This involves changes to the count peaks function as well as the test, and is one of the slower tests, so has its own progress bar
