### 2024-08-02 8c8671d
Updated coin toss to geometric distribution to reflect documentation
