### 2024-08-06 palaeoware/trevosim#63 @RussellGarwood
Apologies, it looks like I forgot to comment on this issue (I've been travelling between Berlin and Manchester, and setting up a new computer, so it's all been a bit chaotic!) - the above commit adds a test for matching peaks by doing doing an exhaustive search and ensuring that the fittest individuals when match peaks is on have the same fitness (and that they don't across multiple repeats when it is off!).
