### 2024-08-06 70a1271
Try and calrify wording for issue #55 around many to one mapping
