### 2024-08-06 palaeoware/trevosim#55 @RussellGarwood
Thanks for catching the typos! Yeah, I have struggled with how to make this as clear as I would like - I've tried with the above push to define many to one mapping when first used, and then given the example in the hope this makes it a little clearer!
