### 2024-08-07 5b6dac0
Updated build instructions on ubuntu to include cmake, and a cmake command that is required on a fresh install
