### 2024-08-11 65b1b28
Change settings wording for mixing for issue #62
