### 2024-08-11 #6722 @RussellGarwood
Hi all! I have just closed a couple of issues, and I think those relating to these reviews are all done (lack of status messages remain a mystery @ms609 - the only solution I can think of is for me to poke around on your system to try and figure it out. Apologies).

Palaeontology, who are holding a paper back while this is in review, have been chasing this, and @alanspencer, who will be doing the mac build for a release is going on vacation towards the middle of next week - so it'd be a great help if we could do the build for a release before he disappears. Thus if you are spending any time on this before then, any further changes required in the source code would be much appreciated, allowing him to do the builds for a release before he leavers (I can action anything not requiring changes to the code after that point, of course).
