### 2024-08-11 7f1f516
Updated docs to reflecct changes for #62
