### 2024-08-12 #6722 @atcribb
Hi @fboehm and @RussellGarwood – I've completed the review and have only very minor comments. To ease response (I hope!), I've created two issues for Documentation and Software Paper and linked them in the checklist where appropriate.
