### 2024-08-14 1c8827e
Upodated text under the About dialog
- update of text under the About dialog window.
