### 2024-08-15 5542f18
Fixes for the Windows console output
- Not a perfect way of doing this but the new code allows Windows builds to output to the console windows the application was called from.
