### 2024-08-15 0643ba4
Switch out Butterfield 2011 for Jones et al 1994 for issue #66
