### 2024-08-15 6431d1a
Clarify the real world implications of expanding playing field for issue #66
