+++ JOSS/paper.md
@@ -97,7 +97,7 @@ A new ecosystem engineering system allows the impact of organism-environment fee
 
 ### Expanding playing field 
 
-This new option alters competition such that it occurs between species rather than between individuals (i.e. it removes intra-species competition from the simulation). When enabled (disabled by default), this is achieved by allowing only one individual from each species to be present in the playing field at any time. As such, the playing field grows to accommodate new species, which appear following the standard speciation rules. On duplication of an individual, juveniles overwrite the previous member of their species in the playing field. Otherwise, the playing field operates as normal, i.e. it is ordered by fitness and duplication of an individual selected using a geometric distribution links fitness to fecundity.  
+This new option alters competition such that it occurs between species rather than between individuals. When enabled (disabled by default), this is achieved by allowing only one individual from each species to be present in the playing field at any time. As such, the playing field grows to accommodate new species, which appear following the standard speciation rules. On duplication of an individual, juveniles overwrite the previous member of their species in the playing field. Otherwise, the playing field operates as normal, i.e. it is ordered by fitness and duplication of an individual selected using a geometric distribution links fitness to fecundity. Enabling the expanding playing field removes intra-species competition from the simulation: this can be used to investigate the impact that intra- v.s. inter-species competition has on evolutionary processes during a run, and on the phylogenetic outcomes of a simulation.
 
 ### Match peaks
 
