### 2024-08-15 b58878f
Clarify the real world implications of match fitness peaks for issue #66
