### 2024-08-15 d5f5276
Clarify the real world implications of no selection for issue #66
