### 2024-08-15 7c7a749
Clarify the real world implications of perturbations for issue #66
