### 2024-08-15 eb9b2c7
Update test information to reflect v3 release version
