### 2024-08-15 d1bcef8
Add statement of need to docs as per suggestion in #65, and also: reduce redundancy/repetition in the landing page and intro; improve cross referencing; and tidy up some of the code
