+++ docs/index.rst
@@ -12,7 +12,7 @@ TREvoSim is an individual-based evolutionary model, focussing on the simulation 
 
 TREvoSim has been released with the intention that it can be used as a multi-purpose platform for the study of many evolutionary phenomena, but in particular, acts as one of a limited number of methods for simulating character data and tree topology at the same time and is one of very few individual-based simulations capable of doing so. We note that, as with REvoSim, this package is complementary to the many other approaches of studying evolution on a range of different timescales, and will be continually developed by the core team to expand its capabilities.
 
-A full description of the algorithm can be found in Keating *et al.* (2020), referenced below. In brief, a simulation employs digital organisms comprising binary strings. These are used to calculate fitness through comparison to against one or several sets of random binary strings (the environment(s)), and ultimately provide the character matrices and trees output by the simulation. Organisms exist in a playing field and compete with each other; one is selected, the probability of which being based on a fitness algorithm, for reproduction each iteration. When this occurs, the organism is mutated, and returned to the playing field, overwriting the least fit organism. A new species is defined based on hamming distance to the parent species, or last species to originate within the lineage. Species are kept in a list, and their genomes are recorded on extinction. The simulation runs until the requested number of species have existed, and then writes data in the requested format.
+You can find more information in the :ref:`introduction`, and in the papers listed below. 
 
 e: `palaeoware@gmail.com <mailto:palaeoware@gmail.com>`_
 
