### 2024-08-16 79c1697
Correct error in EE documentations of single shot vs persistent EE for #65
