### 2024-08-16 10e00d1
Clarify the shockingly unclear wording of EE mechanism and frequency in the docs for issue #66
