### 2024-08-16 93ebf91
Updated fix for the Windows console running mode
