### 2024-08-16 a86d66f
Fix to make the MacOS TREvoSimTest applciation work on the console
