### 2024-08-16 bc7735f
Add new sections to docs - design and data analysis, for #66
