### 2024-08-16 0c59f97
Make limitations of increased mixing as ccurrently implemented more obvious, as per issue #62
