+++ docs_build/_sources/settings2.rst.txt
@@ -26,7 +26,7 @@ Events
 ------
 
 :Apply perturbation during run: TREvoSim offers the option of applying an environmental perturbation during a run. When this is selected, at halfway through a run (after the speciation of the n/2 the requested species, or at half the requested iteration count), all masks for all environments are overwritten with new, random, masks. If playing field masks are independent this results in new random masks for all environments on all playing fields. If masks are set to be identical across playing fields, then new masks are written to all environments, but these are identical across the playing fields. This perturbation then lasts for 10% of the iteration number at which it began, during which the masks gradually return to their pre-perturbation state (specifically, over the course of the perturbation period, 90% of bits across all masks and environments are copied back over the new masks from a copy created at perturbation initiation). As currently coded, during this period, only mutations occurring to the new, perturbation, rather than the previous environment will be applied, and overall the environmental change due to mutation will be lower than the remainder of the run (although rate of change will likely be higher as the perturbation environment is reverting to the pre-perturbation one).
-:Increase mixing for perturbation period: When mixing is enabled between playing fields, this can be accelerated during a perturbation. When this option is selected, mixing between playing fields increases by a factor of ten when a perturbation is occurring, then returns to background rates once it is complete.
+:Increase mixing for perturbation period: When mixing is enabled between playing fields, this can be accelerated during a perturbation. When this option is selected, mixing between playing fields increases when a perturbation is occurring, then returns to background rates once it is complete. Note: At present mixing is defined as a percentage probability (see below), which is multiplied by ten during the increased mixing phase. As such, for any value of mixing greater than ten, increasing mixing will result in mixing occurring once per iteration (i.e. with a 100% probability). This is likely to change in future releases to, e.g. a doubling of the rate or similar: if you are considering using this feature please contact the developers. 
 
 Playing field(s)
 ----------------
