### 2024-08-16 #6722 @RussellGarwood
@ms609 As per discussions on reviosim issue 62, I have just pushed changes to the documentation to highlight the current limitations of increased mixing, and that it is likely to change in the future. This looks like the only remaining item on your checklist, but shout if I have missed anything!
