### 2024-08-16 d1e297d
Update small detail on tests in JOSS documentation
