### 2024-08-17 2eeedcf
Update index to include alt text, tighten wording, add JOSS paper
