### 2024-08-17 142278f
Standardise some text throughout in terms of regs, add some notes, start tidying up introduction
