### 2024-08-17 da5d575
Updates ot algorithm page, including adding alt text to images, captions, and some rewording
