+++ docs/algorithmconcepts.rst
@@ -6,11 +6,11 @@ Algorithm, Concepts, and Experimental Design
 TREvoSim Algorithm Description
 ------------------------------
 
-TREvoSim has been developed across several publications, and the algorithm was first described in Keating et al. (`2020 <https://doi.org/10.1093/sysbio/syaa012>`_, v1.0.0), with modifications outlined in Mongiardino Koch et al (`2021 <https://doi.org/10.1098/rspb.2021.0044>`_, v2.0.0). The simulation approach also shares features with the spatially explicit eco-evolutionary simulation package REvoSim (Garwood et. al. `2019 <https://doi.org/10.1111/pala.12420>`_, Furness et. al. `2023 <https://doi.org/10.21105/joss.05284>`_). The latest version of the algorithm is described here to avoid having to build a picture by accessing multiple publications. Note, this is modified after those previous descriptions, and so employs the same wording in places.
+TREvoSim has been developed across several publications, and the algorithm was first described in Keating et al. (`2020 <https://doi.org/10.1093/sysbio/syaa012>`_, v1.0.0), with modifications outlined in Mongiardino Koch et al. (`2021 <https://doi.org/10.1098/rspb.2021.0044>`_, v2.0.0). The simulation approach also shares features with the spatially explicit eco-evolutionary simulation package REvoSim (`Garwood et. al. 2019 <https://doi.org/10.1111/pala.12420>`_, `Furness et. al. 2023 <https://doi.org/10.21105/joss.05284>`_). The latest version of the algorithm is described here to avoid having to build a picture by accessing multiple publications. Note: this is modified after those previous descriptions, and so employs the same wording in places.
 
 TREvoSim employs digital organisms in the form of binary strings of a user-defined length (e.g. 0010110011101010011011). The zeros and ones comprising the string are here referred to as the genome of the digital organism, and ultimately provide the binary characters that can be used at the end of a simulation for, for example, phylogenetic inference, and they are also used to assign fitness to organisms relative to an environment, as outlined below. When a simulation runs, organisms compete, replicate, and mutate. The simulation incorporates a species concept, and speciation is emergent in the simulation, allowing the software to output a phylogenetic tree showing species relationships during or at the end of a run. 
 
-This overview employs figures from Garwood et al. (2024) -- in all, user defined variables are shown in green font, and the values shown are the default. For clarity, the description eschews listing every potential modification to the algorithm that can be achieved through modifying the software settings, as this would obfuscate important functions in potentially unnecessary detail (although it does mention a number of key choices). Where a setting modifies the algorithm outlined below, this is described fully in the information included on the settings pages in this documentation. The implications of changing a setting are also described in these pages, wherever possible.
+This overview employs figures from Garwood et al. (2024) -- in all, user defined variables are shown in green font, and the values shown are the default. For clarity, the description eschews listing every potential modification to the algorithm that can be achieved through modifying the software settings, as this would obfuscate important functions in potentially unnecessary detail (although it does mention a number of key choices). Where a setting modifies the algorithm outlined below, this is described fully in the information included on the settings pages in this documentation. The implications of changing a setting are also described in those pages, wherever possible.
 
 Data structures
 ^^^^^^^^^^^^^^^
@@ -20,6 +20,9 @@ Organisms reside in a structure called the playing field -- a list of organisms 
 .. figure:: _static/data_structures.png
     :width: 400
     :align: center
+    :alt: Data structures in TREvoSim
+    
+    The data structures used in TREvoSim
 
 If so desired, simulations can employ more than one playing field, which can have the same, or different, environment(s). 
 
[…]
