### 2024-08-17 faaa38d
Updates to window layour page
