### 2024-08-17 353f3dd
Update the first settings page - in particular, change from field sets to definition lists in order to provide more space
