+++ docs/settings.rst
@@ -7,28 +7,49 @@ Clicking on the settings button of the toolbar will launch a settings dialogue w
 
 .. figure:: _static/settings_01.png
     :align: center
+    :width: 550
+    :alt: A screenshot of the TREvoSim settings panel for Organisms and simulation
+
+    The TREvoSim Organisms and Simulations settings panel
 
 Organisms
 ---------
 
-:Number of characters/genome size: This defines the number of bits in character string for each organism, and ultimately the character number of the matrix output by the software.
-:Characters for species identification: This defines the number of characters used to identify species. When this is different to the genome size, only the characters between zero and this number are used in the species delineation parts of the TREvoSim algorithm. By default (when the genome size is changed in the settings) this number will update to be the same value as the genome size, but changing it to a smaller value allows a portion of characters that are not tied to species identity.
-:Characters for fitness calculation: This defines the number of characters used in the fitness algorithm. When this is different to the genome size, on the characters between zero and this number are used to calculate fitness. By default this will update to be the same value as the genome size, but changing it to a smaller value allows for a 'non-coding' portion of the genome, with a more drift-like/stochastic behaviour (note, however, that this portion of genome is still part of an organism that is under selection).
-:Rate of organism mutation: This is the rate of mutation for the organisms in the simulation, in units of mutations per hundred characters per iteration.
-:Stochastic layer: This tick box dictates whether a stochastic layer is used. The stochastic layer is a secondary string of bits that can be used to abstract the fitness calculation from the genome and its mutations, through a mechanism based on many-to-one mapping. When this box is ticked, on settings close, another dialogue is launched that allows you to - for every possible combination of 4 bits - define an output, i.e. 0 or 1, referred to here as the map. When the setting is enabled, each organism stores an internal genome four times the length of the organism's gnome. During the simulation, mutations are applied to the stochastic layer (at the user defined rate per 100 characters), but all other operations employ the standard genome. When mutations are applied, the map is then used to convert bits from the stochastic layer into the genome that is used elsewhere (i.e. for every four stochastic bits, the output is 1 standard bit, based on the map). This abstraction process could be considered, for instance, a highly simplified parallel to e.g. ontogeny: it removes the fitness of the organism from direct control of its genome, and thus facilitates silent mutations and neutral theory of evolution type / less strongly adaptationist dynamics. Note: The stochastic layer only works currently when the character number, characters for species identification, and characters for fitness calculation are the same. As such, selecting this option sets all values to be the same as total character number, and prevents changes when selected to all but total character number. 
-:Depth of layer: In future versions, to allow further abstraction, varying the number of layers created from this map is planned (e.g. if the depth is two, then a second stochastic layer will be created from the first, the second being sixteen times the length of the organismal genome, and dictating the makeup of the first). This is not currently implemented, and so at present, this option is limited to one. 
+Number of characters/genome size
+    This defines the number of bits in character string for each organism, and ultimately the character number of the matrix output by the software.
+Characters for species identification
+    This defines the number of characters used to identify species. When this is different to the genome size, only the characters between zero and this number are used in the species delineation parts of the TREvoSim algorithm. By default (when the genome size is changed in the settings) this number will update to be the same value as the genome size, but changing it to a smaller value allows a portion of characters that are not tied to species identity.
+Characters for fitness calculation
[…]
