### 2024-08-17 2c3d7d9
Clean up logging page, including moving away from field sets and standardising indentation so parenting is clear
