+++ docs/loggingsim.rst
@@ -13,12 +13,18 @@ TREvoSim has a versatile logging system which allows the user to define outputs 
 
 .. figure:: _static/output_01.png
     :align: center
+    :alt: The output settings page for the end run log in TREvoSim
+
+    The TREvoSim End Run Log settings
 
 A typical use case for this is when you would like to use the outputs of a finished simulation to do downstream analyses for e.g. phylogenetic methods. This tab provides options for two custom log files, which are placed, after a run, within the *TREvoSim_output* folder created on the save path. Basename defines the start of the filename, and if you so wish, this can include forward slashes before the start of the file name to denote a path, that TREvoSim will create for you. TREvoSim then adds a number to the start of the filename you have suggested, and this number iterates with subsequent runs. The file extension is also defined in this dialogue in a separate text box. 
 
 The two text boxes allow custom file content to be written: text from these is written as provided to the file -- for example with run instructions or program commands -- and any of a series of keywords defined below included within two vertical bar ( | ) symbols are replaced as a file is written. The default outputs are shown in the figure above -- they create a vanilla nexus file which allows you to, for example, load the character matrix into R, and a TNT file which conducts a parsimony search on TREvoSim data before outputting the most parsimonious trees. 
 
 In addition to these two custom files, TREvoSim can output a tree for each run in a standard nexus format. The options at the bottom of the output tab allow you to define the basename for this file. The output includes the tree and translate block, as well as a comment with the settings of the run written to it. 
+
+.. note::
+    See also :ref:`Data analysis` for pointers on how to analyse TREvoSim outputs. 
 
 Running log
 ^^^^^^^^^^^
@@ -27,116 +33,140 @@ The other tab is a running log which is output during a run:
 
 .. figure:: _static/output_02.png
     :align: center
+    :alt: The output settings page for the running log in TREvoSim
+
+    The TREvoSim Running Log settings
 
 This is of utility for studying processes as they occur during a run, as it is output on a regular basis as the simulation progresses. Any text entered into the Custom running log text box is output to a new file (appended with an iterator) at the requested frequency. This can be used to record anything in the state of the simulation required to study a process of interest. At the top of the dialogue are two options:
 
-:Write running log: By default the running log is not created or written to when a simulation runs. In order to write the files this option needs to be enabled.
-:Frequency: This dictates the frequency, in iterations, with which a new running log file is written. 
+Write running log
+  By default the running log is not created or written to when a simulation runs. In order to write the files this option needs to be enabled.
+Frequency
+  This dictates the frequency, in iterations, with which a new running log file is written. 
 
 Below this is a free text box. This works the same way, and uses the same keywords (shown below) as the end run log, outputting data in the current state for the iteration at which the log is written -- it is only possible to enter text into this box when the Write running log checkbox is ticked.
 
 There are then two further options at the base of the dialogue:
 
[…]
