### 2024-08-18 b3ffa2c
Ensure the tests don't output etraneous files
