+++ testinternal.cpp
@@ -369,7 +369,7 @@ bool testinternal::testThree(QString &outString)
     simSettings.playingfieldNumber = 3;
     simSettings.runForTaxa = 5;
     simSettings.test = 3;
-    simSettings.workingLog = true;
+    //simSettings.workingLog = true;
 
     if (theMainWindow)
         theMainWindow->resizeGrid(simSettings.runForTaxa, simSettings.genomeSize);
