### 2024-08-18 41e801f
Minor improvements to wording throughout the paper, nothing of significance
