+++ JOSS/paper.md
@@ -81,7 +81,7 @@ Simulations provide valuable insights into the patterns and processes of evoluti
 
 # Background 
 
-A range of digital platforms to study evolution and ecology exist, with varied approaches and levels of abstraction [@Dolson_Ofria_2021]. TREvoSim is sister-package to the spatially explicit eco-evolutionary simulation REvoSim [@Garwood_Spencer_Sutton_2019; @Furness_Garwood_Sutton_2023]. v1.0.0 was developed to investigate the accuracy and precision of phylogenetic inference methods [@Keating_Sansom_Sutton_Knight_Garwood_2020]. After further development, TREvoSim v2.0.0 was used to investigate the impact fossils have on phylogenetic inference and evolutionary timescales [@Mongiardino_Koch_Garwood_Parry_2021; @Mongiardino_Koch_Garwood_Parry_2023]. In brief, TREvoSim is a non spatially-explicit model in which organisms — which consist of a genome of binary characters — compete within a structure called the playing-field to echo natural selection (Figure 1). Their chance of replication is dictated by a fitness algorithm that assesses organismal fit against a series of random numbers (masks, constituting an environment). On replication, organisms have a chance of mutation, and descendents overwrite a current member of the playing field. The simulation has a lineage-based species concept, and at the end of a simulation can output trees and characters (species genomes), as well as logging the simulation state as the model runs.
+A range of digital platforms to study evolution and ecology exist, with varied approaches and levels of abstraction [@Dolson_Ofria_2021]. TREvoSim is a sister-package to the spatially explicit eco-evolutionary simulation REvoSim [@Garwood_Spencer_Sutton_2019; @Furness_Garwood_Sutton_2023]. v1.0.0 was developed to investigate the accuracy and precision of phylogenetic inference methods [@Keating_Sansom_Sutton_Knight_Garwood_2020]. After further development, TREvoSim v2.0.0 was used to investigate the impact fossils have on phylogenetic inference and evolutionary timescales [@Mongiardino_Koch_Garwood_Parry_2021; @Mongiardino_Koch_Garwood_Parry_2023]. In brief, TREvoSim is a non spatially-explicit model in which organisms -- which consist of a genome of binary characters -- compete within a structure called the playing-field to echo natural selection (Figure 1). Their chance of replication is dictated by a fitness algorithm that assesses organismal fit against a series of random numbers (masks, constituting an environment). On replication, organisms have a chance of mutation, and descendents overwrite a current member of the playing field. The simulation has a lineage-based species concept, and at the end of a simulation the software can output trees and characters (species genomes), as well as logging the simulation state as the model runs.
 
 ![Figure 1 - A simplified overview of TREvoSim. Green text represents a user-defined variable and the given value is default. The figure is split into data structures, the fitness calculation, the algorithm, and the tree. In the tree, character change is represented by change on the Y axis, time on the X. Any lineage is likely to comprise multiple individuals: modal genomes are shown in solid lines, and non-modal via dashed lines. By default, genomes for each species are recorded on their extinction. A full description of the algorithm is available in the [TREvoSim documentation](https://trevosim.readthedocs.io/en/latest/).](./Figure_01.png)
 
@@ -93,7 +93,7 @@ TREvoSim v3.0.0 includes a suite of new features that allow the investigation of
 
 ### Ecosystem engineering
 
[…]
