### 2024-08-18 e560917
Add batch to command line parser for issue #33 - actual batch is not yet implemetned
