### 2024-08-18 bd817df
Add reviewer acknowledgements to JOSS paper #60
