### 2024-08-19 #6722 @atcribb
Hi @RussellGarwood and @fboehm – these updates look great. In particular, I really appreciate the detail added to the new experimental design and data analysis sections. I've checked off the remaining points on my checklist and I am also happy to recommend this for publication.
