### 2024-08-19 f4edf47
Minor addition to funding details
