### 2024-08-19 d35d3bd
Update README.md
A few tweaks prior to new release
