### 2024-08-21 #6722 @RussellGarwood
Thanks ever so much for all your input @ms609 and @atcribb - as per the Acnowledgements I've added to the paper, this process has really helped improve the software and documentation, and we really appreciate all your time and feedback!

@fboehm - I have a release for v3.0.0 incorporating all changes for this paper lined up ready to go. I think it makes sense for me to do that release, and then link to that on GiHub and Zenodo from the paper (links in the paper currently point towards the release candidate), prior to the next steps.. Alternatively I can update the paper to always point to the latest release on GitHub and Zenodo, but I think linking the specific release for this paper is sensible. Please can you let me know if I missed anything, otherwise I shall do the release, update the JOSS paper draft, and then post here when it is done!
