### 2024-08-22 #6722 @RussellGarwood
@fboehm - Further to yesterday's message, I've been completed a release of v3.0.0 to accompany this paper, and pointed the JOSS manuscript towards both that release on GitHub, and the associated Zenodo repository, so I think all is in order to publish the paper, assuming you are happy to do so.

@ms609 - In the rush to submit this in March, I stupidly tagged the release candidate as v3.0.0 (sans -rc1), which came back to bite me and @alanspencer when doing this release. I've fixed the issue, but assuming you want to keep your TREvoSim repository for future developments (batch mode from command line is almost done!), you may want to either reclone it, or run:

`git fetch --prune --prune-tags`

To avoid issues with clashing tags. Apologies!
