### 2024-08-22 #6722 @RussellGarwood
Thanks @fboehm - most are done now, I'll work through these now and upate the checklist below as I work through them!

- [x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. [Release v3.0.0](https://github.com/palaeoware/trevosim/releases/tag/v3.0.0)
