### 2024-08-22 e101644
Add a warning when replicates set on command line but not file
