+++ main.cpp
@@ -25,11 +25,12 @@ int main(int argc, char *argv[])
     QCoreApplication::setApplicationVersion(version);
 
     // Start GUI version...
-    if (argc < 2) {
-        #if defined( _WIN32 )
-            // hide console window
-            ::ShowWindow( ::GetConsoleWindow(), SW_HIDE );
-        #endif
+    if (argc < 2)
+    {
+#if defined( _WIN32 )
+        // hide console window
+        ::ShowWindow( ::GetConsoleWindow(), SW_HIDE );
+#endif
 
         QPixmap splashPixmap(":/palaeoware_logo_square.png");
         QSplashScreen *splash = new QSplashScreen(splashPixmap, Qt::WindowStaysOnTopHint);
@@ -49,7 +50,8 @@ int main(int argc, char *argv[])
         return e;
     }
     // Start from a console...
-    else {
+    else
+    {
         //Sort out command line option
         QCommandLineParser *parser = new QCommandLineParser();
         parser->setApplicationDescription("REvoSim is an individual-based evolutionary model. You are using the command line option. See documentation or Garwood et al. (2019) Palaeontology for description of software.");
@@ -71,13 +73,19 @@ int main(int argc, char *argv[])
 
         MainWindow *w = new MainWindow;
 
+        if (!parser->isSet(opt_o) && parser->isSet(opt_b))
+        {
+            qInfo() << "You appear to have set batch mode running, but node loaded a file. TREvoSim will run these replicates with the default settings.";
+        }
+
         if (parser->isSet(opt_o))
         {
             QString fileFromCommandLine = QString();
             fileFromCommandLine = parser->value(opt_o);
             qInfo() << QString("Program launched from command line, and will try to open file %1.").arg(fileFromCommandLine);
 
-            if (!fileFromCommandLine.isNull()) {
+            if (!fileFromCommandLine.isNull())
+            {
                 w->runFromCommandLine(fileFromCommandLine);
             }
         }
