### 2024-08-24 6b77f44
Implementation of batch mode for command line issue #31
