+++ mainwindow.cpp
@@ -214,41 +214,79 @@ void MainWindow::setStatus(QString message)
     ui->statusBar->showMessage(message);
 }
 
-void MainWindow::runFromCommandLine(QString fileFromCommandLine)
+void MainWindow::runFromCommandLine(QHash<QString, QString> parsedOptions)
 {
     runFromCommand = true;
+    qInfo().noquote() << QString("Program launched from command line");
 
-    //Check if this works - it is easier to do this using the settings load command than the open and load commands herein
-    QFile settingsFile(fileFromCommandLine);
-    if (!settingsFile.open(QIODevice::ReadOnly) || !simSettings->loadSettings(&settingsFile))
+    if (parsedOptions.contains("fileFromCommandLine"))
     {
-        qInfo().noquote() << "Couldn't load settings from file. Please check the path and try again. " + QString(PRODUCTNAME) + " will now quit.";
-        exit(0);
+        QString fileFromCommandLine = parsedOptions.value("fileFromCommandLine");
+        qInfo().noquote() << QString("Will try to open file %1.").arg(fileFromCommandLine);
+
+        //Check if this works - it is easier to do this using the settings load command than the open and load commands herein
+        QFile settingsFile(fileFromCommandLine);
+        if (!settingsFile.open(QIODevice::ReadOnly) || !simSettings->loadSettings(&settingsFile))
+        {
+            qInfo().noquote() << "Couldn't load settings from file. Please check the path and try again. " + QString(PRODUCTNAME) + " will now quit.";
+            exit(0);
+        }
+        //Open deals with resizing the displays for us to avoid crashes, so call it despite having loaded the settings
+        else
+        {
+            settingsFile.close();
+            open(fileFromCommandLine);
+        }
+
+        qInfo() << "File loaded, settings now as follows:";
+        qInfo() << simSettings->printSettings();
     }
-    //Open deals with resizing the displays for us to avoid crashes, so call it despite having loaded the settings
+    //If we have no file, let's load defaults rather than whatever was last used in a run, as this makes more sense
     else
     {
-        settingsFile.close();
-        open(fileFromCommandLine);
+        defaultSettings();
+        qInfo() << "No file was specified, so TREvoSim will run with the default settings, which are as follows:";
+        qInfo() << simSettings->printSettings();
     }
+    //Note that this doesn't persist as from command line there is no save on exit
 
-    qInfo() << "File loaded - starting simulation.";
-    //First sort GUI
-    startRunGUI();
-
-    bool error = false;
-    //Create a new simulation object - sending it important settings.
-    simulation theSimulation(runs, simSettings, &error, this);
-    //Then set it running - send pointer to main window for GUI and access functions, and run number
-    if (error)
+    int batchReplicates = 0;
+    if (parsedOptions.contains("batchReplicates"))
     {
-        qInfo() << "Failed to initialise simulation. " + QString(PRODUCTNAME) + " will now quit.";
-        exit(0);
+        batchReplicates = parsedOptions.value("batchReplicates").toInt();
     }
-    else error = theSimulation.run();
 
-    if (!error) qInfo().noquote() << QString(PRODUCTNAME) + " has encountered an error. If the nature of this error is not obvious, please file an issue or contact the coders.";
-    else qInfo().noquote() << QString(PRODUCTNAME) + " has finished its run, and will now quit.";
+    if (batchReplicates == 0)
+    {
+
+        qInfo() << "TREvoSim is now going to run your simulation: success or error messages will be printed below.";
+        //First sort GUI
+        startRunGUI();
+
+        bool error = false;
[…]
