### 2024-08-25 214715d
Add speciation mode defines for issue #59
