### 2024-08-27 #6722 @RussellGarwood
Apologies for asking @fboehm, but could I trouble you for an ETA on the publication of this? That way I can let the journal that is holding back another publication waiting for a doi on this that the end is in sight!
