### 2024-08-30 f96e48b
To avoid confusion with the introduction of speciation modes, rename Sansomian speciation to something a bit more obvious. Sorry Rob.
