+++ docs/settings.rst
@@ -45,8 +45,8 @@ Fitness target
     This is the target value for the count of 1s in the fitness calculation. See :ref:`algorithmconcepts` for details of the fitness algorithm. At low and high values (max being mask number multiplied by character number) fewer genomes will have peak fitness, at half max value, a larger number of peak fitness organisms exist. This can be quantified using the :ref:`Fitness histogram` menu command.
 No selection
     If this option is enabled, fitnesses are not calculated for digital organisms in the playing field, and instead every iteration and random individual is selected for replication.
-Sansomian speciation
-    This option dictates when the character string of a species is recorded: at speciation, or at extinction. Sansomian speciation (named because it was Dr Rob Sansom's idea, rather than being a wider concept), the default, is to record the genome/characters of a species when that species goes extinct. This is useful, in particular with respect to species zero -- that used to seed the simulation. Because this tends to survive for longer than other species (see discussion in :ref:`Tree Rooting`), the genome at the start of a species lifespan could differ significantly from that at its extinction. Recording the genome of the last individual of the species ensures in such long-surviving species that the recorded characters are closest that of its sister species (and also the last common ancestor of these two species). When the option is not selected, genome is recorded at speciation.
+Record genomes on extinction
+    This option controls when the character string of a species is recorded, which can occur either at the point of speciation, or at extinction. When this option is checked, which it is by default, the genome/characters of a species are written when that species goes extinct. This is useful, in particular, with respect to species zero -- that used to seed the simulation. Because this tends to survive for longer than other species (see discussion in :ref:`Tree Rooting`), the genome at the start of a species lifespan could differ significantly from that at its extinction. Recording the genome of the last individual of the species ensures in such long-surviving species that the recorded characters are closest that of its sister species (and also the last common ancestor of these two species). When the option is not selected, genome is recorded at the birth of each species.
 Discard deleterious mutations
     Optionally, TREvoSim can accept only mutations which are neutral, or improve the fitness of an organism.
 Strip uninformative characters via subsampling
