### 2024-08-30 2c1e512
Somehoe the last changes removed the QDialogueButtonBox - resinstate with this set of modifications, manually connect slots, rename ambiguous buttonbox in stochastic dialogue, and update image in docs
