### 2024-08-31 #6722 @fboehm
@RussellGarwood - it looks like there are small discrepancies between the author list on the archive (https://zenodo.org/records/13358420) and the author list on the article proof pdf (https://raw.githubusercontent.com/openjournals/joss-papers/joss.06722/joss.06722/10.21105.joss.06722.pdf)

![Screenshot from 2024-08-31 18-45-24](https://github.com/user-attachments/assets/526d8456-a78b-4b99-83d9-24ce1cb9eec9)

Cna you fix the author names so that they match exactly? I think that some authors have middle initials on the pdf and not on the zenodo archive. For example, Luke Parry is an author of the archive, but Luke A. Parry is an author of the manuscript.
