### 2024-09-01 #6722 @RussellGarwood
@fboehm Apologies for missing these - I have fixed them (these were those people that already had Zenodo accounts that I had forgotten to modify beyond the names in that system to match the paper). Thanks for catching the discrepancy :)
