### 2024-09-03 065dd50
Move combo box to correct place, add options to constructor - next add them to the combo box
