### 2024-09-03 e71cd12
Add species mode setting
