### 2024-09-04 4f711e7
Change name of function for checking for speciations to reflect new mode in simulation files
