### 2024-09-05 dd2dd8b
Update tests to reflect change in function name
