### 2024-09-06 #6722 @RussellGarwood
@fboehm I'm being chased by the journal that is holding a publication for this to appear - is there any chance I could an update / an ETA? Specifically they're just after the DOI for the paper - so if that is available before the paper is published, just an idea of when that may be would be really useful to know.  Thanks loads for your help :)
