### 2024-09-09 ccdea37
change function call, function parameters, and the code in the simulation loop
