+++ simulation.h
@@ -94,8 +94,8 @@ private:
     void applyPerturbation();
     void applyPlayingfieldMixing(QVector<Organism *> &speciesList);
     void applyEcosystemEngineering(QVector<Organism *> &speciesList, bool writeEcosystemEngineers);
-    void testForUninformative(QVector <Organism *> &speciesList, QList <int> &uninformativeCoding, QList <int> &uninformativeNonCoding);
-    bool testForCharacterNumber(QList <int> &uninformativeCoding, QList <int> &uninformativeNonCoding);
+    void checkForUninformative(QVector <Organism *> &speciesList, QList <int> &uninformativeCoding, QList <int> &uninformativeNonCoding);
+    bool checkForCharacterNumber(QList <int> &uninformativeCoding, QList <int> &uninformativeNonCoding);
     bool stripUninformativeCharacters(QVector<Organism *> &speciesList, const QList <int> &uninformativeCoding, const QList <int> &uninformativeNonCoding);
     bool checkForUnresolvableTaxa(QVector<Organism *> &speciesList, QString &unresolvableTaxonGroups, int &unresolvableCount);
     bool writeFile(const QString logFileNameBase, const QString logFileExtension, const QString logFileString, const QHash<QString, QString> &outValues, const QVector<Organism *> &speciesList);
@@ -110,7 +110,7 @@ private:
     int genomeDifference(const Organism *organismOne, const Organism *organismTwo);
     QHash<QString, QVector<int> > checkForExtinct(const QVector <Organism *> &speciesList);
     void speciesExtinction(Organism *speciesListOrganism, const Organism *playingFieldOrganism, int extinctIteration, bool samsonian, bool stochastic, bool test = false);
-    int checkForSpeciation(const Organism *organismOne, int runSelectSize);
+    bool checkForSpeciation(const Organism *organismOne, int runSelectSize, int runSpeciesDifference, int speciationMode);
 };
 
 #endif // SIMULATION_H
