### 2024-09-10 #6722 @fboehm
Author lists (ie, names & ordering) now match exactly in the manuscript pdf and the zenodo archive.

Author affiliations (ie, institution names) don't match exactly, but I don't see the differences as important. Note that there are 11 affiliation numbers in zenodo archive and 10 in the manuscript pdf. The differences, however, come down to, for example, one author lists both a university name and a college name or solely a university name.

thus, this looks good.

**pdf manuscript author list** is now:

Russell J. Garwood 1,2¶ , Alan R. T. Spencer 2,3, Charles T. M. Bates 1 ,
4
 Leah M. Callender-Crowe 4 , Frances S. Dunn 5 , Thomas J. D.
5
 Halliday 6 , Joseph N. Keating 7 , Nicolás Mongiardino Koch 8 , Luke A.
6
 Parry 9 , Robert S. Sansom 1 , Thomas J. Smith 9 , Mark D. Sutton 3 ,
7
 and Thomas Vanteghem 10

**AND AUTHOR LIST ON ZENODO ARCHIVE** Is now:

[Garwood, Russell J1, 2](https://zenodo.org/search?q=metadata.creators.person_or_org.name%3A%22Garwood,+Russell+J%22)
ORCID icon

[Spencer, Alan R.T.3, 2](https://zenodo.org/search?q=metadata.creators.person_or_org.name%3A%22Spencer,+Alan+R.T.%22)
ORCID icon

[Bates, Charles, T. M.1](https://zenodo.org/search?q=metadata.creators.person_or_org.name%3A%22Bates,+Charles,+T.+M.%22)
ORCID icon

[Callender-Crowe, Leah M.4](https://zenodo.org/search?q=metadata.creators.person_or_org.name%3A%22Callender-Crowe,+Leah+M.%22)
ORCID icon

[Dunn, Frances S.5](https://zenodo.org/search?q=metadata.creators.person_or_org.name%3A%22Dunn,+Frances+S.%22)
ORCID icon

[Halliday, Thomas J.D.6](https://zenodo.org/search?q=metadata.creators.person_or_org.name%3A%22Halliday,+Thomas+J.D.%22)
ORCID icon

[Keating, Joseph, N.7](https://zenodo.org/search?q=metadata.creators.person_or_org.name%3A%22Keating,+Joseph,+N.%22)
ORCID icon

[Mongiardino Koch, Nicolás8](https://zenodo.org/search?q=metadata.creators.person_or_org.name%3A%22Mongiardino+Koch,+Nicol%C3%A1s%22)
ORCID icon

[Parry, Luke A.9](https://zenodo.org/search?q=metadata.creators.person_or_org.name%3A%22Parry,+Luke+A.%22)
ORCID icon

[Sansom, Robert S.1](https://zenodo.org/search?q=metadata.creators.person_or_org.name%3A%22Sansom,+Robert+S.%22)
ORCID icon

[Smith, Thomas J.10](https://zenodo.org/search?q=metadata.creators.person_or_org.name%3A%22Smith,+Thomas+J.%22)
ORCID icon

[Sutton, Mark D.3](https://zenodo.org/search?q=metadata.creators.person_or_org.name%3A%22Sutton,+Mark+D.%22)
ORCID icon

[Vanteghem, Thomas11](https://zenodo.org/search?q=metadata.creators.person_or_org.name%3A%22Vanteghem,+Thomas%22)
ORCID icon
