### 2024-09-10 #6722 @fboehm
I have one small request, @RussellGarwood, for the manuscript text:

line 72 of the pdf: you use the word "v.s.". I think it should be "vs."

I think overall it reads well, and I think only that one very small change is needed before we proceed further.
