### 2024-09-10 e9c6915
OK, we need organism to store a list of parent and speciation genomes now
