### 2024-09-10 6098114
SOrt out parent genomes in the organism class
