### 2024-09-10 #6722 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
