### 2024-09-11 cb55c9e
Deal with new species and print playingfield with new parentgenomes setup - need to check latter. Next need to deal with speciation modes in check for new species function in main sim
