### 2024-09-11 #6722 @Kevin-Mattheus-Moerman
On the paper:
- [ ] In  your affiliations please spell out all country names. So use United Kingdom instead of UK, and use Unites States (or United States of America) instead of USA.
