### 2024-09-11 8018b12
Update country names in JOSS paper
