### 2024-09-11 #6722 @RussellGarwood
@fboehm - thanks loads for all of your help throughout, it is much appreciated, and this process has really imrpvoed the software. I'm really pleased with the progress the JOSS process has inspired!

@Kevin-Mattheus-Moerman Thanks for doing these checks. With the last merge (I forgot to pull first, apologies) I have made the change you requested!
