### 2024-09-14 bc1eb13
sorting out checkforspeciation - a bit of a mess at the moment but no more time
