### 2024-09-16 #6722 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
