### 2024-10-10 palaeoware/trevosim#14 @RussellGarwood
Thanks again for this suggestions @ms609 - sadly my coding time has been limited since the start of september due to teaching, but with the above push I have just included a codemeta.json. A great tool!
